\name{AllCountries2e}
\alias{AllCountries2e}
\docType{data}
\title{
AllCountries - 2e
}
\description{
Data on the countries of the world
}
%\usage{data(AllCountries2e)}
\format{
  A dataset with 215 observations on the following 25 variables.
  \tabular{rl}{
    \code{Country} \tab {Name of the country}\cr
    \code{LandArea} \tab {Size in 1000 sq. kilometers}\cr
    \code{Population} \tab {Population in millions}\cr
    \code{Density} \tab {Number of people per square kilometer}\cr
    \code{GDP} \tab {Gross Domestic Product (in $US) per capita}\cr
    \code{Rural} \tab {Percentage of population living in rural areas}\cr
    \code{CO2} \tab {CO2 emissions (metric tons per capita)}\cr
    \code{PumpPrice} \tab {Price for a liter of gasoline ($US)}\cr
    \code{Military} \tab {Percentage of government expenditures directed toward the military}\cr
    \code{Health} \tab {Percentage of government expenditures directed towards healthcare}\cr
    \code{ArmedForces} \tab {Number of active duty military personnel (in 1,000's)}\cr
    \code{Internet} \tab {Percentage of the population with access to the internet}\cr
    \code{Cell} \tab {Cell phone subscriptions (per 100 people)}\cr
    \code{HIV} \tab {Percentage of the population with HIV}\cr
    \code{Hunger} \tab {Percent of the population considered undernourished}\cr
    \code{Diabetes} \tab {Percent of the population diagnosed with diabetes}\cr
    \code{BirthRate} \tab {Births per 1000 people}\cr
    \code{DeathRate} \tab {Deaths per 1000 people}\cr
    \code{ElderlyPop} \tab {Percentage of the population at least 65 years old}\cr
    \code{LifeExpectancy} \tab {Average life expectancy (years)}\cr
    \code{FemaleLabor} \tab {Percent of females 15 - 64 in the labor force}\cr
    \code{Unemployment} \tab {Percent of labor force unemployed}\cr
    \code{Energy} \tab {Energy usage (kilotons of oil equivalent)}\cr
    \code{Electricity} \tab {Electric power consumption (kWh per capita)}\cr
    \code{Developed} \tab {Categories for kilowatt hours per capita, 1= under 2500, 2=2500 to 5000, 3=over 5000}\cr
  }
}
\details{
Data for each variable were collected for years between 2012 and 2014.  Within a variable all country measurements are from the same year, but the year may vary between different variables depending on availability. \cr
** From 2e - dataset has been updated for 3e **
}
\source{
Data collected from the World Bank website, \emph{worldbank.org}.
}

\keyword{datasets}
