\name{APMultipleChoice}
\alias{APMultipleChoice}
\docType{data}
\title{
AP Multiple Choice
}
\description{
Correct responses on Advanced  Placement multiple choice exams
}
%\usage{data(APMultipleChoice)}
\format{
  A dataset with 400 observations on the following variable.
  \tabular{rl}{
    \code{Answer} \tab {Correct response: \code{A},  \code{B},  \code{C}, \code{D}, or \code{E}}\cr
  }
}
\details{
Correct responses from multiple choice sections for a sample of released Advanced Placement exams
}
\source{
Sample exams from several disciplines at \emph{http://apcentral.collegeboard.com}
}

\keyword{datasets}
