\name{April14Temps}
\alias{April14Temps}
\docType{data}
\title{April 14th Temperatures}
\description{
Temperatures in Des Moines, IA and San Francisco, CA on April 14th
}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{1995 to 2024}
    \item{\code{DesMoines}}{Temperature in Des Moines (degrees F)}
    \item{\code{SanFrancisco}}{Temperature in San Francisco (degrees F)}
  }
}
\details{
Average temperature for the day of April 14th in each of 30 years from 1995-2024 \cr
** Data set updated for 4e (earlier versions are now April14Temps3e, April14Temps2e, and April14Temps1e) **
}
\source{
Original data downaded from the University of Dayton Average Daily Temperature Archive at
\emph{https://academic.udayton.edu/kissock/http/Weather/citylistUS.htm} \cr

Recent updates from \url{https://www.wunderground.com/history/daily/us/ca/san-francisco/KSFO} and \url{https://www.wunderground.com/history/daily/us/ia/des-moines/KDSM}
}
\keyword{datasets}
