\name{April14Temps1e}
\alias{April14Temps1e}
\docType{data}
\title{
April 14th Temperatures -1e
}
\description{
Temperatures in Des Moines, IA and San Francisco, CA on April 14th
}
%\usage{data(April14Temps)}
\format{
  A dataset with 16 observations on the following 3 variables.
  \tabular{rl}{
    \code{Year} \tab {1995-2010} \cr
    \code{DesMoines} \tab {Temperature in Des Moines (degrees F)} \cr
    \code{SanFrancisco} \tab {Temperature in San Francisco (degrees F)}
  }
}
\details{
Average temperature for the day of April 14th in each of 16 years from 1995-2010 \cr
** From 1e - dataset has been updated for 2e **
}
\source{
The University of Dayton Average Daily Temperature Archive at \cr
\emph{http://academic.udayton.edu/kissock/http/Weather/citylistUS.htm}
}

\keyword{datasets}
