\name{April14Temps2e}
\alias{April14Temps2e}
\docType{data}
\title{
April 14th Temperatures - 2e
}
\description{
Temperatures in Des Moines, IA and San Francisco, CA on April 14th
}
%\usage{data(April14Temps2e)}
\format{
  A dataset with 21 observations on the following 3 variables.
  \tabular{rl}{
    \code{Year} \tab {1995 to 2015} \cr
    \code{DesMoines} \tab {Temperature in Des Moines (degrees F)} \cr
    \code{SanFrancisco} \tab {Temperature in San Francisco (degrees F)}
  }
}
\details{
Average temperature for the day of April 14th in each of 21 years from 1995-2015 \cr
** From 2e - dataset has been updated for 3e **
}
\source{
The University of Dayton Average Daily Temperature Archive at \cr
\emph{http://academic.udayton.edu/kissock/http/Weather/citylistUS.htm}
}

\keyword{datasets}
