\name{BaseballHits2014}
\alias{BaseballHits2014}
\docType{data}
\title{
Baseball Hits - 2014
}
\description{
Number of hits, wins, and other stats for MLB teams - 2014
}
%\usage{data(BaseballHits2014)}
\format{
  A dataset with 30 observations on the following 14 variables.
  \tabular{rl}{
    \code{Team} \tab {Name of baseball team (3-character code)}\cr
    \code{League} \tab {Either \code{AL} or \code{NL}}\cr
    \code{Wins} \tab {Number of wins for the season}\cr
    \code{Runs} \tab {Number of runs scored}\cr
    \code{Hits} \tab {Number of hits}\cr
    \code{Doubles} \tab {Number of doubles}\cr
    \code{Triples} \tab {Number of triples}\cr
    \code{HomeRuns} \tab {Number of home runs}\cr
    \code{RBI} \tab {Number of runs batted in}\cr
    \code{StolenBases} \tab {Number of stolen bases}\cr
    \code{CaughtStealing} \tab {Number of times caught stealing}\cr
    \code{Walks} \tab {Number of walks}\cr
    \code{Strikeouts} \tab {Number of strikeouts}\cr
    \code{BattingAvg} \tab {Team batting average}\cr
  }
}
\details{
Data from the 2014 Major League Baseball regular season. \cr
** From 2e - dataset has been updated for 3e **
}
\source{
\emph{http://www.baseball-reference.com/leagues/MLB/2014-standard-batting.shtml}
}

\keyword{datasets}
