\name{BaseballHits2019}
\alias{BaseballHits2019}
\docType{data}
\title{Baseball Team Statistics (2019)}
\description{
Number of hits, wins, and other stats for MLB teams in 2019
}
\format{
  A data frame with 30 observations on the following 14 variables.
  \describe{
    \item{\code{Team}}{Name of baseball team (3-character code)}
    \item{\code{League}}{Either \code{AL} or \code{NL}}
    \item{\code{Wins}}{Number of wins for the season}
    \item{\code{Runs}}{Number of runs scored}
    \item{\code{Hits}}{Number of hits}
    \item{\code{Doubles}}{Number of doubles}
    \item{\code{Triples}}{Number of triples}
    \item{\code{HomeRuns}}{Number of home runs}
    \item{\code{RBI}}{Number of runs batted in}
    \item{\code{StolenBases}}{Number of stolen bases}
    \item{\code{CaughtStealing}}{Number of times caught stealing}
    \item{\code{Walks}}{Number of walks}
    \item{\code{Strikeouts}}{Number of strikeouts}
    \item{\code{BattingAvg}}{Team batting average}
  }
}
\details{
Offensive team statistics for the 2019 Major League Baseball regular season. \cr
** Updated for 3e (earlier versions are now BaseballHits2014 and BaseballHits1e)
}
\source{
\url{https://www.baseball-reference.com/leagues/MLB/2019-standard-batting.shtml}
}
\keyword{datasets}
