\name{BaseballSalaries2015}
\alias{BaseballSalaries2015}
\docType{data}
\title{
MLB Player Salaries in 2015
}
\description{
Opening Day salaries for all Major League Baseball players in 2015
}
%\usage{data(BikeCommute)}
\format{
  A dataset with 868 observations on the following 4 variables.
  \tabular{rl}{
    \code{Name} \tab {Player's name}\cr
    \code{Salary} \tab {2015 season salary (in millions)}\cr
    \code{Team} \tab {Abbreviated team name}\cr
    \code{Position} \tab {Code for player's main position}\cr
  }
}
\details{
Yearly salary (in millions of dollars) for all players on the rosters of Major League Baseball teams at the start of the 2015 season.\cr
** From 2e - dataset has been updated for 3e **
}
\source{
\emph{ http://www.usatoday.com/sports/mlb/salaries}
}


\keyword{datasets}
