\name{BaseballSalaries2019}
\alias{BaseballSalaries2019}
\docType{data}
\title{MLB Player Salaries in 2019}
\description{
Opening Day salaries for all Major League Baseball players in 2019
}
\format{
  A data frame with 877 observations on the following 4 variables.
  \describe{
    \item{\code{Name}}{Player's name}
    \item{\code{Salary}}{2019 season salary (in millions)}
    \item{\code{Team}}{Abbreviated team name}
    \item{\code{POS}}{Code for player's main position}
  }
}
\details{
Yearly salary (in millions of dollars) for all players on the rosters of Major League Baseball teams at the start of the 2019 season. \cr
** Updated for 3e (earlier version for 2015 is at BaseballSalaries2015). **
}
\source{
\url{https://databases.usatoday.com/mlb-salaries/}
}
\keyword{datasets}
