\name{BaseballSalaries2024}
\alias{BaseballSalaries2024}
\docType{data}
\title{MLB Player Salaries in 2024}
\description{
Opening Day salaries for all Major League Baseball players in 2024
}
\format{
  A data frame with 952 observations on the following 4 variables.
  \describe{
    \item{\code{Name}}{Player's name}
    \item{\code{Salary}}{2024 season salary (in millions)}
    \item{\code{Team}}{Team name}
    \item{\code{Pos}}{Code for player's main position}
  }
}
\details{
Yearly salary (in millions of dollars) for all players on the rosters of
Major League Baseball teams at the start of the 2024 season. \cr
** Updated for 4e (earlier versions are BaseballSalaries2019 and BaseballSalaries2015). **
}
\source{
Downloaded from USA Today at
\url{https://databases.usatoday.com/major-league-baseball-salaries-2024/}.
(March 2025)
}
\keyword{datasets}
