\name{BaseballTimes}
\alias{BaseballTimes}
\docType{data}
\title{Baseball Game Times (2024)}
\description{
Information for a sample of 50 Major League Baseball games played during the 2024 season
}
%\usage{data(BaseballTimes)}
\format{
  A dataset with 50 observations on the following 9 variables.
  \tabular{rl}{
    \code{Away} \tab {Away team name}\cr
    \code{Home} \tab {Home team name}\cr
    \code{Runs} \tab {Total runs scored (both teams)}\cr
    \code{Margin} \tab {Margin of victory}\cr
    \code{Hits} \tab {Total number of hits (both teams)}\cr
    \code{Errors} \tab {Total number of errors (both teams)}\cr
    \code{Pitchers} \tab {Total number of pitchers used (both teams)}\cr
    \code{Walks} \tab {Total number of walks (both teams)}\cr
    \code{Time} \tab {Elapsed time for game (in minutes)}\cr
  }
}
\details{
Data from a sample of boxscores for Major League Baseball games played during the 2024 season. Games include all played on May 6th, June 6th, July 13th, and August 16th,
}
\source{
Data obtained from boxscores at \url{https://www.baseball-reference.com/boxes/index.fcgi}
}

\keyword{datasets}
