\name{Benford}
\alias{Benford}
\docType{data}
\title{
Benford data
}
\description{
Two examples to test Benford's Law
}
%\usage{data(Benford)}
\format{
  A dataset with 9 observations on the following 4 variables.
  \tabular{rl}{
    \code{Digit} \tab {Leading digit (1-9)}\cr
    \code{BenfordP} \tab {Expected proportion according to Benford's law}\cr
    \code{Address} \tab {Frequency as a first digit in an address}\cr
    \code{Invoices} \tab {Frequency as the first digit in invoice amounts}\cr
  }
}
\details{
Leading digits from 1188 addresses sampled from a phone book and 7273 amounts from invoices sampled at a company.
}
\source{
Thanks to Prof. Richard Cleary for providing the data
}

\keyword{datasets}
