\name{BikeCommute}
\alias{BikeCommute}
\docType{data}
\title{
Bike Commute
}
\description{
Commute times for two kinds of bicycle
}
%\usage{data(BikeCommute)}
\format{
  A dataset with 56 observations on the following 9 variables.
  \tabular{rl}{
    \code{Bike} \tab {Type of material \code{Carbon} or \code{Steel}}\cr
    \code{Date} \tab {Date of the bike commute}\cr
    \code{Distance} \tab {Length of commute (in miles)}\cr
    \code{Time} \tab {Total commute time (hours:minutes:seconds)}\cr
    \code{Minutes} \tab {Time converted to minutes}\cr
    \code{AvgSpeed} \tab {Average speed during the ride (miles per hour)}\cr
    \code{TopSpeed} \tab {Maximum speed (miles per hour)}\cr
    \code{Seconds} \tab {Time converted to seconds}\cr
    \code{Month} \tab { Categories: \code{1Jan} \code{2Feb} \code{3Mar} \code{4Apr} \code{5May} \code{6June} \code{7July}}\cr
  }
}
\details{
Data from a personal experiment to compare commuting time based on a randomized selection between two bicycles made of different materials. 
}
\source{
Thanks to Dr. Groves for providing his data.
}
\references{
Bicycle weight and commuting time: randomised trial, in British Medical Journal, BMJ 2010;341:c6801.
}

\keyword{datasets}
