\name{BMI}
\alias{BMI}
\docType{data}
\title{Body Mass Index and Exercise}
\description{
Body Mass Index (BMI) and exercsie indicator from the 2023 BRFSS survey
}
\format{
  A data frame with 392,788 observations on the following 2 variables.
  \describe{
    \item{\code{BMI}}{Body Mass Index = weight (in kg) / (height (in m))^2}
    \item{\code{Exercise}}{Have you exercised in last 30 days? (\code{Yes} or \code{No})}
}
}
\details{
Data from the 2023 Behavioral Risk Factor Surveillance System (BRFSS) survey.  BMI is calculated as weight (in kg) divided by height (in meters) squared. Values above 25 are considered overweight and over 30 are clasified as obese.
}

\source{
Centers for Disease Control and Prevention (CDC), Behavioral Risk Factor Surveillance System Survey Data, U.S. Department of Health and Human Services, Centers for Disease Control and Prevention, Atlanta, 2023. \url{https://www.cdc.gov/brfss/annual_data/annual_2023.html}

}
\keyword{datasets}
