\name{BodyFat}
\alias{BodyFat}
\docType{data}
\title{Body Measurements}
\description{
Percent fat and other body measurements for a sample of men
}
%\usage{data(BodyFat)}
\format{
  A dataset with 100 observations on the following 10 variables.
  \tabular{rl}{
    \code{Bodyfat} \tab {Percent body fat}\cr
    \code{Age} \tab {Age in years}\cr
    \code{Weight} \tab {Weight in pounds}\cr
    \code{Height} \tab {Height in inches}\cr
    \code{Neck} \tab {Neck circumference in cm.}\cr
    \code{Chest} \tab {Chest circumference in cm.}\cr
    \code{Abdomen} \tab {Abdomen cirumference in cm.}\cr
    \code{Ankle} \tab {Ankle circumference in cm.}\cr
    \code{Biceps} \tab {Exended biceps circumference in cm.}\cr
    \code{Wrist} \tab {Wrist circumference in cm. }
  }
}
\details{This is a subset of a larger sample of men who each had a percent body fat estimated by an underwater weighing technique.  Other measurements were taken to see how they might be used to predict the body fat percentage.

}
\source{
These data were contributed by Roger Johnson,  then at Carleton University, to the Datasets Archive at the Journal of Statistics Education. \cr \emph{https://ww2.amstat.org/publications/jse/v4n1/datasets.johnson.html} \cr
The data were originally  supplied by Dr. A. Garth Fisher, Human Performance Research Center, Brigham Young University, Provo, Utah 84602.
}

\keyword{datasets}
