\name{BodyTemp50}
\alias{BodyTemp50}
\docType{data}
\title{Body Temperatures}
\description{
Sample of 50 body temperatures
}
\format{
  A data frame with 50 observations on the following 3 variables.
  \describe{
    \item{\code{BodyTemp}}{Body temperature in degrees F}
    \item{\code{Pulse}}{Pulse rates (beat per minute)}
    \item{\code{Sex}}{F=Female, M=Male}
  }
}
\details{
Body temperatures and pulse rates for a sample of 50 healthy adults. Note the Sex variable was labeled as Gender in earlier versions of this dataset. We acknowledge that this binary dichotomization is not a complete or inclusive representation of reality.
}
\source{
Shoemaker, "What's Normal: Temperature, Gender and Heartrate", Journal of Statistics Education, Vol. 4, No. 2 (1996) \cr
\url{https://jse.amstat.org/v4n2/datasets.shoemaker.html}
}
\keyword{datasets}
