\name{CAOSExam}
\alias{CAOSExam}
\docType{data}
\title{CAOS Exam Scores}
\description{
Scores on a pre-test and post-test of basic statistics concepts
}
%\usage{data(CAOSExam)}
\format{
  A datset with 10 observations on the following 3 variables.
  \tabular{rl}{
    \code{Student} \tab {ID code for student}\cr
    \code{Pretest} \tab {CAOS Pretest score}\cr
    \code{Posttest} \tab {CAOS Posttest score}\cr
  }
}
\details{
The CAOS (Comprehensive Assessment of Outcomes in First Statistics Course) exam is designed to measure comprehension 
basic statistical ideas in an introductory statistics course.  This dataset has scores for ten students 
who took the CAOS pre-test at the start of a course and the post-test during the course itself.  
Each exam consists of 40 multiple choice questions and the score is the percentage correct. 
}
\source{
A sample of 10 students from an introductory statisics course. 
}
\references{
Find out more about the CAOS exam at http://app.gen.umn.edu/artist/caos.html
}

\keyword{datasets}
