\name{CarbonDioxide}
\alias{CarbonDioxide}
\docType{data}
\title{Carbon Dioxide Levels}
\description{
Atmospheric carbon dioxide levels by year
}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Every five years from 1960 to 2015}
    \item{\code{C02}}{Carbon dioxide level in parts per million}
  }
}
\details{
Carbon dioxide levels in the atmosphere over a 55 year span from 1960-2015. \cr
** Updated for 3e (earlier version is now CarbonDioxide2e) **

}
\source{
Dr. Pieter Tans, NOAA/ESRL. Values recorded at the Mauna Loa Observatory in Hawaii. \url{https://gml.noaa.gov/ccgg/trends/}
}
\keyword{datasets}
