\name{CarbonDioxide2e}
\alias{CarbonDioxide2e}
\docType{data}
\title{
Carbon Dioxide Levels - 2e
}
\description{
Atmospheric carbon dioxide levels by year
}
%\usage{data(Carbondioxide2e)}
\format{
  A dataset with 11 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year} \tab {Every five years from 1960 to 2010}\cr
    \code{C02} \tab {Carbon dioxide level in parts per million}\cr
  }
}
\details{
Carbon dioxide levels in the atmosphere over a 50 year span from 1960-2010.\cr
** From 2e - dataset has been updated for 3e **
}
\source{
Dr. Pieter Tans, NOAA/ESRL (\emph{www.esrl.noaa.gov/gmd/ccgg/trends/}). Values recorded at the Mauna Loa Observatory
in Hawaii.
}

\keyword{datasets}
