\name{CarDepreciation}
\alias{CarDepreciation}
\docType{data}
\title{
Car Depreciation
}
\description{
Depreciation for 20 car models (2024).
}
%\usage{data(CarDepreciation)}
\format{
  A dataset with 20 observations on the following 5 variables.
  \tabular{rl}{
    \code{Car} \tab {Name of the car model}\cr
    \code{New} \tab {Price of a new car}\cr
    \code{Used} \tab {Value after one year}\cr
    \code{Depreciation} \tab {Drop in value after one year}\cr
    \code{PctDrop} \tab {Percent depreciation after one year}\cr
  }
}
\details{
Twenty car models were selected at random.  Original price (in dollars) and value after one year (and 12,000 miles) were recorded for each model.  The depreciation is the difference (New-Used). Updated for 4e (earlier version is now CarDepreciation3e)
}
\source{
New and used costs determined using models selected from \url{https://caredge.com/depreciation}.
}

\keyword{datasets}
