\name{Cars2015}
\alias{Cars2015}
\docType{data}
\title{
2015 Car Models
}
\description{
Information about new car models in 2015
}
%\usage{data(Cars2015)}
\format{
  A dataset with 110 observations on the following 24 variables.
  \tabular{rl}{
    \code{Make} \tab {Manufacturer (e.g. Chevrolet, Toyota, etc.)}\cr
    \code{Model} \tab {Car model  (e.g. Impala, Prius, ...)}\cr
    \code{Type} \tab {Vehicle category (\code{Small}, \code{Hatchback}, \code{Sedan}, \code{Sporty}, \code{Wagon}, \code{SUV}, \code{7Pass})}\cr
    \code{LowPrice} \tab {Lowest MSRP (in $1,000)}\cr
    \code{HighPrice} \tab {Highest MSRP (in $1,000)}\cr
    \code{Drive} \tab {Type of drive (\code{FWD}, \code{RWD}, \code{AWD})}\cr
    \code{CityMPG} \tab {City miles per gallon (EPA)}\cr
    \code{HwyMPG} \tab {Highway miles per gallon (EPA)}\cr
    \code{FuelCap} \tab {Fuel capacity (in gallons)}\cr
    \code{Length} \tab {Length (in inches)}\cr
    \code{Width} \tab {Width (in inches)}\cr
    \code{Height} \tab {Height (in inches)}\cr
    \code{Wheelbase} \tab {Wheelbase (in inches)}\cr
    \code{UTurn} \tab {Diameter (in feet) needed for a U-turn}\cr
    \code{Weight} \tab {Curb weight (in pounds)}\cr
    \code{Acc030} \tab {Time (in seconds) to go from 0 to 30 mph}\cr
    \code{Acc060} \tab {Time (in seconds) to go from 0 to 60 mph}\cr
    \code{QtrMile} \tab {Time (in seconds) to go ¼ mile}\cr
    \code{PageNum} \tab {Page number in the Consumer Reports New Car Buying Guide }\cr
    \code{Size} \tab {\code{Small}, \code{Midsized}, or \code{Large}}\cr
  }
}
\details{
Data for a set of 110 new car models in 2015 based on information in the Consumer Reports New Car Buying Guide.
}
\source{
Consumer Reports 2015 New Car Buying Guide \cr
\emph{http://www.magastack.com/issue/6053-consumer-reports-new-car-buying-guide-february-2015?page=1}
\emph{http://www.consumerreports.org/cro/cars/compare.htm?add=true&product=new/chevrolet/impala}
}

\keyword{datasets}
