\name{Cars2020}
\alias{Cars2020}
\docType{data}
\title{2020 Car Models}
\description{
Information about new car models in 2020
}
\format{
  A data frame with 110 observations on the following 21 variables.
  \describe{
    \item{\code{Make}}{Manufacturer (e.g. Chevrolet, Toyota, etc.)}
    \item{\code{Model}}{Car model  (e.g. Impala, Highlander, ...)}
    \item{\code{Type}}{Vehicle category (\code{Hatchback}, \code{Minivan}, \code{Sedan}, \code{Sporty}, \code{SUV}, or \code{Wagon})}
    \item{\code{LowPrice}}{Lowest MSRP (in $1,000)}
    \item{\code{HighPrice}}{Highest MSRP (in $1,000)}
    \item{\code{CityMPG}}{City miles per gallon (EPA)}
    \item{\code{HwyMPG}}{Highway miles per gallon (EPA)}
    \item{\code{Seating}}{Seating capacity}
    \item{\code{Drive}}{Type of drive (\code{AWD}, \code{FWD}, or \code{RWD})}
    \item{\code{Acc030}}{Time (in seconds) to go from 0 to 30 mph}
    \item{\code{Acc060}}{Time (in seconds) to go from 0 to 60 mph}
    \item{\code{QtrMile}}{Time (in seconds) to go ¼ mile}
    \item{\code{Braking}}{Distance to stop from 60 mph (dry pavement)}
    \item{\code{FuelCap}}{Fuel capacity (in gallons)}
    \item{\code{Length}}{Length (in inches)}
    \item{\code{Width}}{Width (in inches)}
    \item{\code{Height}}{Height (in inches)}
    \item{\code{Wheelbase}}{Wheelbase (in inches)}
    \item{\code{UTurn}}{Diameter (in feet) needed for a U-turn}
    \item{\code{Weight}}{Curb weight (in pounds)}
    \item{\code{Size}}{\code{Large}, \code{Midsized}, or \code{Small}}
  }
}
\details{
Data for a set of 110 new car models in 2020 based on information in the Consumer Reports. \cr
** Updated for 3e (an earlier version from 2015 is at Cars2015). **
}
\source{
Data on new car models in 2020 accessed from Consumer Reports website. \url{https://www.consumerreports.org/cars/}
}
\keyword{datasets}
