\name{Cereal}
\alias{Cereal}
\docType{data}
\title{Breakfast Cereals}
\description{
Nutrition information for a sample of 30 breakfast cereals
}
%\usage{data(Cereal)}
\format{
  A dataset with 30 observations on the following 10 variables.
  \tabular{rl}{
    \code{Name} \tab {Brand name of cereal}\cr
    \code{Company} \tab {Manufacturer coded as \code{G}=General Mills, \code{K}=Kellog's or \code{Q}=Quaker}\cr
    \code{Serving} \tab {Servind sixze (in cups)}\cr
    \code{Calories} \tab {Calories (per cup)}\cr
    \code{Fat} \tab {Fat (grams per cup)}\cr
    \code{Sodium} \tab {Sodium (mg per cup)}\cr
    \code{Carbs} \tab {Carbohydrates (grams per cup)}\cr
    \code{Fiber} \tab {Dietary Fiber (grams per cup)}\cr
    \code{Sugars} \tab {Sugars (grams per cup)}\cr
    \code{Protein} \tab {Protein (grams per cup)}\cr
  }
}
\details{
Nutrition contents for a sample of breakfast cererals, derived from nutrition lables.  
Values are per cup of cereal (rather than per serving).
}
\source{
Cereal data obtained from nutrition labels at http://www.nutritionresource.com/foodcomp2.cfm?id=0800
}

\keyword{datasets}
