\name{CityTemps}
\alias{CityTemps}
\docType{data}
\title{City Temperatures}
\description{
Mean monthly temperature in Moscow, Melbourne, and San Francisco for 2014 and 2015
}
%\usage{data(Cereal)}
\format{
  A dataset with 24 observations on the following 5 variables.
  \tabular{rl}{
    \code{Year} \tab {2014 or 2015}\cr
    \code{Month} \tab {1=January to 12=December}\cr
    \code{Moscow} \tab {Monthly temperatures in Moscow (Russia)}\cr
    \code{Melbourne} \tab {Monthly temperatures in Melbourne (Australia)}\cr
    \code{SanFrancisco} \tab {Monthly temperatures in San Francisco (United States)}\cr
  }
}
\details{
Mean monthly temperatures in degrees Celsius for the years 2014 and 2015 in each of three cities.
}
\source{
KNMI Climate Explorer at \emph{https://climexp.knmi.nl/selectstation.cgi?id=someone@somewhere}
}

\keyword{datasets}
