\name{CityTemps3e}
\alias{CityTemps3e}
\docType{data}
\title{City Temperatures - 3e}
\description{
Mean monthly temperature in Moscow, Melbourne, and San Francisco for 2017 and 2018
}
\format{
  A data frame with 24 observations on the following 5 variables.
  \describe{
    \item{\code{Year}}{2017 or 2018}
    \item{\code{Month}}{1=January through 12=December}
    \item{\code{Moscow}}{Monthly temperatures in Moscow (Russia)}
    \item{\code{Melbourne}}{Monthly temperatures in Melbourne (Australia)}
    \item{\code{San.Francisco}}{Monthly temperatures in San Francisco (United States)}
  }
}
\details{
Mean monthly temperatures in degrees C for the years 2017 and 2018 in each of three cities. \cr
** Updated for 3e (an earlier version for 2014 and 2015 is at CityTemps2e). **
}
\source{
Source:  KNMI Climate Explorer at \url{https://climexp.knmi.nl/selectstation.cgi?id=someone@somewhere} Use station codes 94866 (Melbourne), 72494 (San Francisco), 27612 (Moscow).

}
\keyword{datasets}
