\name{CO2}
\alias{CO2}
\docType{data}
\title{
CO2
}
\description{
Atmospheric carbon dioxide levels by year 
}
\usage{data(CO2)}
\format{
  A data frame with 11 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Every five years from 1960-2010}
    \item{\code{C02}}{Carbon dioxide level in parts per million}
  }
}
\details{
Carbon dioxide levels in the atmosphere over a 50 year span from 1960-2010.
}
\source{
Dr. Pieter Tans, NOAA/ESRL (www.esrl.noaa.gov/gmd/ccgg/trends/). Values recorded at the Mauna Loa Observatory
in Hawaii.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(CO2)
}
\keyword{datasets}
