\name{CocaineTreatment}
\alias{CocaineTreatment}
\docType{data}
\title{
Coacaine Treatment
}
\description{
Relapse/no relapse responses to three different treatments for cocaine addiction
}
%\usage{data(CocaineTreatment)}
\format{
  A dataset with 72 observations on the following 2 variables.
  \tabular{rl}{
    \code{Drug} \tab {Treatment drug: \code{Desipramine}, \code{Lithium}, or \code{Placebo}}\cr
    \code{Relapse} \tab {Did the patient relapse? \code{no} or \code{yes}}\cr
  }
}
\details{
Data from an experiment to investigate the effectiveness of the two drugs, desipramine
and lithium, in the treatment of cocaine addiction. Subjects (cocaine addicts seeking treatment) 
were randomly assigned to take one of the treatment drugs or a placebo. 
The response variable is whether or not the subject relapsed (went back to using cocaine) after the treatment. 
}
\source{
Gawin, F., et.al., "Desipramine Facilitation of Initial Cocaine Abstinence", Archives of General Psychiatry, 1989; 46(2):
117 - 121.
}

\keyword{datasets}
