\name{CommuteStLouis}
\alias{CommuteStLouis}
\docType{data}
\title{
Comute times in St. Louis
}
\description{
Commute times and distance for a sample of 500 people in St. Louis
}
%\usage{data(CommuteStLouis)}
\format{
  A dataset with 500 observations on the following 5 variables.
  \tabular{rl}{
    \code{City} \tab { \code{St. Louis}}\cr
    \code{Age} \tab {Age of the respondent (in years)}\cr
    \code{Distance} \tab {Commute distance (in miles)}\cr
    \code{Time} \tab {Commute time (in minutes)}\cr
    \code{Sex} \tab {  \code{F} or \code{M}}\cr
  }
}
\details{
Data from the US Census Bureau's American Housing Survey (AHS) which contains information about housing and living conditions for samples from certain metropolitan areas. 
These data were extracted respondents in the St. Louis metropolitan area. They include only cases where the respondent worked somewhere other than home. 
Values show the time (in minutes) and distance (in miles) that respondents typically traveled on their commute to work each day as well as age and sex.
}
\source{
Sample chosen using DataFerret at http://www.thedataweb.org/index.html.
}

\keyword{datasets}
