\name{CompassionateRats}
\alias{CompassionateRats}
\docType{data}
\title{Compassionate Rats}
\description{
Would a rat attempt to free a trapped rat?
}
%\usage{data(CompassionateRats)}
\format{
  A dataset with 30 observations on the following 2 variables.
  \tabular{rl}{
    \code{Sex} \tab {Sex of the rat: coded as \code{F} or  \code{M}}\cr
    \code{Empathy} \tab {Freed the trapped rat?  \code{no} or  \code{yes}}\cr
  }
}
\details{ 
In a recent study, some rats showed compassion by freeing  another trapped rat,
even when chocolate served as a distraction and even when the rats would then have to share the
chocolate with their freed companion. 
}
\source{
Bartal I.B., Decety J., and Mason P., "Empathy and Pro-Social Behavior in Rats," Science, 2011; 224(6061):1427-1430.
}

\keyword{datasets}
