\name{DDS}
\alias{DDS}
\docType{data}
\title{Develomental Services}
\description{
Funding for individuals by the California Department of Developmental Services (DDS),
}
%\usage{data(DDS)}
\format{
  A dataset with 1000 observations on the following 6 variables.
  \tabular{rl}{
    \code{ID} \tab {ID code for subject}\cr
    \code{AgeCohort} \tab {Age group (\code{0-5}, \code{6-12}, \code{13-17}, \code{18-21}, \code{22-50}, \code{50+})}\cr
    \code{Age} \tab {Age in years}\cr
    \code{Expenditures} \tab {Annual expenditures in dollars}\cr
    \code{Ethnicity} \tab {Ethnic group}\cr
  }
}
\details{
The California Department of Developmental Services (DDS) allocates funds to support developmentally disabled
California residents (such as those with autism, cerebral palsy, or intellectual disabilities) and their
families. We refer to those supported by DDS as DDS consumers. The dataset DDS includes data on annual
expenditure (in $), ethnicity, age, and gender for 1000 DDS consumers.
}
\source{
Taylor, S.A. and Mickel, A. E. (2014). "Simpson's Paradox: A Data Set and Discrimination Case Study Exercise," Journal
of Statistics Education, 22(1). The dataset has been altered slightly for privacy reasons, but is based on actual DDS consumers.
}


\keyword{datasets}
