\name{DietDepression}
\alias{DietDepression}
\docType{data}
\title{Diet and Depression}
\description{
Results from a study of a short-term diet intervention on depression.
}
\format{
  A data frame with 75 observations on the following 10 variables.
  \describe{
    \item{\code{Group}}{\code{Control} or \code{Diet}}
    \item{\code{CESD1}}{CESD depression score on Day 1}
    \item{\code{CESD21}}{CESD depression score on Day 21}
    \item{\code{CESDDiff}}{Change in CESD depression score}
    \item{\code{DASS1}}{DASS depression score on Day 1}
    \item{\code{DASS21}}{DASS depression score on Day 21}
    \item{\code{DASSDiff}}{Change in DASS depression score}
    \item{\code{BMI1}}{Body Mass Index on Day 1}
    \item{\code{BMI21}}{Body Mass Index on Day 21}
    \item{\code{BMIDiff}}{Change in Body Mass Index}
  }
}
\details{
A group of researchers in Australia conducted a short (three-week) dietary intervention in a randomized controlled experiment. In the study, 75 college-age students with elevated depression symptoms and relatively poor diet habits were randomly assigned to either a healthy diet intervention group or a control group.  The researchers recorded the change over the three-week period on two different numeric scales of depression (the CESD scale and the DASS scale).  The CESD (Centre for Epidemiological Studies Depression) score is based more on clinical observations, while the DASS (Depression, Anxiety, and Stress Scale) depends more on self-reported information.  They also recorded body mass index (BMI) at the start and end of the 21 day period.
}
\source{
Francis HM, et al., "A brief diet intervention can reduce symptoms of depression in young adults - A randomised controlled trial," PLoS ONE, 14(10), October 2019.
}
\keyword{datasets}
