\name{DogOwner}
\alias{DogOwner}
\docType{data}
\title{
Dog/Owner matches
}
\description{
Experiment to match dogs with owners
}
\usage{data(DogOwner)}
\format{
  A data frame with 25 observations on the following variable.
  \describe{
    \item{\code{Match}}{a factor with levels \code{no} \code{yes}}
  }
}
\details{
Each dog-owner pair was viewed by 28 naive undergraduate judges, and the pairing was deemed "correct" (yes) if the majority
of judges (more than 14) chose the correct dog to go with the owner.
}
\source{
Roy and Christenfeld, Do Dogs Resemble their Owners?, Psychological Science, Vol. 15, No. 5, 2004, pp. 361 - 363. 
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(DogOwner)
}
\keyword{datasets}
