\name{DrugResistance}
\alias{DrugResistance}
\docType{data}
\title{
Drug Resistance
}
\description{
Effect on drug resistance by level of treatment in mice.
}
%\usage{data(DrugResistance)}
\format{
  A dataset with 72 observations on the following 5 variables.
  \tabular{rl}{
    \code{Treatment} \tab {\code{Untreated}, \code{Light}, \code{Moderate}, or \code{Aggressive}} \cr
    \code{Weight} \tab {Mouse weight in grams} \cr
    \code{RBC} \tab {Red blood cell density} \cr
    \code{ResistantDensity} \tab {Density of resistant parasites} \cr
    \code{DaysInfectious} \tab {Days infectious with resistant parasites} \cr
  }
}
\details{ In an experiment to study drug resistance in mice, groups of 18 mice were injected with a mixture of drug-resistant and drug-susceptible malaria parasites.
One group received no treatment while the others got limited, moderate, or aggressive amounts of anti-malarial treatment.  The weight and red blood cell density reflect the initial health
of the mice.  Density of resistant parasites and number of days infectious measure the effectiveness of the treatment.
}
\source{
Huijben S, Bell AS, Sim DG, Tomasello D, Mideo N, Day T, Read AF (2013) Aggressive chemotherapy and the selection of drug resistant pathogens. PLoS Pathogens 9(9): e1003578. \cr
\emph{http://dx.doi.org/10.1371/journal.ppat.1003578} \cr
Huijben S, et al., (2013). Data from: Aggressive chemotherapy and the selection of drug resistant pathogens. Dryad Digital\cr
Repository. \emph{http://dx.doi.org/10.5061/dryad.09qc0}
}

\keyword{datasets}
