\name{EducationLiteracy2e}
\alias{EducationLiteracy2e}
\docType{data}
\title{
Education Literacy - 2015
}
\description{
Education spending and literacy rates for countries.
}
%\usage{data(EducationLiteracy2e)}
\format{
  A dataset with 188 observations on the following 3 variables.
  \tabular{rl}{
    \code{Country} \tab {Name of country} \cr
    \code{Education} \tab {Education spending (as a percentage of GDP)} \cr
    \code{Literacy} \tab {Literacy rate} \cr
  }
}
\details{ For each country, we have public spending on education (as a percentage of GDP) and literacy rate (percentage of the population who can read and write). \cr
** From 2e - dataset has been updated for 3e **
}
\source{
Most recent data (as of 2015) for each country obtained from \emph{worldbank.org} and \emph{http:\\www.knoema.com}
}

\keyword{datasets}
