\name{EducationLiteracy3e}
\alias{EducationLiteracy3e}
\docType{data}
\title{Education and  Literacy - 2019}
\description{
Education spending and literacy rates for countries.
}
\format{
  A data frame with 170 observations on the following 4 variables.
  \describe{
    \item{\code{Country}}{Name of country}
    \item{\code{Code}}{Three-letter code for country}
    \item{\code{Education}}{Education spending (as a percentage of GDP)}
    \item{\code{Literacy}}{Literacy rate}
  }
}
\details{
For each country, we have public spending on education (as a percentage of GDP) and literacy rate (percentage of the population who can read and write). \cr
** Updated for 3e (an earlier version is at EducationLiteracy2e).  **
}
\source{
Most recent data (as of 2019) for each country obtained from \url{https://www.worldbank.org/ext/en/home}.
}
\keyword{datasets}
