\name{ElectionMargin1e}
\alias{ElectionMargin1e}
\docType{data}
\title{
Election Margin - 1e
}
\description{
Approval rating and election margin for recent presidential elections
}
%\usage{data(ElectionMargin)}
\format{
  A data frame with 11 observations on the following 5 variables.
  \tabular{rl}{
    \code{Year} \tab {Certain election years from 1940-2004} \cr
    \code{Candidate} \tab {Incumbent US president} \cr
    \code{Approval} \tab {Presidential approval rating at time of election} \cr
    \code{Margin} \tab {Margin of victory/defeat (as a percentage)} \cr
    \code{Result} \tab {Outcome of the election for the incumbent: \code{Lost} or \code{Won}} \cr
  }
}
\details{
Data include US Presidential elections since 1940 in which an incumbent was running for president.
The approval rating for the sitting president is compared to the margin of victory/defeat in the election. \cr
** From 1e - dataset has been updated for 2e **
}
\source{
Silver, Nate, "Approval Ratings and Re-Election Odds", fivethirtyeight.com, posted January 28, 2011.
}

\keyword{datasets}
