\name{ElectionMargin2012}
\alias{ElectionMargin2012}
\docType{data}
\title{
Election Margin 1940-2012
}
\description{
Approval rating and election margin for recent US presidential elections with an incumbent president
}
%\usage{data(ElectionMargin)}
\format{
  A dataset with 12 observations on the following 5 variables.
  \tabular{rl}{
    \code{Year} \tab {Certain election years from 1940-2020} \cr
    \code{Candidate} \tab {Incumbent US president} \cr
    \code{Approval} \tab {Presidential approval rating at time of election} \cr
    \code{Margin} \tab {Margin of victory/defeat (as a percentage)} \cr
    \code{Result} \tab {Outcome of the election for the incumbent: \code{Lost} or \code{Won}} \cr
  }
}
\details{
Data include US Presidential elections since 1940 in which an incumbent was running for president.
The approval rating for the sitting president is compared to the margin of victory/defeat in the election. \cr

}
\source{
Silver, Nate, "Approval Ratings and Re-Election Odds", \emph{fivethirtyeight.com}, posted January 28, 2011 and \emph{http:\\realclearpolitics.org}
}

\keyword{datasets}
