\name{ExerciseHours}
\alias{ExerciseHours}
\docType{data}
\title{Exercise Hours}
\description{
Amount of exercise per week for students (and other variables)
}
\format{
  A data frame with 50 observations on the following 7 variables.
  \describe{
    \item{\code{Year}}{Year in school (1=First year,..., 4=Senior)}
    \item{\code{Sex}}{\code{F} or \code{M}}
    \item{\code{Hand}}{Left (\code{l}) or Right (\code{r}) handed?}
    \item{\code{Exercise}}{Hours of exercise per week}
    \item{\code{TV}}{Hours of TV viewing per week}
    \item{\code{Pulse}}{Resting pulse rate (beats per minute)}
    \item{\code{Pierces}}{Number of body piercings}
  }
}
\details{
Data from an in-class survey of statistics students asking about amount of exercise,
TV viewing, handedness, sex, pulse rate, and number of body piercings. Note the Sex variable was labeled as Gender in earlier versions of this dataset. We acknowledge that this binary dichotomization is not a complete or inclusive representation of reality.
}
\source{
In-class student survey.
}
\keyword{datasets}
