\name{FacebookFriends}
\alias{FacebookFriends}
\docType{data}
\title{Facebook Friends}
\description{
Data on number of Facebook friends and grey matter density in brain regions 
related to social perception and associative memory.
}
%\usage{data(FacebookFriends)}
\format{
  A dataset with 40 observations on the following 2 variables.
  \tabular{rl}{
    \code{GMdensity} \tab {Normalized z-scores of grey matter density in certain brain regions}\cr
    \code{FBfriends} \tab {Number of friends on Facebook}\cr
  }
}
\details{
A recent study in Great Britain examines the relationship between the number of friends an individual
has on Facebook and grey matter density in the areas of the brain associated with social perception
and associative memory. The study included 40 students at City University London.
}
\source{
Kanai, R., Bahrami, B., Roylance, R., and Rees, G., "Online social network size is reflected in human brain structure,"
Proceedings of the Royal Society, 7 April 2012; 279(1732): 1327-1334. Data approximated from information in the article.
}

\keyword{datasets}
