\name{FatMice18}
\alias{FatMice18}
\docType{data}
\title{Fat Mice 18}
\description{
Weight gain for mice with different nighttime light conditions
}
%\usage{data(FatMice18)}
\format{
  A dataset with 18 observations on the following 2 variables.
  \tabular{rl}{
    \code{Light} \tab {Light treatment: \code{LD}= normal light/dark cycle OR \code{LL}=bright light at night}\cr
    \code{WgtGain4} \tab {Weight gain (grams over a four week period)}\cr
  }
}
\details{
This is a subset of the LightatNight dataset, showing body mass gain in mice after 4 weeks for two of the treatment conditions: 
a normal light/dark cycle (LD) or a bright light on at night (LL). 
}
\source{
Fonken, L., et. al., "Light at night increases body mass by shifting time of food intake," Proceedings of the National
Academy of Sciences, October 26, 2010; 107(43): 18664-18669.
}

\keyword{datasets}
