\name{FireAnts}
\alias{FireAnts}
\docType{data}
\title{Fire Ants}
\description{
Reactions of lizards to the presence of fire ants.
}
%\usage{data(FireAnts)}
\format{
  A dataset with 80 observations on the following 3 variables.
  \tabular{rl}{
    \code{Invasion} \tab {Coded as \code{Uninvaded} or \code{Invaded}, depending on if the lizard comes from a region with fire ants}\cr
    \code{Twitches} \tab {Number of twitches the lizard makes when encountering fire ants}\cr
    \code{Flee} \tab {Time for the lizard to flee in seconds (more than one minute is recorded as 61).}\cr
  }
}
\details{
The red imported fire ant, \emph{Solenopsis invicta}, is native to South America, but has an expansive invasive range, including much of the southern United States (invasion of this ant is predicted to go global). In the United States, these ants occupy similar habitats as fence lizards. The ants eat the lizards and the lizards eat the ants, and in either scenario the venom from the fire ant can be fatal to the lizard. The study explored the question of whether lizards learn to adapt their behavior if their environment has been invaded by fire ants by taking lizards from an uninvaded habitat (eastern Arkansas) and lizards from an invaded habitat (southern Alabama, which has been invaded for more than 70 years), exposing them to fire ants, and measuring how long it takes each lizard to flee and the number of twitches each lizard does.
}
\source{
Langkilde, T. (2009). "Invasive fire ants alter behavior and morphology of native lizards"", Ecology, 90(1): 208-217.
Thanks to Dr. Langkilde for providing the data.
}

\keyword{datasets}
