\name{FisherIris}
\alias{FisherIris}
\docType{data}
\title{Fisher's Iris Data}
\description{
Measurements of three iris species
}
%\usage{data(FisherIris)}
\format{
  A dataset with 150 observations on the following 5 variables.
  \tabular{rl}{
    \code{Type} \tab {Species of iris, \code{Setosa}, \code{Virginica}, or \code{Versicolor} }\cr
    \code{PetalLength} \tab {Petal length in mm.}\cr
    \code{PetalWidth} \tab {Petal width in mm.}\cr
    \code{SepalLength} \tab {Sepal length in mm.}\cr
    \code{SepalWidth} \tab {Sepal width in mm.}\cr
  }
}
\details{
Data used in Fisher's 1936 paper, this famous dataset looks at measurements for samples of three different species of iris.  The petal is part of the flower itself and the sepals are green leaves, directly under the petals, providing support.
}

\source{
R. A. Fisher (1936). "The use of multiple measurements in taxonomic problems". Annals of Eugenics 7 (2): 179–188. doi:10.1111/j.1469-1809.1936.tb02137.x.
}

\keyword{datasets}
