\name{FishGills12}
\alias{FishGills12}
\docType{data}
\title{
Fish Respiration and Calcium - Full Data
}
\description{
An experiment to look at fish respiration rates in water with different levels of calcium.
}
%\usage{data(FishGills12)}
\format{
  A dataset with 360 observations on the following 2 variables.
  \tabular{rl}{
    \code{Calcium} \tab {Amount of calcium in the water (mg/L)} \cr
    \code{GillRate} \tab {Respiration rate (beats per minute)} \cr
  }
}
\details{
Fish were randomly assigned to twelve tanks with different levels (measured in mg/L)  of calcium.
Respiration rate was measured as number of gill beats per minute.
}
\source{
Thanks to Prof. Brad Baldwin for supplying the data.
}

\keyword{datasets}
