\name{FishGills3}
\alias{FishGills3}
\docType{data}
\title{
Fish Respiration and Calcium
}
\description{
Respiration rate for fish in three levels of calcium.
}
%\usage{data(FishGills3)}
\format{
  A dataset with 90 observations on the following 2 variables.
  \tabular{rl}{
    \code{Calcium} \tab {Level of calcium \code{Low} 0.71 mg/L, \code{Medium} 5.24 mg/L, or \code{High} 18.24 mg/L} \cr
    \code{GillRate} \tab {Respiration rate (beats per minute)} \cr
  }
}
\details{
Fish were randomly assigned to three tanks with different levels (low, medium and high) levels of calcium.
Respiration rate was measured as number of gill beats per minute.
}
\source{
Thanks to Prof. Brad Baldwin for supplying the data.
}

\keyword{datasets}
