\name{Flight179}
\alias{Flight179}
\docType{data}
\title{
Flight times
}
\description{
Flight times for Flight 179 (Boston-SF) and Flight 180 (SF-Boston).
}
%\usage{data(Flight179)}
\format{
  A dataset with 36 observations on the following 3 variables.
  \tabular{rl}{
    \code{Date} \tab {Date of the flight (5th, 15th and 25th of each month in 2010} \cr
    \code{Flight179} \tab {Flying time (Boston-SF) in minutes} \cr
    \code{Flight180} \tab {Flying time (SF-Boston) in minutes} \cr
  }
}
\details{
United Airlines Flight 179 was a daily flight from Boston to San Francisco.
Flight 180 goes in the other direction (SF to Boston). The data show the airborne flying times
for each flight on the three dates each month (5th, 15th and 25th) in 2010. \cr
** In first edition, but not in 2e - replaced by Flight433 **
}
\source{
Data collected from the Bureau of Transportation Statistics website at \cr
\emph{http://www.bts.gov/xml/ontimesummarystatistics/src/dstat/OntimeSummaryAirtime.xml}
}

\keyword{datasets}
