\name{Flight433}
\alias{Flight433}
\docType{data}
\title{
Flight 433
}
\description{
Flight times for Flight 433 (Boston-SF) in January 2016.
}
%\usage{data(Flight433)}
\format{
  A dataset with 31 observations on the following 1 variable.
  \tabular{rl}{
    \code{Airtime} \tab {Airborne flying fime (in  minutes) for Flight 433, Boston to San Francisco} \cr
  }
}
\details{
United Airlines Flight 433 was a daily flight from Boston to San Francisco.
The data show the airborne flying times
for the flight on each day of January 2016. \cr
** New to second edition, replaces Flight179 **
}
\source{
Data collected from the Bureau of Transportation Statistics website at \cr
\emph{http://www.bts.gov/xml/ontimesummarystatistics/src/dstat/OntimeSummaryAirtime.xml}
}

\keyword{datasets}
