\name{Flight433_2e}
\alias{Flight433_2e}
\docType{data}
\title{
Flight 433 - 2e
}
\description{
Flight times for Flight 433 (Boston-SF) in January 2016.
}
%\usage{data(Flight433_2e)}
\format{
  A dataset with 31 observations on the following 1 variable.
  \tabular{rl}{
    \code{Airtime} \tab {Airborne flying time (in  minutes) for Flight 433, Boston to San Francisco} \cr
  }
}
\details{
United Airlines Flight 433 was a daily flight from Boston to San Francisco.
The data show the airborne flying times
for the flight on each day of January 2016. \cr
** From 2e - dataset has been updated for 3e **
}
\source{
Data collected from the Bureau of Transportation Statistics website at \cr
\emph{http://www.bts.gov/xml/ontimesummarystatistics/src/dstat/OntimeSummaryAirtime.xml}
}

\keyword{datasets}
