\name{Flight433_3e}
\alias{Flight433_3e}
\docType{data}
\title{Flight 433 - 3e}
\description{
Flight times for Flight 433 (Boston-SF) in January 2019.
}
\format{
  A data frame with 28 observations on the following variable.
  \describe{
    \item{\code{AirTime}}{Airborne flying time (in  minutes) for Flight 433, Boston to San Francisco}
  }
}
\details{
United Airlines Flight 433 was a daily flight from Boston to San Francisco.
The data show the airborne flying times
for the flight on each day of January 2019. \cr
**Updated for 3e (earlier version from 2016 is in Flight433_2e) **
}
\source{
Data collected from the Bureau of Transportation Statistics website at
\url{https://www.transtats.bts.gov/}
}
\keyword{datasets}
