\name{Flight475}
\alias{Flight475}
\docType{data}
\title{Flight 475}
\description{
Flight times for Delta Flight 475 (Boston-SF) in October 2024.
}
\format{
  A data frame with 31 observations on the following variable.
  \describe{
    \item{\code{AirTime}}{Airborne flying time (in  minutes) for Flight 475, Boston to San Francisco}
  }
}
\details{
Delta Airlines Flight 475 was a daily flight from Boston to San Francisco.
The data show the airborne flying times
for the flight on each day of October 2024. \cr
**Updated for 4e (earlier version are Flight433_3e (2019) and Flight433_2e (2016)) **
}
\source{
Data collected from the Bureau of Transportation Statistics website at
\url{https://www.transtats.bts.gov/}
}
\keyword{datasets}
