\name{FootballBrain}
\alias{FootballBrain}
\docType{data}
\title{
Football Brain Measurements
}
\description{
Brain measurements for non-football players, football players with no concussion history, and football players with a concussion history.
}
%\usage{data(FootballBrain)}
\format{
  A dataset with 75 observations on the following 5 variables.
  \tabular{rl}{
    \code{Group} \tab {\code{Control}=no football, \code{FBNoConcuss}=football player but no concussions,}\cr
    \code{ } \tab {or \code{FBConcuss}=football player with concussion history} \cr
    \code{Hipp} \tab {Total hippocampus volume, in microL} \cr
    \code{LeftHipp} \tab {Left hippocampus volume, in microL} \cr
    \code{Years} \tab {Number of years playing football} \cr
    \code{Cognition} \tab {Cognitive testing composite reaction time score, given as a percentile } \cr
  }
}
\details{
The study included 3 groups, with 25 cases in each group.  The control group consisted of healthy individuals with no history of brain trauma who were comparable to the other groups in age, sex, and education.  The second group consisted of NCAA Division 1 college football players with no history of concussion, while the third group consisted of NCAA Division 1 college football players with a history of concussion.
High resolution MRI was used to collect brain hippocampus volume.  Data were collected between June 2011 and August 2013.  The data values given here are estimated from information given in the paper.
}
\source{
Singh R, Meier T, Kuplicki R, Savitz J, et al.,  "Relationship of Collegiate Football Experience and Concussion With Hippocampal Volume and Cognitive Outcome,"  JAMA, 311(18), 2014
}

\keyword{datasets}
