\name{GeneticDiversity}
\alias{GeneticDiversity}
\docType{data}
\title{
Genetic Diversity
}
\description{
Genetic diversity for different populations are compared to the distance from East Africa.
}
%\usage{data(GeneticDiversity)}
\format{
  A dataset with 52 observations on the following 5 variables.
  \tabular{rl}{
    \code{Population} \tab {Identifier for each population} \cr
    \code{Country} \tab {Main country where the population is found} \cr
    \code{Continent} \tab {Continent where the population is found} \cr
    \code{GeneticDiversity} \tab {A measure of genetic diversity in the population} \cr
    \code{Distance} \tab {Distance by land to East Africa (in km)} \cr
  }
}
\details{
The data give a measure of genetic diversity for different populations and the geographic distance of each population
from East Africa (Addis Ababa, Ethiopia), as one would travel over the surface of the earth by land (migration long ago is thought to have happened by land).
}
\source{
Calculated using data from S Ramachandran, O Deshpande, CC Roseman, NA Rosenberg, MW Feldman, LL Cavalli-Sforza.
"Support from the relationship of genetic and geographic distance in human populations for a serial founder effect originating
in Africa,"" Proceedings of the National Academy of Sciences, 2005, 102: 15942-15947.
}

\keyword{datasets}
