\name{GlobalInternet}
\alias{GlobalInternet}
\docType{data}
\title{
Global Internet Usage
}
\description{
Internet usage for several countries
}
\usage{data(GlobalInternet)}
\format{
  A data frame with 9 observations on the following 3 variables.
  \describe{
    \item{\code{Country}}{Name of country}
    \item{\code{PercentFastConnection}}{Percent of internet users with a fast connection}
    \item{\code{HoursOnline}}{Average number of hours online in February 2011}
  }
}
\details{
The Nielsen Company measured connection speeds on home computers in nine different countries.  Variables include the  percent of internet users with a fast connection (defined as 2Mb/sec
or faster) and the average amount of time spent online, defined as total hours connected to the web from a home computer during the month of February 2011. 
}
\source{
NielsenWire, "Swiss Lead in Speed: Comparing Global Internet Connections", April 1, 2011
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%data(GlobalInternet)
}
\keyword{datasets}
