\name{GlobalInternet2019}
\alias{GlobalInternet2019}
\docType{data}
\title{Global Internet Usage - 2019}
\description{
Internet usage for several countries
}
\format{
  A data frame with 9 observations on the following 3 variables.
  \describe{
    \item{\code{Country}}{Name of country}
    \item{\code{InternetSpeed}}{Average download speed (in Mb)}
    \item{\code{HoursOnline}}{Average hours online per day}
  }
}
\details{
The Worldwide Broadband Speed League tests internet speeds at millions of access points around the world. The average download speed for each country is derived from those data. The DataReportal site provides summaries of country level data on internet usage obtained from various sources.  The average number of hours spent online for each country is based on survey data reported at that site.  \cr
** Updated for 3e (earlier version from 2011 is at GlobalInternet2011).
}
\source{
Internet speeds for 2019 downloaded from \url{https://bestbroadbanddeals.co.uk/broadband/speed/worldwide-speed-league/} \cr
Online hours for 2019 downloaded from \url{https://datareportal.com/library}
}
\keyword{datasets}
