\name{GlobalInternet2024}
\alias{GlobalInternet2024}
\docType{data}
\title{Global Internet Usage - 2024}
\description{
Internet usage for several countries in 2024
}
\format{
  A data frame with 9 observations on the following 3 variables.
  \describe{
    \item{\code{Country}}{Name of country}
    \item{\code{InternetSpeed}}{Average download speed (in Mb)}
    \item{\code{HoursOnline}}{Average hours online per day for internet users}
  }
}
\details{
The Worldwide Broadband Speed League tests internet speeds at millions of access points around the world. The average download speed for each country is derived from those data. The DataReportal site provides summaries of country level data on internet usage obtained from various sources.  The average number of hours spent online for each country is based on survey data reported at that site.  \cr
** Updated for 4e (earlier versions are GlobalInternet2019 and GlobalInternet2011).
}
\source{
Internet speeds for 2024 downloaded from \url{https://bestbroadbanddeals.co.uk/broadband/speed/worldwide-speed-league/} \cr
Online hours for 2024 downloaded from \url{https://datareportal.com/library}
}
\keyword{datasets}
