\name{GolfRound}
\alias{GolfRound}
\docType{data}
\title{Golf Round}
\description{
Scorecard for 18 holes of golf
}
\format{
  A data frame with 18 observations on the following 4 variables.
  \describe{
    \item{\code{Hole}}{Hole number (1 to 18)}
    \item{\code{Distance}}{Length of the hole (in yards)}
    \item{\code{Par}}{Par for the hole}
    \item{\code{Score}}{Actual number of stokes needed in this round}
  }
}
\details{
Data come from a scorecard for one round of golf at the Potsdam Country Club.  Par is the expected number of strokes a good golfer should need to complete the hole.
}
\source{
Personal file
}
\keyword{datasets}
