\name{GPAGender}
\alias{GPAGender}
\docType{data}
\title{GPA and Gender}
\description{
Data from a survey of introductory statistics students.
}
%\usage{data(GPAGender)}
\format{
  A dataset with 343 observations on the following 6 variables.
  \tabular{rl}{
    \code{Exercise} \tab {Hours of exercise (per week)}\cr
    \code{SAT} \tab {Combined SAT scores (out of 1600)}\cr
    \code{GPA} \tab {Grade Point Average (0.00-4.00 scale)}\cr
    \code{Pulse} \tab {Pulse rate (beats per minute)}\cr
    \code{Piercings} \tab {Number of body piercings}\cr
    \code{GenderCode} \tab {0=female ir 1=male}\cr
  }
}
\details{
This is a subset of the StudentSurvey dataset where cases with missing values have been dropped 
and gender is coded as a 0/1 indicator variable. 
}
\source{
A first day survey over several different introductory statistics classes. 
}

\keyword{datasets}
