\name{GroundhogDay}
\alias{GroundhogDay}
\docType{data}
\title{Groundhog Day}
\description{
  Yearly data on US March temperature and Puxsutawney Phil's forecast on Groundhog Day
}
\format{
  A data frame with 122 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (1903-2025) }
    \item{\code{Shadow}}{Phil saw his shadow? (\code{Yes} or \code{No})}
    \item{\code{USMarchTemp}}{Average US temperature in March of that year}
   }
}
\details{
Every February 2nd (Groundhog Day) people gather at Gobbler's Knob in Puxsutawney, Pennsylvannia to see if an emerging groundhog (Phil) casts a shadow. Legend has it that a shadow fortells six more weks of winter, but no shadow indicates an early coming of spring.  This dataset shows the outcome of Phil's shadow viewing for each year from 1903 to 2025, along with the average temerature in the US that March.  One might expect six more week's of winter would mean colder temperatures in March.
}
\source{
Data through 2016 from  \url{https://www.kaggle.com/datasets/groundhogclub/groundhog-day} (May 2025).\\

More recent years from \url{https://www.groundhog.org/} and \url{https://www.ncei.noaa.gov/}.
}
\keyword{datasets}
