\name{GSWarriors2016}
\alias{GSWarriors2016}
\docType{data}
\title{
Golden State Warriors Basketball - 2016
}
\description{
Game log data for the Golden State Warriors basketball team in 2015-2016
}
%\usage{data(GSWarriors2016)}
\format{
  A dataset with 82 observations on the following 33 variables.
  \tabular{rl}{
    \code{Game} \tab {ID number for each game} \cr
    \code{Date} \tab {Date the game was played} \cr
    \code{Location} \tab {\code{Away}  or \code{Home}} \cr
    \code{Opp} \tab {Opponent team} \cr
    \code{Win} \tab {Game result: \code{L} or \code{W}} \cr
    \code{FG} \tab {Field goals made} \cr
    \code{FGA} \tab {Field goals attempted} \cr
    \code{FG3} \tab {Three-point field goals made} \cr
    \code{FG3A} \tab {Three-point field goals attempted} \cr
    \code{FT} \tab {Free throws made} \cr
    \code{FTA} \tab {Free throws attempted} \cr
    \code{Rebounds} \tab {Total rebounds} \cr
    \code{OffReb} \tab {Offensive rebounds} \cr
    \code{Assists} \tab {Number of assists} \cr
    \code{Steals} \tab {Number of steals} \cr
    \code{Blocks} \tab {Number of shots blocked} \cr
    \code{Turnovers} \tab {Number of turnovers} \cr
    \code{Fouls} \tab {Number of fouls} \cr
    \code{Points} \tab {Number of points scored} \cr
    \code{OppFG} \tab {Opponent's field goals made} \cr
    \code{OppFGA} \tab {Opponent's Field goals attempted} \cr
    \code{OppFG3} \tab {Opponent's Three-point field goals made} \cr
    \code{OppFG3A} \tab {Opponent's Three-point field goals attempted} \cr
    \code{OppFT} \tab {Opponent's Free throws made} \cr
    \code{OppFTA} \tab {Opponent's Free throws attempted} \cr
    \code{OppRebounds} \tab {Opponent's Total rebounds} \cr
    \code{OppOffReb} \tab {Opponent's Offensive rebounds} \cr
    \code{OppAssists} \tab {Opponent's assists} \cr
    \code{OppSteals} \tab {Opponent's steals} \cr
    \code{OppBlocks} \tab {Opponent's shots blocked} \cr
    \code{OppTurnovers} \tab {Opponent's turnovers} \cr
    \code{OppFouls} \tab {Opponent's fouls} \cr
    \code{OppPoints} \tab {Opponent's points scored} \cr
  }
}
\details{
Information from online boxscores for all 82 regular season games played by the Golden State Warriors basketball team during the 2015-2016 season. \cr
** From 2e - dataset has been updated for 3e **
}
\source{
Data for the 2015-2016 Golden State games downloaded from \cr \emph{http://www.basketball-reference.com/teams/GSW/2016/gamelog/}
}

\keyword{datasets}
