\name{GSWarriors2019}
\alias{GSWarriors2019}
\docType{data}
\title{Golden State Warriors Basketball (2019)}
\description{
Game log data for the Golden State Warriors basketball team in 2018-2019
}
\format{
  A data frame with 82 observations on the following 33 variables.
  \describe{
    \item{\code{Game}}{ID number for each game}
    \item{\code{Date}}{Date the game was played (mm/dd/yyy)}
    \item{\code{Location}}{\code{Away}  or \code{Home}}
    \item{\code{Opp}}{Opponent team}
    \item{\code{Win}}{Game result: \code{L} or \code{W}}
    \item{\code{Points}}{Number of points scored}
    \item{\code{FG}}{Field goals made}
    \item{\code{FGA}}{Field goals attempted}
    \item{\code{FG3}}{Three-point field goals made}
    \item{\code{FG3A}}{Three-point field goals attempted}
    \item{\code{FT}}{Free throws made}
    \item{\code{FTA}}{Free throws attempted}
    \item{\code{Rebounds}}{Total rebounds}
    \item{\code{OffReb}}{Offensive rebounds}
    \item{\code{Assists}}{Number of assists}
    \item{\code{Steals}}{Number of steals}
    \item{\code{Blocks}}{Number of shots blocked}
    \item{\code{Turnovers}}{Number of turnovers}
    \item{\code{Fouls}}{Number of fouls}
    \item{\code{OppPoints}}{Opponent's points scored}
    \item{\code{OppFG}}{Opponent's field goals made}
    \item{\code{OppFGA}}{Opponent's field goals attempted}
    \item{\code{OppFG3}}{Opponent's three-point field goals made}
    \item{\code{OppFG3A}}{Opponent's three-point field goals attempted}
    \item{\code{OppFT}}{Opponent's free throws made}
    \item{\code{OppFTA}}{Opponent's free throws attempted}
    \item{\code{OppRebounds}}{Opponent's total rebounds}
    \item{\code{OppOffReb}}{Opponent's offensive rebounds}
    \item{\code{OppAssists}}{Opponent's assists}
    \item{\code{OppSteals}}{Opponent's steals}
    \item{\code{OppBlocks}}{Opponent's shots blocked}
    \item{\code{OppTurnovers}}{Opponent's turnovers}
    \item{\code{OppFouls}}{Opponent's fouls}
  }
}
\details{
Information from online boxscores for all 82 regular season games played by the Golden State Warriors basketball team during the 2018-2019 season. \cr
** Updated for third edition (2e version is now GSWarriors2016, 1e version is MiamiHeat dataset)  **
}
\source{
Data for the 2018-2019 Golden State games downloaded from
\url{https://www.basketball-reference.com/teams/GSW/2019/gamelog/}
}
\keyword{datasets}
