\name{HappyPlanetIndex}
\alias{HappyPlanetIndex}
\docType{data}
\title{Happy Planet Index}
\description{
Measurements related to happiness and well-being for 147 countries.
}
%\usage{data(HappyPlanetIndex)}
\format{
  A dataset with 147 observations on the following 9 variables.
 \tabular{rl}{
    \code{Country} \tab {Name of country}\cr
    \code{Region} \tab {\code{1}=Latin America, \code{2}=N. America and Oceania,
         \code{3}=Western Europe, \code{4}=Middle East and N. Africa, } \cr
     \tab {\code{5}=Sub-Saharan Africa, \code{6}=South Asia,
          \code{7}=Eastern Europe and Central Asia, \code{8}=East Asia}\cr
    \code{HPI} \tab {Happy Planet Index (0-100 scale)}\cr
    \code{HPIRank} \tab {HPI rank for the country}\cr
    \code{LifeExpectancy} \tab {Average life expectancy (in years)}\cr
    \code{Footprint} \tab {Ecological footprint - a measure of the (per capita) ecological impact}\cr
    \code{LOL} \tab {Ladder of Life (a measure of wellbeing on 0-10 scale)}\cr
    \code{GDPperCapita} \tab {Gross Domestic Product (per capita)}\cr
    \code{Population} \tab {Population (in millions)}\cr
  }
}
\details{
Data for 147 countries from the Happy Planet Index Project that works
to quantify indicators of happiness, well-being, and ecological footprint at a country level. \cr
** Updated for 4e (earlier version is now HappyPlanetIndex2010) **
}
\source{
 Abdallah, S. & Hoffman, A. (2024)  The Happy Planet Index 2024 Data File. \cr
 Accessed from \url{https://happyplanetindex.org/} March 2025
}

\keyword{datasets}
