\name{HappyPlanetIndex2010}
\alias{HappyPlanetIndex2010}
\docType{data}
\title{Happy Planet Index 2010}
\description{
Measurements related to happiness and well-being for 143 countries.
}
%\usage{data(HappyPlanetIndex)}
\format{
  A dataset with 143 observations on the following 11 variables.
 \tabular{rl}{
    \code{Country} \tab {Name of country}\cr
    \code{Region} \tab {\code{1}=Latin America, \code{2}=Western nations, \code{3}=Middle East, \code{4}=Sub-Saharan Africa, } \cr
     \tab {\code{5}=South Asia, \code{6}=East Asia, \code{7}=former Communist countries}\cr
    \code{Happiness} \tab {Score on a 0-10 scale for average level of happiness (10 is happiest)}\cr
    \code{LifeExpectancy} \tab {Average life expectancy (in years)}\cr
    \code{Footprint} \tab {Ecological footprint - a measure of the (per capita) ecological impact}\cr
    \code{HLY} \tab {Happy Life Years - combines life expectancy with well-being}\cr
    \code{HPI} \tab {Happy Planet Index (0-100 scale)}\cr
    \code{HPIRank} \tab {HPI rank for the country}\cr
    \code{GDPperCapita} \tab {Gross Domestic Product (per capita)}\cr
    \code{HDI} \tab {Human Development Index}\cr
    \code{Population} \tab {Population (in millions)}\cr
  }
}
\details{
Data for 143 countries from the Happy Planet Index Project that works
to quantify indicators of happiness, well-being, and ecological footprint at a country level.
}
\source{
Marks, N., "The Happy Planet Index", www.TED.com/talks, August 29, 2010. \cr
Data downloaded from \emph{http://www.happyplanetindex.org/data/}
}

\keyword{datasets}
