\name{HeatCognition}
\alias{HeatCognition}
\docType{data}
\title{Heat and Cognition}
\description{
Effect of heat on cognitive ability
}
\format{
  A data frame with 46 observations on the following 3 variables.
  \describe{
    \item{\code{AC}}{Whether the student had air conditioning on in the room, \code{No} or \code{Yes}}
    \item{\code{MathZRT}}{Z-score of reaction time solving math problems}
    \item{\code{ColorsZRT}}{Z-score of reaction time solving STROOP color problems}
  }
}
\details{
Forty-six college students were asked to solve cognitive problems first thing in the morning during a heat wave in their Northeastern city.  Twenty of the students had air-conditioning in their rooms and twenty-six did not.  Z-scores of reaction times are given for math problems and for color dissonance problems.
}
\source{
Cedeo Laurent JG, Williams A, Oulhote Y, Zanobetti A, Allen JG, Spengler JD
"Reduced cognitive function during a heat wave among residents of non-air-conditioned buildings: An observational study of young adults in the summer of 2016." PLoS Med 15(7): e1002605, July 10, 2018. \url{https://journals.plos.org/plosmedicine/article?id=10.1371/journal.pmed.1002605}.  (Dataset is simplified from the repeated measures design used in the original study.)
}
\keyword{datasets}
