\name{HeightData}
\alias{HeightData}
\docType{data}
\title{Height Data}
\description{
Heights measured for the same 94 children over 18 years.
}
%\usage{data(HeightData)}
\format{
  A dataset with 94 observations on the following 33 variables.
  \tabular{rl}{
    \code{ID} \tab {Identification number)}\cr
    \code{Sex} \tab {\code{M} or \code{F}}\cr
    \code{Year_1} \tab {Height (in cm.) at age 1 year}\cr
    \code{Year_1.25} \tab {Height (in cm.) at age 1.25 years}\cr
    \code{Year_1.5} \tab {Height (in cm.) at age 1.5 years}\cr
    \code{Year_1.75} \tab {Height (in cm.) at age 1.75 years}\cr
    \code{Year_2} \tab {Height (in cm.) at age 2 years}\cr
    \code{Year_3} \tab {Height (in cm.) at age 3 years}\cr
    \code{Year_4} \tab {Height (in cm.) at age 4 years}\cr
    \code{Year_5} \tab {Height (in cm.) at age 5 years}\cr
    \code{      }\tab {See below for full list of years...}\cr
    \code{Year_17.5} \tab {Height (in cm.) at age 17.5 years}\cr
    \code{Year_18} \tab {Height (in cm.) at age 18 years}\cr
  }
}
\details{
In the 1940's and 1950's, the heights of 39 boys and 54 girls, in centimeters, were measured at 30 different
time points between the ages of 1 and 18 years as part of the University of California Berkeley growth
study.  Ages for measurement are 1, 1,25, 1,5, 1,75, 2, 3, 4, 5, 6, 7, 8, 8.5, 9, 9.5, 10, 10.5, 11, 11,5, 12, 12.5, 13, 13.5, 14, 14.5, 15, 15.5, 16, 16.5, 17, 17.5, 18.
}
\source{
Tuddenham, R. D., and Snyder, M. M. (1954) "Physical growth of California boys and girls from birth to age 18", University of California Publications in Child Development, 1, 183-364.
}

\keyword{datasets}
