\name{HockeyPenalties}
\alias{HockeyPenalties}
\docType{data}
\title{
Hockey Penalties
}
\description{
Penalty minutes (per game) for NHL teams in 2010-11
}
\usage{data(HockeyPenalties)}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{Team}}{Name of the team}
    \item{\code{PIMperG}}{Average penaly minutes per game}
  }
}
\details{
Data give the average numeber of penalty minutes for each of the 30 National HOckey League (NHL) teams during the 2010-11 regualar season. 
}
\source{
Data obtained online at www.nhl.com
}
\references{
%%
}
\examples{
%%data(HockeyPenalties)
}
\keyword{datasets}
