\name{HockeyPenalties2011}
\alias{HockeyPenalties2011}
\docType{data}
\title{
Hockey Penalties - 2011
}
\description{
Penalty minutes (per game) for NHL teams in 2010-11
}
%\usage{data(HockeyPenalties2011)}
\format{
  A dataset with 30 observations on the following 2 variables.
  \tabular{rl}{
    \code{Team} \tab {Name of the team}\cr
    \code{PIMperG} \tab {Average penalty minutes per game} \cr
  }
}
\details{
Data give the average number of penalty minutes for each of the 30 National Hockey League (NHL) teams during the 2010-11 regular season.\cr
** From 2e - dataset has been updated for 3e **
}
\source{
Data obtained online at \emph{www.nhl.com}
}

\keyword{datasets}
