\name{HockeyPenalties2019}
\alias{HockeyPenalties2019}
\docType{data}
\title{Hockey Penalties (2019)}
\description{
Penalty minutes (per game) for NHL teams in 2018-2019
}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{Team}}{Name of the team}
    \item{\code{PIM}}{Average penalty minutes per game}
    \item{\code{OppPIM}}{Average opponent's penalty minutes per game}
    \item{\code{Playoff}}{Did the team make the playoffs? (\code{N} or \code{Y})}
  }
}
\details{
Data give the average number of penalty minutes for each of the 30 National Hockey League (NHL) teams (and their opponents) during the 2018-2019 regular season. \cr
** Updated for 3e (earlier version from 2010-11 is at HockeyPenalties2011). **
}
\source{
Data obtained online at \url{https://www.hockey-reference.com/leagues/NHL_2019.html#all_stats}
}
\keyword{datasets}
