\name{HollywoodMovies2011}
\alias{HollywoodMovies2011}
\docType{data}
\title{Hollywood Movies in 2011}
\description{
Data on movies released in Hollywood in 2011
}
%\usage{data(HollywoodMovies2011)}
\format{
  A dataset with 136 observations on the following 14 variables.
  \tabular{rl}{
    \code{Movie} \tab {Title of movie} \cr
    \code{LeadStudio} \tab {Studio that released the movie} \cr
    \code{RottenTomatoes} \tab {Rotten Tomatoes rating (reviewers)} \cr
    \code{AudienceScore} \tab {Audience rating (via Rotten Tomatoes)} \cr
    \code{Story} \tab {General theme - one of 21 themes} \cr
    \code{Genre} \tab { \code{Action} \code{Adventure} \code{Animation} \code{Comedy} \code{Drama} \code{Fantasy} \code{Horror} \code{Romance} \code{Thriller}} \cr
    \code{TheatersOpenWeek} \tab {Number of screens for opening weekend} \cr
    \code{BOAverageOpenWeek} \tab {Average opening week box office income (per theater)} \cr
    \code{DomesticGross} \tab {Gross income for domestic viewers (in $ millions)} \cr
    \code{ForeignGross} \tab {Gross income for foreign viewers (in $ millions)} \cr
    \code{WorldGross} \tab {Gross income for all viewers (in $ millions)} \cr
    \code{Budget} \tab {Production budget (in $ millions)} \cr
    \code{Profitability} \tab {WorldGross as a percentage of Budget} \cr
    \code{OpeningWeekend} \tab {Opening weekend gross (in $ millions)} \cr
  }
}
\details{
Information from 136 movies released from Hollywood in 2011. \cr
** This dataset has been updated for 2e with more years of data (in HollywoodMovies) **
}
\source{
McCandless, D., "Most Profitable Hollywood Movies" from "Information is Beautiful" at  \cr
\emph{http://www.informationisbeautiful,net.data/} and \cr
\emph{http://bit.ly/hollywoodbudgets}.
}

\keyword{datasets}
