\name{HomesForSale}
\alias{HomesForSale}
\docType{data}
\title{Homes For Sale (2019)}
\description{
Data on homes for sale in four states in 2019
}
\format{
  A data frame with 120 observations on the following 5 variables.
  \describe{
    \item{\code{State}}{Location of the home (\code{CA}, \code{NJ}, \code{NY}, or \code{PA})}
    \item{\code{Price}}{Asking price (in $1,000's)}
    \item{\code{Size}}{Area of all rooms (in 1,000's sq. ft.)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
  }
}
\details{
Data for samples of homes for sale in each state, selected from \emph{zillow.com}. \cr
** Updated for 3e (earlier version from 2010 is in HomesForSale2e). **
}
\source{
Data collected from \url{https://www.zillow.com/} in 2019.
}
\keyword{datasets}
