\name{HomesForSale2e}
\alias{HomesForSale2e}
\docType{data}
\title{
Home for Sale - 2e
}
\description{
Data on homes for sale in four states
}
%\usage{data(HomesForSale2e)}
\format{
  A dataset with 120 observations on the following 5 variables.
  \tabular{rl}{
    \code{State} \tab {Location of the home: \code{CA} \code{NJ} \code{NY} \code{PA}} \cr
    \code{Price} \tab {Asking price (in $1,000's)} \cr
    \code{Size} \tab {Area of all rooms (in 1,000's sq. ft.)}\cr
    \code{Beds} \tab {Number of bedrooms}\cr
    \code{Baths} \tab {Number of bathrooms}\cr
  }
}
\details{
Data for samples of homes for sale in each state, selected from \emph{zillow.com}. \cr
** From 2e - dataset has been updated for 3e **
}
\source{
Data collected from \emph{www.zillow.com} in 2010.
}

\keyword{datasets}
