\name{HomesForSaleCanton}
\alias{HomesForSaleCanton}
\docType{data}
\title{Homes For Sale in Canton, NY (2025)}
\description{
Data for a sample of homes sold in Canton, NY in 2025
}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Price}}{Asking price (in $1,000's)}
    \item{\code{Size}}{Area of all rooms (in sq. ft.)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
  }
}
\details{
Data for a sample of homes recently sold in Canton, NY, selected from \emph{zillow.com}. \cr
** Updated for 4e (earlier versions are HomesForSaleCanton3e (2019) and HomesForSaleCanton2e (2010)). **
}
\source{
Data collected from \emph{https://www.zillow.com/} in March 2025.
}
\keyword{datasets}
