\name{HomesForSaleCanton2e}
\alias{HomesForSaleCanton2e}
\docType{data}
\title{
Homes for sale in Canton, NY - 2e
}
\description{
Prices of homes for sale in Canton, NY
}
%usage{data(HomesForSaleCanton2e)}
\format{
  A dataset with 10 observations on the following variable.
  \tabular{rl}{
    \code{Price} \tab {Asking price for the home (in $1,000's)} \cr
  }
}
\details{
Data for samples of homes for sale in Canton, NY, selected from \emph{zillow.com}.\cr
** From 2e - dataset has been updated for 3e **
}
\source{
Data collected from \emph{www.zillow.com} in 2010.
}

\keyword{datasets}
