\name{HomesForSaleCanton3e}
\alias{HomesForSaleCanton3e}
\docType{data}
\title{Homes For Sale in Canton, NY (2019)}
\description{
Data for a sample of homes offered for sale in Canton, NY
}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Price}}{Asking price (in $1,000's)}
    \item{\code{Size}}{Area of all rooms (in 1,000's sq. ft.)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
  }
}
\details{
Data for a sample of homes for sale in Canton, NY, selected from \emph{zillow.com}. \cr
** Updated for 3e (earlier version from 2010 is in HomesForSaleCanton2e). **
}
\source{
Data collected from \emph{https://www.zillow.com/} in 2019.
}
\keyword{datasets}
