\name{HomesForSaleNY}
\alias{HomesForSaleNY}
\docType{data}
\title{Homes For Sale in New York (2025)}
\description{
Data for a sample of homes sold in New York (state)
}
\format{
  A data frame with 30 observations on the following 6 variables.
  \describe{
    \item{\code{State}}{Location of the home (\code{NY})}
    \item{\code{Price}}{Asking price (in $1,000's)}
    \item{\code{Size}}{Area of all rooms (in sq. ft.)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
    \item{\code{Zip}}{Zip code for the home}
  }
}
\details{
Data for a sample of 30 homes recently sold in New York, selected from random zip codes at \emph{zillow.com}. This is a subset of the HomesForSale dataset. \cr
** Updated for 4e (earlier versions are HomesForSaleNY3e (2019) and HomesForSaleNY2e (2010)). **
}
\source{
Data collected from \emph{https://www.zillow.com/} in 2025.
}
\keyword{datasets}

