\name{HomesForSaleNY3e}
\alias{HomesForSaleNY3e}
\docType{data}
\title{Homes For Sale in New York (2019)}
\description{
Data for a sample of homes offered for sale in New York (state)
}
\format{
  A data frame with 30 observations on the following 5 variables.
  \describe{
    \item{\code{State}}{Location of the home (\code{NY})}
    \item{\code{Price}}{Asking price (in $1,000's)}
    \item{\code{Size}}{Area of all rooms (in 1,000's sq. ft.)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
  }
}
\details{
Data for a sample of homes for sale in New York, selected from \emph{zillow.com}. This is a subset of the HomesForSale dataset. \cr
** Updated for 3e (earlier version from 2010 is in HomesForSaleNY2e). **
}
\source{
Data collected from \emph{https://www.zillow.com/} in 2019.
}
\keyword{datasets}

