\name{HomingPigeons}
\alias{HomingPigeons}
\docType{data}
\title{Homing Pigeons}
\description{
Results from the 2019 Midwest Classic Homing Pigeon race
}
\format{
  A data frame with 1412 observations on the following 5 variables.
  \describe{
    \item{\code{Position}}{Finishing position in the race}
    \item{\code{Loft}}{Name of the pigeon's home loft}
    \item{\code{Sex}}{\code{C}=cock (male) or \code{H}=hen (female)}
    \item{\code{Distance}}{Distance (in miles) from release point to home loft}
    \item{\code{Speed}}{Speed (in yards per minute)}
  }
}
\details{
Finishing results from 1412 pigeons completing the 2019 Midwest Classic race for homing pigeons on June 30, 2019.  Each loft may enter multiple pigeons. 
}
\source{
Final race report from the Midwest Homing Pigeon Association, downloaded from \url{http://www.midwesthpa.com/MIDFinalReports.htm}
}
\keyword{datasets}
