\name{Honeybee}
\alias{Honeybee}
\docType{data}
\title{Honeybee Colonies}
\description{
Number of honeybee colonies (1995-2012)
}
%\usage{data(Honeybee)}
\format{
  A dataset with 18 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year} \tab {Year}\cr
    \code{Colonies} \tab {Estimated number of honeybee colonies in the US (in thousands)}\cr
  }
}
\details{
Data collected from the USDA on the estimated number of honeybee colonies in the US for the years 1995 through 2012.
}
\source{
USDA National Agriculture and Statistical Services, \cr
\emph{http://usda.mannlib.cornell.edu/MannUsda/viewDocumentInfo.do?documentID=1191} Accessed September 2015.
}

\keyword{datasets}
