\name{Honeybee2023}
\alias{Honeybee2023}
\docType{data}
\title{Honeybee Colonies - 2023}
\description{
Number of honeybee colonies (2008-2023)
}
%\usage{data(Honeybee2023)}
\format{
  A dataset with 16 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year} \tab {Year}\cr
    \code{Colonies} \tab {Estimated number of honeybee colonies in the US (in thousands)}\cr
  }
}
\details{
Data collected from the USDA on the estimated number of honeybee colonies in the US for the years 2008 through 2023.
Updated for 4e (earlier version is now Honeybee2012 with data from 1995-2012).
}
\source{
USDA National Agriculture and Statistical Services, \url{https://quickstats.nass.usda.gov/} (Accessed February 2025)
}

\keyword{datasets}
