\name{HoneybeeCircuits}
\alias{HoneybeeCircuits}
\docType{data}
\title{Honeybee Circuits}
\description{
Number of circuits for honeybee dances and nest quality
}
%\usage{data(HoneybeeCircuits)}
\format{
  A dataset with 78 observations on the following 2 variables.
  \tabular{rl}{
    \code{Circuits} \tab {Number of waggle dance circuits for a returning scout bee}\cr
    \code{Quality} \tab {Quality of the nest site: \code{High} or  \code{Low}}\cr
  }
}
\details{
When honeybees are looking for a new home, they send out scouts to explore options. When a scout
returns, she does a "waggle dance" with multiple circuit repetitions to tell the swarm about the option
she found. The bees then decide between the options and pick the best one. Scientists wanted to
find out how honeybees decide which is the best option, so they took a swarm of honeybees to an
island with only two possible options for new homes: one of very high honeybee quality and one of low
quality. They then kept track of the scouts who visited each option and counted the number of waggle
dance circuits each scout bee did when describing the option.
}
\source{
Seeley, T., Honeybee Democracy, Princeton University Press, Princeton, NJ, 2010, p. 128
}

\keyword{datasets}
