\name{HoneybeeWaggle}
\alias{HoneybeeWaggle}
\docType{data}
\title{Honeybee Waggle}
\description{
Honeybee dance duration and distance to nesting site
}
%\usage{data(HoneybeeWaggle)}
\format{
  A dataset with 7 observations on the following 2 variables.
  \tabular{rl}{
    \code{Distance} \tab {Distance to the potential nest site (in meters)}\cr
    \code{Duration} \tab {Duration of the waggle dance (in seconds)}\cr
  }
}
\details{
When honeybee scouts find a food source or a nice site for a new home, they communicate the location
to the rest of the swarm by doing a "waggle dance." They point in the direction of the site and dance
longer for sites farther away. The rest of the bees use the duration of the dance to predict distance
to the site. 
}
\source{
Seeley, T., Honeybee Democracy, Princeton University Press, Princeton, NJ, 2010, p. 128
}

\keyword{datasets}
