\name{HotDogs2015}
\alias{HotDogs2015}
\docType{data}
\title{Hot Dog Eating Contest - 2015}
\description{
Winning number of hot dogs consumed in an eating contest
}
%\usage{data(HotDogs2015)}
\format{
  A dataset with 14 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year} \tab {Year of the contest: 2002-2015}\cr
    \code{HotDogs} \tab {Winning number of hot dogs consumed}\cr
  }
}
\details{
Every Fourth of July, Nathan's Famous in New York City holds a hot dog eating contest, in which
contestants try to eat as many hot dogs (with buns) as possible in ten minutes.
The winning number of hot dogs are given for each year from 2002-2015.  \cr
** From 2e - dataset has been updated for 3e **
}
\source{
Downloaded from \emph{https://en.wikipedia.org/wiki/Nathan's_Hot_Dog_Eating_Contest}
}

\keyword{datasets}
