\name{HotDogs2019}
\alias{HotDogs2019}
\docType{data}
\title{Hot Dog Eating Contest}
\description{
Winning number of hot dogs consumed in an eating contest (2002-2019)
}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Year of the contest: 2002 to 2019}
    \item{\code{HotDogs}}{Winning number of hot dogs consumed}
  }
}
\details{
Every Fourth of July, Nathan's Famous in New York City holds a hot dog eating contest, in which
contestants try to eat as many hot dogs (with buns) as possible in ten minutes.
The winning number of hot dogs are given for each year from 2002-2019.  \cr
** Data set updated for 3e (earlier versions are HotDogs2015 and HotDogs1e) **
}
\source{
Downloaded from \url{https://en.wikipedia.org/wiki/Nathan's_Hot_Dog_Eating_Contest}
}
\keyword{datasets}
