\name{HotDogs2024}
\alias{HotDogs2024}
\docType{data}
\title{Hot Dog Eating Contest - 2024}
\description{
Winning number of hot dogs consumed in an eating contest (2002-2024)
}
\format{
  A data frame with 23 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Year of the contest: 2002 to 2024}
    \item{\code{HotDogs}}{Winning number of hot dogs consumed}
  }
}
\details{
Every Fourth of July, Nathan's Famous in New York City holds a hot dog eating contest, in which
contestants try to eat as many hot dogs (with buns) as possible in ten minutes.
The winning number of hot dogs are given for each year from 2002-2024.  \cr
** Data set updated for 4e (earlier versions are HotDogs2019, HotDogs2015, and HotDogs1e) **
}
\source{
Downloaded from \url{https://en.wikipedia.org/wiki/Nathan's_Hot_Dog_Eating_Contest}
}
\keyword{datasets}
