\name{HouseStarts}
\alias{HouseStarts}
\docType{data}
\title{Housing Starts}
\description{
Quarterly housing starts in the United States from 2000-2015
}
%\usage{data(HouseStarts)}
\format{
  A dataset with 64 observations on the following 3 variables.
  \tabular{rl}{
    \code{Year} \tab {Year (2000 to 2015)}\cr
    \code{Quarter} \tab {\code{Q1}=Jan-Mar, \code{Q2}=Apr-June, \code{Q3}=July-Sept, \code{Q4}=Oct-Dec} \cr
    \code{Houses} \tab {New US residential house construction starts (in thousands)}\cr
  }
}
\details{
Number of new homes started in the US for each quarter from 2000-2015.
}
\source{
Census.gov website \emph{https://www.census.gov/econ/currentdata/} \cr
\emph{https://www.census.gov/econ/currentdata/dbsearch?program=RESCONST&startYear=2000 &endYear=2016&categories=STARTS&dataType=SINGLE&geoLevel=US&notAdjusted=1&submit=GET+DATA&releaseScheduleId=}
}

\keyword{datasets}
