\name{HouseStarts2018}
\alias{HouseStarts2018}
\docType{data}
\title{Housing Starts (2000-2018)}
\description{
Quarterly housing starts in the United States from 2000-2018
}
\format{
  A data frame with 76 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (2000 to 2018)}
    \item{\code{Quarter}}{\code{Q1}=Jan-Mar, \code{Q2}=Apr-June, \code{Q3}=July-Sept, \code{Q4}=Oct-Dec}
    \item{\code{Houses}}{New US residential house construction starts (in thousands)}
  }
}
\details{
Number of new homes started in the US for each quarter from 2000-2018. \cr
Updated for 3e (earlier version is in HouseStarts2015)
}
\source{
Census.gov website \url{https://www.census.gov/econ/currentdata/}

\url{https://www.census.gov/econ/currentdata/dbsearch?program=RESCONST&startYear=2000&endYear=2018&categories=STARTS&dataType=SINGLE&geoLevel=US&notAdjusted=1&submit=GET+DATA&releaseScheduleId=}

}
\keyword{datasets}
