\name{HumanTears50}
\alias{HumanTears50}
\docType{data}
\title{Human Tears - Testosterone}
\description{
Differences in testosterone levels for 50 men in a matched pairs experiment, where the differences are between sniffing female tears and sniffing a placebo
}
\format{
  A data frame with 50 observations on the following 3 variables.
  \describe{
    \item{\code{Placebo}}{Testosterone level after sniffing a placebo}
    \item{\code{Tears}}{Testosterone level after sniffing female tears}
    \item{\code{Difference}}{Difference in testosterone level (Placebo - Tears)}
  }
}
\details{
Fifty men had a pad attached to their upper lip that contained either female tears collected from women who watched a sad film or a salt solution (as a placebo) that had been trickled down the same women's faces.  The data were collected following a double-blind matched pairs design, where the order was randomized and the data were collected on consecutive days.  After sniffing each substance (placebo or tears), men had their salivary testosterone levels measured, in pg/ml.  Data are approximated from information given in the article.
}
\source{
Gelstein, S, et al., "Human Tears Contain a Chemosignal," Science, 331(6014), 226-230, January 14, 2011.
}
\keyword{datasets}
