\name{Hurricanes}
\alias{Hurricanes}
\docType{data}
\title{Hurricanes}
\description{
Hurricanes making landfall on the US east coast each year (1914-2014)
}
%\usage{data(Hurricanes)}
\format{
  A dataset with 64 observations on the following 3 variables.
  \tabular{rl}{
    \code{Year} \tab {Year (1914 to 2014)}\cr
    \code{Hurricanes} \tab {Number of hurricanes making landfall on US East coast} \cr
  }
}
\details{
Number of hurricanes making landfall on the East coast of the United States - yearly 1914-2014

}
\source{
Weather Underground website at \emph{https://www.wunderground.com/hurricane/hurrarchive.asp}
}

\keyword{datasets}
