\name{Hurricanes2018}
\alias{Hurricanes2018}
\docType{data}
\title{Hurricanes (1914 to 2018)}
\description{
Hurricanes in the North Atlantic each year (1914-2018)
}
\format{
  A data frame with 105 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Year (1914 to 2018)}
    \item{\code{Hurricanes}}{Number of North Atlantic hurricanes}
  }
}
\details{
Number of North Atlantic hurricanes - yearly 1914-2018. \cr
** Updated for 3e (earlier version through 2014 is in Hurricanes2014). **
}
\source{
Weather Underground website at \url{https://www.wunderground.com/hurricane/archive}
}
\keyword{datasets}
