\name{Hurricanes2024}
\alias{Hurricanes2024}
\docType{data}
\title{Hurricanes (1914 to 2024)}
\description{
Hurricanes in the North Atlantic each year (1914-2024)
}
\format{
  A data frame with 111 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Year (1914 to 2024)}
    \item{\code{Hurricanes}}{Number of North Atlantic hurricanes}
  }
}
\details{
Number of North Atlantic hurricanes - yearly 1914-2024. \cr
** Updated for 4e (earlier versions are in Hurricanes2018 and Hurricanes2014). **
}
\source{
Weather Underground website at \url{https://www.wunderground.com/hurricane/archive}
}
\keyword{datasets}
