\name{InkjetPrinters}
\alias{InkjetPrinters}
\docType{data}
\title{Inkjet Printers}
\description{
Data from online reviews of inkjet printers (2025)
}
%\usage{data(InkjetPrinters)}
\format{
  A dataset with 25 observations on the following 7 variables.
  \tabular{rl}{
    \code{Model} \tab {Model name of printer}\cr
    \code{PPM} \tab {Printing rate (pages per minute) for a benchmark set of print jobs}\cr
    \code{PhotoTime} \tab {Time (in seconds) to print 4x6 color photos}\cr
    \code{Price} \tab {Typical retail price (in dollars)}\cr
    \code{CostBW} \tab {Cost per page (in cents) for printing in black & white}\cr
    \code{CostColor} \tab {Cost per page (in cents) for printing in color}\cr
    \code{Year} \tab {Year printer was first released}\cr
  }
}
\details{
Information from reviews of all-one inkjet printers at Rtings.com in January 2025.
}
\source{
Inkjet printer data found at \url{https://www.rtings.com/printer/tools/table/157814}, January 2025.
}

\keyword{datasets}
