\name{LifeExpectancyVehicles}
\alias{LifeExpectancyVehicles}
\docType{data}
\title{Life Expectancy and Vehicle Registrations (2017)}
\description{
Yearly US life expectancy and number of registered vehicles (1970-2017)
}
\format{
  A data frame with 48 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (1970 to 2017)}
    \item{\code{LifeExpectancy}}{Average life expectancy (in years) for babies born in the year}
    \item{\code{Vehicles}}{Number of motor vehicles registered in the US (in millions)}
  }
}
\details{
Life expectancy (in years for babies born each year) and number of vehicles registered in the US for each year from 1970 to 2017.  \cr
** Updated for 3e (earlier versions are LifeExpectancyVehicles2e and  LifeExpectancyVehicles1e) **

}
\source{
Vehicle registrations from the Federal Highway Administration,
\url{https://www.fhwa.dot.gov/policyinformation/statistics.cfm}. \cr

Lifetime data from the Centers for Disease Control and Prevention, National Center for Health Statistics \url{https://www.cdc.gov/nchs/hus/contents2019.htm?search=Life_expectancy,}.
}
\keyword{datasets}
