\name{LifeExpectancyVehicles2e}
\alias{LifeExpectancyVehicles2e}
\docType{data}
\title{Life Expectancy and Vehicle Registrations - 2e}
\description{
Yearly US life expectancy and number of registered vehicles (1970-2013)
}
%\usage{data(LifeExpectancyVehicles2e)}
\format{
  A dataset with 44 observations on the following 3 variables.
  \tabular{rl}{
    \code{Year} \tab {Year} \cr
    \code{LifeExpectancy} \tab {Average life expectancy (in years) for babies born in the year} \cr
    \code{Vehicles} \tab {Number of motor vehicles registered in the US (in millions) } \cr
  }
}
\details{
Life expectancy (in years for babies born each year) and number of vehicles registered in the US for each year from 1970 to 2013. \cr
** From 2e - dataset has been updated for 3e **
}
\source{
Vehicle registrations from US Census Bureau,  \emph{http://www.census.gov/compendia/statab/cats/transportation.html}
Lifetime data from the Centers for Disease Control and Prevention, National Center for Health Statistics, Health Data Interactive, \emph{www.cdc.gov/nchs/hdi.htm}
}

\keyword{datasets}
