\name{LightatNight}
\alias{LightatNight}
\docType{data}
\title{Light at Night for Mice}
\description{
Data on body mass gain from an experiment with mice having different nighttime light conditions
}
%\usage{data(LightatNight)}
\format{
  A dataset with 18 observations on the following 2 variables.
  \tabular{rl}{
    \code{Group} \tab {\code{Light}=dim light at night or \code{Dark}=dark at night}\cr
    \code{BMGain} \tab {Body mass gain (in grams over a three week period)}\cr
  }
}
\details{
In this study, 18 mice were randomly split into two groups. One group was on a normal light/dark
cycle (\code{Dark}) and the other group had light during the day and dim light at night (\code{Light}).
The dim light was equivalent to having a television set on in a room. The mice in
darkness ate most of their food during their active (nighttime) period, matching the behavior of mice in the
wild. The mice with dim light at night, however, consumed much of their food during
the well-lit rest period, when most mice are usually sleeping. The change in body mass was recorded after three weeks.\cr
** See also LightatNight4Weeks or LightatNight8Weeks for more variables measured at other points in the same experiment,
with a third experimental condition which had 9 additional mice with a bright light on all the time. **
}
\source{
Fonken, L., et. al., "Light at night increases body mass by shifting time of food intake," Proceedings of the National
Academy of Sciences, October 26, 2010; 107(43): 18664-18669.
}

\keyword{datasets}
