\name{Lock5Data-package}
\alias{Lock5Data-package}
\alias{Lock5Data}
\docType{package}
\title{
Lock5 Datasets
}
\description{
Datasets for Statistics: Unlocking the Power of Data \cr
by Lock, Lock, Lock, Lock and Lock
}
\details{
\tabular{ll}{
Package: \tab Lock5Data\cr
Type: \tab Package\cr
Version: \tab 2.6\cr
Date: \tab 2012-08-20\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
%%~~ An overview of how to use the package, including the most important ~~
%%~~ functions ~~
}
\author{
Robin Lock

Maintainer: Robin Lock <rlock@stlawu.edu>
}


\keyword{ package }
