\name{MalevolentUniformsNFL}
\alias{MalevolentUniformsNFL}
\docType{data}
\title{
Malevolent Uniforms NFL
}
\description{
Perceived malevolence of uniforms and penalies for National Football League (NFL) teams
}
\usage{data(MalevolentUniformsNFL)}
\format{
  A data frame with 28 observations on the following 7 variables.
  \describe{
    \item{\code{NFLTeam}}{Team name}
    \item{\code{NFL_Malevolence}}{Score reflecting the "malevolence" of a team's uniform}
    \item{\code{ZPenYds}}{Z-score for penalty yards}
  }
}
\details{
Participants with no knowledge of the teams rated the jerseys 
on characteristics such as timid/aggressive, nice/mean and good/bad. The averages of these responses
produced a "malevolence" index with higher scores signifying impressions of more malevolent uniforms. To
measure aggressiveness, the authors used the amount of penalty yards converted to z-scores and averaged for each team over the seasons from 1970-1986.
}
\source{
Frank and Gilovich, "The Dark Side of Self- and Social Perception: Black Uniforms and Aggression in Professional Sports",
Journal of Personality and Social Psychology, Vol. 54, No. 1, 1988, p. 74-85.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(MalevolentUniformsNFL)
}
\keyword{datasets}
