\name{MalevolentUniformsNHL}
\alias{MalevolentUniformsNHL}
\docType{data}
\title{
Malevolent Uniforms NHL
}
\description{
Perceived malevolence of uniforms and penalies for National Hockey League (NHL) teams
}
\usage{data(MalevolentUniformsNHL)}
\format{
  A data frame with 28 observations on the following 3 variables.
  \describe{
    \item{\code{NHLTeam}}{Team name}
    \item{\code{NHL_Malevolence}}{Score reflecting the "malevolence" of a team's uniform}
    \item{\code{ZPenMin}}{Z-score for penalty minutes}
  }
}
\details{
Participants with no knowledge of the teams rated the jerseys 
on characteristics such as timid/aggressive, nice/mean and good/bad. The averages of these responses
produced a "malevolence" index with higher scores signifying impressions of more malevolent uniforms. To
measure aggressiveness, the authors used the amount of penalty minutes converted to z-scores and averaged for each team over the seasons from 1970-1986.
}
\source{
Frank and Gilovich, "The Dark Side of Self- and Social Perception: Black Uniforms and Aggression in Professional Sports",
Journal of Personality and Social Psychology, Vol. 54, No. 1, 1988, p. 74-85.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(MalevolentUniformsNHL)
}
\keyword{datasets}
