\name{MammalLongevity}
\alias{MammalLongevity}
\docType{data}
\title{
Mammal Longevity
}
\description{
Longevity and gestation period for mammals
}
\usage{data(MammalLongevity)}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
    \item{\code{Animal}}{Species of mammal}
    \item{\code{Gestation}}{Time from fertilization until birth (in days)}
    \item{\code{Longevity}}{Average lifespan (in years)}
  }
}
\details{
Dataset with average lifespan (in years) and typical gestation period (in days) for 40 different species of mammals. 
}
\source{
2010 World Almanac, pg. 292.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(MammalLongevity)
}
\keyword{datasets}
