\name{ManhattanApartments}
\alias{ManhattanApartments}
\docType{data}
\title{Manhattan Apartment Prices (2025)}
\description{
Apartment prices for sale in Manhattan in 2025
}
\format{
  A data frame with 20 observations on the following variable.
  \describe{
    \item{\code{Rent}}{Monthly rent (in dollars)}
  }
}
\details{
Monthly rents for a sample of 20 one-bedroom apartments in Manhattan, NY in March 2025
}
\source{
Apartments newly available on Zillow at \emph{https://www.zillow.com/manhattan-new-york-ny/rentals/}, March, 2025.
}
\keyword{datasets}
