\name{ManhattanApartments2011}
\alias{ManhattanApartments2011}
\docType{data}
\title{Manhattan Apartment Prices - 2011}
\description{
Monthly rent for one-bedroom apartments in Manhattan, NY
}
%\usage{data(ManhattanApartments2011)}
\format{
  A dataset with 20 observations on the following variable.
  \tabular{rl}{
    \code{Rent} \tab {Montly rent in dollars}\cr
  }
}
\details{
Monthly rents for a sample of 20 one-bedroom apartments in Manhattan, NY that were advertised on Craig's List in July, 2011.\cr
** From 2e - dataset has been updated for 3e **
}
\source{
Apartments advertised on Craig's List at \emph{newyork.craigslist.org}, July 5, 2011.
}

\keyword{datasets}
