\name{ManhattanApartments2019}
\alias{ManhattanApartments2019}
\docType{data}
\title{Manhattan Apartment Prices (2019)}
\description{
Apartment prices for sale in Manhattan in 2019
}
\format{
  A data frame with 20 observations on the following variable.
  \describe{
    \item{\code{Rent}}{Monthly rent (in dollars)}
  }
}
\details{
Monthly rents for a sample of 20 one-bedroom apartments in Manhattan, NY that were advertised on Craig's List in November, 2019.
}
\source{
Apartments newly advertised on Craig's List at \url{https://newyork.craigslist.org/}, November, 2019.
}
\keyword{datasets}
