\name{MarriageAges}
\alias{MarriageAges}
\docType{data}
\title{
Marriage Ages
}
\description{
Ages for husbands and wives from marriage licenses
}
\usage{data(MarriageAges)}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{Husband}}{Age of husband at marriage}
    \item{\code{Wife}}{Age of wife at marriage}
  }
}
\details{
Data from a sample of 100 marriage licences in St. Lawrence County, NY gives the ages of husbands and wives for newly married couples.
}
\source{
Thanks to Linda Casserly, St. Lawrence County Clerk's Office
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%%data(MarriageAges)
}
\keyword{datasets}
