\name{MastersGolf}
\alias{MastersGolf}
\docType{data}
\title{Masters Golf Scores}
\description{
Scores from the 2011 Masters golf tournament
}
%\usage{data(MastersGolf)}
\format{
  A dataset with 20 observations on the following 2 variables.
  \tabular{rl}{
    \code{First} \tab {First round score (in relation to par)}\cr
    \code{Final} \tab {Final four round score (in relation to par)}\cr
  }
}
\details{
Data for a random sample of 20 golfers who made the cut at the 2011 Masters golf tournament.
}
\source{
2011 Masters tournament results at http://www.masters.com/en_US/discover/past_winners.html
}

\keyword{datasets}
