\name{MentalMuscle}
\alias{MentalMuscle}
\docType{data}
\title{Mental Muscle}
\description{
Comparing actual movements to mental imaging movements
}
%\usage{data(MentalMuscle)}
\format{
  A dataset with 32 observations on the following 3 variables.
  \tabular{rl}{
    \code{Action} \tab {Treatment: \code{} \code{Actual} motions or \code{Mental} imaging motions}\cr
    \code{PreFatigue} \tab {Time (in seconds) to complete motions before fatigue}\cr
    \code{PostFatigue} \tab {Time (in seconds) to complete motions after fatigue}\cr
  }
}
\details{
In this study, participants were asked to either perform actual arm pointing motions or to mentally
imagine equivalent arm pointing motions. Participants then developed muscle fatigue by holding a
heavy weight out horizontally as long as they could. After becoming fatigued, they were asked to
repeat the previous mental or actual motions. Eight participants were assigned to each group, and
the time in seconds to complete the motions was measured before and after fatigue. 
}
\source{
Data approximated from summary statistics in: Demougeot L. and Papaxanthis C., "Muscle Fatigue Affects Mental Simulation
of Action," The Journal of Neuroscience, July 20, 2011, 31(29):10712-10720.
}

\keyword{datasets}
