\name{MiamiHeat}
\alias{MiamiHeat}
\docType{data}
\title{
Miami Heat basketball
}
\description{
Game log data for the Miami Heat basketball team in 2010-11
}
\usage{data(MiamiHeat)}
\format{
  A data frame with 82 observations on the following 33 variables.
  \describe{
    \item{\code{Game}}{ID number for each game}
    \item{\code{Date}}{Data the game was played}
    \item{\code{Location}}{\code{Away}  or \code{Home}}
    \item{\code{Opp}}{Opponent tream}
    \item{\code{Win}}{Game result: \code{L} or \code{W}}
    \item{\code{FG}}{Field goals made}
    \item{\code{FGA}}{Field goals attempted}
    \item{\code{FG3}}{Three-point field goals made}
    \item{\code{FG3A}}{Three-point field goals attempted}
    \item{\code{FT}}{Free throws made}
    \item{\code{FTA}}{Free throws attempted}
    \item{\code{Rebounds}}{Total rebounds}
    \item{\code{OffReb}}{Offensive rebounds}
    \item{\code{Assists}}{Number of assists}
    \item{\code{Steals}}{Number of steals}
    \item{\code{Blocks}}{Number of shots blocked}
    \item{\code{Trunovers}}{Number of turnovers}
    \item{\code{Fouls}}{Number of fouls}
    \item{\code{Points}}{Number of points scored}
    \item{\code{OppFG}}{Opponet's field goals made}
    \item{\code{OppFGA}}{Opponent's Field goals attempted}
    \item{\code{OppFG3}}{Opponent's Three-point field goals made}
    \item{\code{OppFG3A}}{Opponent's Three-point field goals attempted}
    \item{\code{OppFT}}{Opponent's Free throws made}
    \item{\code{OppFTA}}{Opponent's Free throws attempted}
    \item{\code{OppOffReb}}{Opponent's Total rebounds}
    \item{\code{OppRebounds}}{Opponent's Offensive rebounds}
    \item{\code{OppAssists}}{Opponent's assists}
    \item{\code{OppSteals}}{Opponent's steals}
    \item{\code{OppBlocks}}{Opponent's shots blocked}
    \item{\code{OppTurnovers}}{Opponent's turnovers}
    \item{\code{OppFouls}}{Opponent's fouls}
    \item{\code{OppPoints}}{Opponent's points scored}
  }
}
\details{
Information from online boxscores for all 82 regular season games payed by the Miami Heat basketball team during the 2010-11 regular season. 
}
\source{
Data for the 2010-11 Miami games downloaded from http://www.basketball-reference.com/teams/MIA/2011/gamelog/
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(MiamiHeat)
}
\keyword{datasets}
