\name{MinistersRum}
\alias{MinistersRum}
\docType{data}
\title{
Ministers and Rum
}
\description{
Number of ministers and rum imports to New England between 1860 and 1940
}
\usage{data(MinistersRum)}
\format{
  A data frame with 17 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Every five years from 1860-1940}
    \item{\code{Ministers}}{Number of Methodist ministers in New England}
    \item{\code{Rum}}{Number of barrels of rum imported into Boston}
  }
}
\details{
Data every five years from 1860 to 1940 on he number of Methodist ministers working in New England and the annual rum imports into Boston. 
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%data(MinistersRum)
}
\keyword{datasets}
