\name{MustangPrice1e}
\alias{MustangPrice1e}
\docType{data}
\title{
Mustang Prices (2008)
}
\description{
Price, age, and mileage for used Mustang cars at an internet website (2008)
}
%\usage{data(MustangPrice)}
\format{
  A dataset with 25 observations on the following 3 variables.
  \tabular{rl}{
    \code{Age} \tab {Age of the car (in years)} \cr
    \code{Miles} \tab {Mileage on the car (in 1,000's)} \cr
    \code{Price} \tab {Asking price (in $1,000's)} \cr
  }
}
\details{
A statistics student, Gabe McBride, was interested in prices for used Mustang cars being offered for sale on an internet site.
He sampled 25 cars from the website and recorded the age (in years), mileage (in thousands of miles) and asking price (in $1,000's) for each car in his sample.
}
\source{
Student project with data collected from \emph{autotrader.com} in 2008.
}

\keyword{datasets}
