\name{NBAPlayers2011}
\alias{NBAPlayers2011}
\docType{data}
\title{
NBA Players data for 2010-11 Season
}
\description{
Data from the 2010-2011 regular season for 176 NBA basketball players.  
}
%\usage{data(NBAPlayers2011)}
\format{
  A dataset with 176 observations on the following 25 variables.
  \tabular{rl}{
    \code{Player} \tab {Name of player}\cr
    \code{Age} \tab {Age in years)}\cr
    \code{Team} \tab {Team name}\cr
    \code{Games} \tab {Games played (out of 82)}\cr
    \code{Starts} \tab {Games started}\cr
    \code{Mins} \tab {Minutes played}\cr
    \code{MinPerGame} \tab {Minutes per game}\cr
    \code{FGMade} \tab {Field goals made}\cr
    \code{FGAttempt} \tab {Field goals attempted}\cr
    \code{FGPct} \tab {Field goal percentage}\cr
    \code{FG3Made} \tab {Three-point field goals made}\cr
    \code{FG3Attempt} \tab {Three-point field goals attempted}\cr
    \code{FG3Pct} \tab {Three-point field goal percentage}\cr
    \code{FTMade} \tab {Free throws made}\cr
    \code{FTAttempt} \tab {Free throws attempted}\cr
    \code{FTPct} \tab {Free throw percentage}\cr
    \code{OffRebound} \tab {Offensive rebounds}\cr
    \code{DefRebound} \tab {Defensive rebounds}\cr
    \code{Rebounds} \tab {Total rebounds}\cr
    \code{Assists} \tab {Number of assists}\cr
    \code{Steals} \tab {Number of steals}\cr
    \code{Blocks} \tab {Number of blocked shots}\cr
    \code{Turnovers} \tab {Number of turnovers}\cr
    \code{Fouls} \tab {Number of personal fouls}\cr
    \code{Points} \tab {Number of points scored}\cr
  }
}
\details{
The dataset includes information for all National Basketball Association (NBA) players 
who averaged more than 24 minutes per game during the 2010-2011 regualr season.
}
\source{
http://www.basketball-reference.com/leagues/NBA_2011_stats.html
}

\keyword{datasets}
