\name{NBAPlayers2015}
\alias{NBAPlayers2015}
\docType{data}
\title{
NBA Players Data for 2014-15 Season
}
\description{
Data from the 2014-2015 regular season for 182 NBA basketball players.
}
%\usage{data(NBAPlayers2011)}
\format{
  A dataset with 182 observations on the following 25 variables.
  \tabular{rl}{
    \code{Player} \tab {Name of player} \cr
    \code{Position} \tab {\code{PG}=point guard, \code{SG}=shooting guard, \code{PF}=power forward, \code{SF}=small forward, \code{C}=center} \cr
    \code{Age} \tab {Age (in years)} \cr
    \code{Team} \tab {Team name} \cr
    \code{Games} \tab {Games played (out of 82)} \cr
    \code{Starts} \tab {Games started} \cr
    \code{Mins} \tab {Minutes played} \cr
    \code{MinPerGame} \tab {Minutes per game} \cr
    \code{FGMade} \tab {Field goals made} \cr
    \code{FGAttempt} \tab {Field goals attempted} \cr
    \code{FGPct} \tab {Field goal percentage} \cr
    \code{FG3Made} \tab {Three-point field goals made} \cr
    \code{FG3Attempt} \tab {Three-point field goals attempted} \cr
    \code{FG3Pct} \tab {Three-point field goal percentage} \cr
    \code{FTMade} \tab {Free throws made} \cr
    \code{FTAttempt} \tab {Free throws attempted} \cr
    \code{FTPct} \tab {Free throw percentage} \cr
    \code{OffRebound} \tab {Offensive rebounds} \cr
    \code{DefRebound} \tab {Defensive rebounds} \cr
    \code{Rebounds} \tab {Total rebounds} \cr
    \code{Assists} \tab {Number of assists} \cr
    \code{Steals} \tab {Number of steals} \cr
    \code{Blocks} \tab {Number of blocked shots} \cr
    \code{Turnovers} \tab {Number of turnovers} \cr
    \code{Fouls} \tab {Number of personal fouls} \cr
    \code{Points} \tab {Number of points scored} \cr
  }
}
\details{
Data for 182 NBA basketball players from the 2014-2015 regular season.  Includes all players who averaged more than 24 minutes per game that season. \cr
** Data set updated for 2e (original is NBAPlayers2011) **
}
\source{
\emph{http://www.basketball-reference.com/leagues/NBA_2015_stats.html}
}

\keyword{datasets}
