\name{NBAStandings}
\alias{NBAStandings}
\docType{data}
\title{
NBA 2010-11 Regular Season Standings
}
\description{
Won-Loss record and statistics for NBA Teams
}
\usage{data(HomesForSaleCA)}
\format{
  A data frame with 30 observations on the following 5 variables.
  \describe{
    \item{\code{Team}}{Team name}
    \item{\code{Wins}}{Number of wins in an 82 game regular season}
    \item{\code{Losses}}{Number of losses}
    \item{\code{WinPct}}{Proportion of games won}
    \item{\code{PtsFor}}{Average points scored per game}
    \item{\code{PtsAgainst}}{Average points allowed per game}
  }
}
\details{
Won-Loss record and regular season statistics for 30 teams in the National Basketball Association
for the 2010-2011 season
}
\source{
Data downloaded from http://www.basketball-reference.com/leagues/NBA_2011_games.html
}
\references{
%%
}
\examples{
%% data(NBAStandings)
}
\keyword{datasets}
