\name{NBAStandings2011}
\alias{NBAStandings2011}
\docType{data}
\title{
NBA 2010-11 Regular Season Standings
}
\description{
Won-Loss record and statistics for NBA Teams in 2010-2011
}
%\usage{data(NBAStandings2011)}
\format{
  A dataset with 30 observations on the following 6 variables.
  \tabular{rl}{
    \code{Team} \tab {Team name} \cr
    \code{Wins} \tab {Number of wins in an 82 game regular season} \cr
    \code{Losses} \tab {Number of losses} \cr
    \code{WinPct} \tab {Proportion of games won} \cr
    \code{PtsFor} \tab {Average points scored per game} \cr
    \code{PtsAgainst} \tab {Average points allowed per game} \cr
  }
}
\details{
Won-Loss record and regular season statistics for 30 teams in the National Basketball Association
for the 2010-2011 season. \cr
** From 1e - dataset has been updated for 2e and 3e**
}
\source{
Data downloaded from \emph{http://www.basketball-reference.com/leagues/NBA_2011_games.html}
}

\keyword{datasets}
