\name{NBAStandings2019}
\alias{NBAStandings2019}
\docType{data}
\title{NBA 2018-2019 Regular Season Standings}
\description{
Won-Loss record and statistics for NBA Teams in 2018-2019
}
\format{
  A data frame with 30 observations on the following 6 variables.
  \describe{
    \item{\code{Team}}{Team name}
    \item{\code{Wins}}{Number of wins in an 82 game regular season}
    \item{\code{Losses}}{Number of losses}
    \item{\code{WinPct}}{Proportion of games won}
    \item{\code{PtsFor}}{Average points scored per game}
    \item{\code{PtsAgainst}}{Average points allowed per game}
  }
}
\details{
Won-Loss record and regular season statistics for 30 teams in the National Basketball Association
for the 2018-2019 season. \cr
** Data set updated for 3e (earlier version are NBAStandings2016 and NBAStandings1e) **
}
\source{
Data downloaded from \emph{http://www.basketball-reference.com/leagues/NBA_2019_games.html}
}
\keyword{datasets}
