\name{NFLPreSeason2014}
\alias{NFLPreSeason2014}
\docType{data}
\title{
Wins for NFL Teams (2005-2014)
}
\description{
Number of preseason and regular season wins for NFL teams, each year from 2005 to 2014.
}
%\usage{data(NFLPreSeason2014)}
\format{
  A dataset with 320 observations on the following 4 variables.
  \tabular{rl}{
    \code{Team} \tab {Code for one of 32 NFL teams}\cr
    \code{Season} \tab {Year between 2005 and 2014}\cr
    \code{Preseason} \tab {Number of preseason wins (out of 4 games)}\cr
    \code{RegularWins} \tab {Number of regular season wins (out of 16 games)}\cr
  }
}
\details{
Number of wins in the preseason (out of 4 preseason games) and regular season (out of 16 regular season games) for each of the 32 National Football (NFL) teams over a ten year period from 2005 to 2014.\cr
** From 2e - dataset has been updated for 3e **
}
\source{
Data available at \emph{http://www.pro-football-reference.com/}.
}

\keyword{datasets}
