\name{NFLPreseason2019}
\alias{NFLPreseason2019}
\docType{data}
\title{Wins for NFL Teams (2005-2019)}
\description{
Number of preseason and regular season wins for NFL teams, each year from 2005 to 2019.
}
\format{
  A data frame with 480 observations on the following 4 variables.
  \describe{
    \item{\code{Team}}{Code for one of 32 NFL teams}
    \item{\code{Season}}{Year between 2005 and 2019}
    \item{\code{Preseason}}{Number of preseason wins (out of 4 games)}
    \item{\code{RegularWins}}{Number of regular season wins (out of 16 games)}
  }
}
\details{
Number of wins in the preseason (out of 4 preseason games) and regular season (out of 16 regular season games) for each of the 32 National Football (NFL) teams over a fifteen year period from 2005 to 2019. \cr
** Updated for 3e (earlier version is now NFLPreseason2014). **
}
\source{
Data available at \url{https://www.pro-football-reference.com/}.
}
\keyword{datasets}
