\name{NFLScores2011}
\alias{NFLScores2011}
\docType{data}
\title{NFL Game Scores in 2011}
\description{
Results for all NFL games for the 2011 regular season
}
%\usage{data(NFLScores2011)}
\format{
  A dataset with 256 observations on the following 11 variables.
  \tabular{rl}{
    \code{Week} \tab {a numeric vector}\cr
    \code{HomeTeam} \tab {Home team name}\cr
    \code{AwayTeam} \tab {Visiting team name}\cr
    \code{HomeScore} \tab {Points scored by the home team}\cr
    \code{AwayScore} \tab {Points scored by the visiting team}\cr
    \code{HomeYards} \tab {Yards gained by the home team}\cr
    \code{AwayYards} \tab {Yards gained by the visting team}\cr
    \code{HomeTO} \tab {Turnovers lost by the home team}\cr
    \code{AwayTO} \tab {Turnovers lost by the visting team}\cr
    \code{Date} \tab {Date of the game}\cr
    \code{Day} \tab {Day of the week: \code{Mon},  \code{Sat},  \code{Sun}, or \code{Thu}}\cr
  }
}
\details{
Data for all 256 regular season games in the National Football League (NFL) for the 2011 season. 
}
\source{
NFL scores and game statistics found at http://www.pro-football-reference.com/years/2011/games.htm.
}

\keyword{datasets}
