\name{NFLScores2018}
\alias{NFLScores2018}
\docType{data}
\title{NFL Scores in 2018}
\description{
Results for all NFL games for the 2018 regular season
}
\format{
  A data frame with 256 observations on the following 11 variables.
  \describe{
    \item{\code{Week}}{Week of the season (1 through 17)}
    \item{\code{HomeTeam}}{Home team name}
    \item{\code{AwayTeam}}{Visiting team name}
    \item{\code{HomeScore}}{Points scored by the home team}
    \item{\code{AwayScore}}{Points scored by the visiting team}
    \item{\code{HomeYards}}{Yards gained by the home team}
    \item{\code{AwayYards}}{Yards gained by the visiting team}
    \item{\code{HomeTO}}{Turnovers lost by the home team}
    \item{\code{AwayTO}}{Turnovers lost by the visiting team}
    \item{\code{Date}}{Date of the game}
    \item{\code{Day}}{Day of the week (\code{Mon}, \code{Sat}, \code{Sun}, or \code{Thu})}
  }
}
\details{
Data for all 256 regular season games in the National Football League (NFL) for the 2018 season. \cr
** Updated for 3e (earlier version is NFLScores2011). **
}
\source{
NFL scores and game statistics found at
\url{https://www.pro-football-reference.com/years/2018/games.htm}.
}
\keyword{datasets}
