\name{Ninja}
\alias{Ninja}
\docType{data}
\title{Ninja Warrior Times}
\description{
Times for students doing a Ninja Warrior obstacle course
}
\format{
  A data frame with 99 observations on the following 9 variables.
  \describe{
    \item{\code{Name}}{First name of the child}
    \item{\code{Age}}{Age of the child (in years)}
    \item{\code{Sex}}{Sex of the child (\code{F} or \code{M})}
    \item{\code{Time1}}{Time to complete the course, in seconds, the first time the child was timed on the course}
    \item{\code{Time2}}{Time, in seconds, for a second run through the course}
    \item{\code{Time3}}{Time, in seconds, for a third run through the course}
    \item{\code{Time4}}{Time, in seconds, for a fourth run through the course}
    \item{\code{Time5}}{Time, in seconds, for a fifth run through the course}
    \item{\code{AvgTime}}{Average time for all runs the child completed}
   }
}

\details{
"Ninja Warrior" is a new sport based off the TV show "American Ninja Warrior" in which people go through an obstacle course as fast as they can.  The sport is growing in popularity and set to debut at the 2028 Summer Olympic Games under the name "Obstacle Course Racing" as part of the Modern Pentathlon.  \\

This dataset contains times from a 2025 Ninja Warrior class for children. Due to time constraints some children only did the course once, while others were timed for up to five different runs.
}
\source{
Data collected in 2025 from a class at Centre Ninja, \url{https://thecentrepa.com/about-ninjacor/}.  Thanks to Coach Brett Corbishley for collecting the data, and to Axel and Cal Lock Morgan for entering the data.
}
\keyword{datasets}
