\name{OlympicMarathon}
\alias{OlympicMarathon}
\docType{data}
\title{
2012 Olympic Men's Marathon
}
\description{
Times for all finishers in the men's marathon at the 2012 Olympics
}
%\usage{data(OlympicMarathon)}
\format{
  A data frame with 85 observations on the following 4 variables.
  \tabular{rl}{
    \code{Athlete} \tab {Name of marathoner} \cr
    \code{Country} \tab {Nationality of marathoner (3 letter country code)} \cr
    \code{Time} \tab {Time as H:MM:SS} \cr
    \code{Minutes} \tab {Time in minutes}
  }
}
\details{
Results for all finishers in the 2012 Men's Olympic marathon in London, England. \cr
** This is an updated version (previous is now in OlympicMarathon1e) **
}
\source{
\emph{http://www.olympic.org/olympic-results/london-2012/athletics/marathon-m}, accessed October 2015.
}

\keyword{datasets}
