\name{OlympicMarathon1e}
\alias{OlympicMarathon1e}
\docType{data}
\title{
2008 Olympic Men's Marathon
}
\description{
Times for all finishers in the men's marathon at the 2008 Olympics
}
%\usage{data(OlympicMarathon)}
\format{
  A data frame with 76 observations on the following 5 variables.
  \tabular{rl}{
    \code{Rank} \tab {Order of finish} \cr
    \code{Athlete} \tab {Name of marathoner} \cr
    \code{Nationality} \tab {Country of marathoner} \cr
    \code{Time} \tab {Time as H:MM:SS} \cr
    \code{Minutes} \tab {Time in minutes}
  }
}
\details{
Results for all finishers in the 2008 Men's Olympic marathon in Beijing, China. \cr
** This 1e version has been updated for 2e **
}
\source{
\emph{http://2008olympics.runnersworld.com/2008/08/mens-marathon-results.html}
}

\keyword{datasets}
