\name{OlympicMarathon2016}
\alias{OlympicMarathon2016}
\docType{data}
\title{2016 Olympic Men's Marathon}
\description{
Times for all finishers in the men's marathon at the 2016 Olympics
}
\format{
  A data frame with 140 observations on the following 4 variables.
  \describe{
    \item{\code{Athlete}}{Name of marathoner}
    \item{\code{Country}}{Nationality of marathoner (3 letter country code)}
    \item{\code{Time}}{Time as H:MM:SS}
    \item{\code{Minutes}}{Time in minutes}
  }
}
\details{
Results for all finishers in the 2016 Men's Olympic marathon in Rio de Janeiro, Brazil. \cr
** Updated for 3e (earlier versions are now in OlympicMarathon2012 and OlympicMarathon2008) **
}
\source{
Downloaded from \emph{https://olympics.com/en/olympic-games/rio-2016/results/athletics/marathon-men}
}
\keyword{datasets}
