\name{OlympicMarathon2024}
\alias{OlympicMarathon2024}
\docType{data}
\title{2024 Olympic Men's Marathon}
\description{
Times for all finishers in the men's marathon at the 2024 Olympics
}
\format{
  A data frame with 71 observations on the following 4 variables.
  \describe{
    \item{\code{Athlete}}{Name of marathoner}
    \item{\code{Country}}{Nationality of marathoner (3 letter country code)}
    \item{\code{Time}}{Time as H:MM:SS}
    \item{\code{Minutes}}{Time in minutes}
  }
}
\details{
Results for all finishers in the 2024 Men's Olympic marathon in Paris, France. \cr
** Updated for 4e (earlier versions are now in OlympicMarathon 2016, OlympicMarathon2012, and OlympicMarathon2008) **
}
\source{
Downloaded from \emph{https://www.olympics.com/en/olympic-games/paris-2024/results/athletics/men-marathon}
}
\keyword{datasets}
