\name{OttawaSenators}
\alias{OttawaSenators}
\docType{data}
\title{
Ottaw Senators hockey team
}
\description{
Data for 24 players on the 2009-10 Ottawa Senators
}
\usage{data(OttawaSenators)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Points}}{Number of points (goals + assists) scored}
    \item{\code{PenMins}}{Number of penalty minutes}
  }
}
\details{
Points scored and penaly minutes for 24 players (excluding goalies) playing ice hockey for the Ottawa Senators during the 2009-10 NHL regular season. 
}
\source{
Data obtained from http://senators.nhl.com/club/stats.htm.
}
\references{
%%
}
\examples{
%% data(OttawaSenators)
}
\keyword{datasets}
