\name{OttawaSenators2010}
\alias{OttawaSenators2010}
\docType{data}
\title{
Ottawa Senators Hockey Team - 2010
}
\description{
Data for 24 players on the 2009-10 Ottawa Senators
}
%\usage{data(OttawaSenators2010)}
\format{
  A dataset with 24 observations on the following 2 variables.
  \tabular{rl}{
    \code{Points} \tab {Number of points (goals + assists) scored}\cr
    \code{PenMins} \tab {Number of penalty minutes}\cr
  }
}
\details{
Points scored and penalty minutes for 24 players (excluding goalies) playing ice hockey for the Ottawa Senators during the 2009-10 NHL regular season. \cr
** From 1e - dataset has been updated for 2e and 3e **
}
\source{
Data obtained from \emph{http://senators.nhl.com/club/stats.htm}.
}

\keyword{datasets}
