\name{OttawaSenators}
\alias{OttawaSenators}
\docType{data}
\title{
Ottawa Senators Hockey Team (2014-2015)
}
\description{
Data for 24 players on the 2014-2105 Ottawa Senators NHL team
}
%\usage{data(OttawaSenators)}
\format{
  A dataset with 24 observations on the following 10 variables.
  \tabular{rl}{
    \code{Player} \tab {Players name} \cr
    \code{Position} \tab {\code{D}=defense, \code{C}=center, \code{RW}=right wing, \code{LW}=left wing} \cr
    \code{Age} \tab {Age (in years)} \cr
    \code{Games} \tab {Games played in the 2014-15 NHL season (out of 82)} \cr
    \code{Goals} \tab {Goals} \cr
    \code{Assists} \tab {Assists} \cr
    \code{Points} \tab {Goals + Assists} \cr
    \code{PlusMinus} \tab {Difference between (even strength) goals for and against while on ice} \cr
    \code{PenMins} \tab {Number of penalty minutes} \cr
    \code{MinPerGame } \tab {Average minutes on the ice per game} \cr
  }
}
\details{
 Data for all players (except goalies) who played at least 10 games with the Ottawa Senators hockey team in the 2014-15 NHL season. \cr
 ** This is an updated version (previous version is now in OttawaSenators1e) **
}
\source{
\emph{http://www.hockey-reference.com/teams/OTT/2015.html}, accessed October 2015.
}

\keyword{datasets}
