\name{OttawaSenators2019}
\alias{OttawaSenators2019}
\docType{data}
\title{Ottawa Senators Hockey Team (2018-2019)}
\description{
Data for 26 players on the 2018-2109 Ottawa Senators NHL team
}
\format{
  A data frame with 26 observations on the following 10 variables.
  \describe{
    \item{\code{Player}}{Players name}
    \item{\code{Position}}{\code{D}=defense, \code{C}=center, \code{RW}=right wing, \code{LW}=left wing}
    \item{\code{Age}}{Age (in years)}
    \item{\code{Games}}{Games played in the 2018-19 NHL season (out of 82)}
    \item{\code{Goals}}{Goals}
    \item{\code{Assists}}{Assists}
    \item{\code{Points}}{Goals + Assists}
    \item{\code{PlusMinus}}{Difference between (even strength) goals for and against while on ice}
    \item{\code{PenMins}}{Number of penalty minutes}
    \item{\code{MinPerGame}}{Average minutes on the ice per game}
  }
}
\details{
 Data for all players (except goalies) who played at least 10 games with the Ottawa Senators hockey team in the 2018-2019 NHL season. \cr
  ** Updated for 3e (previous versions are now OttawaSenators2015 and OttawaSenators1e) **
}
\source{
\url{https://www.hockey-reference.com/teams/OTT/2019.html}
}
\keyword{datasets}
