\name{OttawaSenators2024}
\alias{OttawaSenators2024}
\docType{data}
\title{Ottawa Senators Hockey Team (2023-2024)}
\description{
Data for 23 players on the 2023-2024 Ottawa Senators NHL team
}
\format{
  A data frame with 23 observations on the following 10 variables.
  \describe{
    \item{\code{Player}}{Players name}
    \item{\code{Position}}{\code{D}=defense, \code{C}=center, \code{RW}=right wing, \code{LW}=left wing}
    \item{\code{Age}}{Age (in years)}
    \item{\code{Games}}{Games played in the 2023-24 NHL season (out of 82)}
    \item{\code{Goals}}{Goals}
    \item{\code{Assists}}{Assists}
    \item{\code{Points}}{Goals + Assists}
    \item{\code{PlusMinus}}{Difference between (even strength) goals for and against while on ice}
    \item{\code{PenMins}}{Number of penalty minutes}
    \item{\code{MinPerGame}}{Average minutes on the ice per game}
  }
}
\details{
 Data for all players (except goalies) who played at least 20 games with the Ottawa Senators hockey team in the 2023-2024 NHL season. \cr
  ** Updated for 4e (previous versions are now OttawaSenators2019, OttawaSenators2015, and OttawaSenators1e) **
}
\source{
Data downloaded from \url{https://www.hockey-reference.com/teams/OTT/2024.html} (March 2025)
}
\keyword{datasets}
