\name{PASeniors2019}
\alias{PASeniors2019}
\docType{data}
\title{Pennsylvania High School Seniors (2010-2019)}
\description{
Information on a sample of high school seniors from the state of Pennsylvania between 2010 and 2019.
}
\format{
  A data frame with 457 observations on the following 36 variables.
  \describe{
    \item{\code{Year}}{Year student submitted data}
    \item{\code{Gender}}{\code{Female} or \code{Male}}
    \item{\code{Age}}{Age (in years)}
    \item{\code{Hand}}{Dominant hand (\code{Left}, \code{Right}, or \code{Both})}
    \item{\code{Height}}{Height (in cm)}
    \item{\code{Foot}}{Foot length (in cm)}
    \item{\code{Armspan}}{Armspan (in cm)}
    \item{\code{Languages}}{Languages spoken}
    \item{\code{GetToSchool}}{Main mode of transportation to school (\code{Bus}, \code{Car}, or \code{Walk} - Walk includes bicycle)}
    \item{\code{TravelTime}}{Travel time to school (in minutes)}
    \item{\code{ReactionTime}}{Time (in seconds) to click when a color changes}
    \item{\code{MemoryScore}}{Score in an online memory game}
    \item{\code{Activity}}{Favorite physical activity}
    \item{\code{Music}}{Favorite genre of music}
    \item{\code{BirthMonth}}{Birth month}
    \item{\code{Season}}{Favorite season}
    \item{\code{Allergies}}{Have allergies? (\code{No} or \code{Yes})}
    \item{\code{Vegetarian}}{Vegetarian?  (\code{No} or \code{Yes})}
    \item{\code{FavFood}}{Favorite food}
    \item{\code{Drink}}{Beverage used most often during the day}
    \item{\code{FavSubject}}{Favorite subject in school}
    \item{\code{Sleep1}}{Typical hours of sleep on a school night}
    \item{\code{Sleep2}}{Typical hours of sleep on a non-school night}
    \item{\code{Occupants}}{Number of occupants at home}
    \item{\code{Communicate}}{Most often method to communicate with friends}
    \item{\code{TextsSent}}{Number of texts sent (previous day)}
    \item{\code{HangHours}}{Hours last week spent hanging out with friends}
    \item{\code{HWHours}}{Hours last week spent doing homework}
    \item{\code{SportsHours}}{Hours last week spent playing sports or outdoor activities}
    \item{\code{VideoGameHours}}{Hours last week spent playing computer/video games}
    \item{\code{ComputerHours}}{Hours last week spent using a computer}
    \item{\code{TVHours}}{Hours last week spent watching TV}
    \item{\code{WorkHours}}{Hours last week spent working at a paid job}
    \item{\code{SchoolPressure}}{Amount of pressure due to schoolwork}
    \item{\code{Superpower}}{Most desired superpower (\code{Fly}, \code{Freeze time}, \code{Invisibility}, \code{Super strength}, or \code{Telepathy})}
    \item{\code{Preference}}{Prefers to be \code{Famous}, \code{Happy}, \code{Healthy}, or \code{Rich}}
  }
}
\details{
The dataset gives responses for a sample of high school seniors in Pennsylvania who participated in the Census at Schools project.
}
\source{
Data from U.S. Census at School (\emph{https://ww2.amstat.org/censusatschool/}) downloaded and used with the permission of the American Statistical Association.
}
\keyword{datasets}
