\name{PizzaGirl}
\alias{PizzaGirl}
\docType{data}
\title{Pizza Girl Tips}
\description{
Data on tips for pizza deliveries
}
%\usage{data(PizzaGirl)}
\format{
  A dataset with 24 observations on the following 2 variables.
  \tabular{rl}{
    \code{Tip} \tab {Amount of tip (in dollars)}\cr
    \code{Shift} \tab {Data coillected over three different shifts}\cr
  }
}
\details{
"Pizza Girl" collected data on her deliveries and tips over three different evening shifts. 
}
\source{
Pizza Girl: Statistical Analysis at http://slice.seriouseats.com/archives/2010/04/statistical-analysis-of-a-pizza-delivery-shift-20100429.html.
}

\keyword{datasets}
