\name{ProcessedFoods}
\alias{ProcessedFoods}
\docType{data}
\title{
Ultra-Processed vs Unprocessed Food Diets}

\description{
Results from an experiment comparing diets with ultra-processed food to unprocessed foods
}

\format{
  A data frame with 20 observations on the following 14 variables.
  \describe{
    \item{\code{WeightGain}}{Difference in weight gain}
    \item{\code{Kcal}}{Difference in average daily caloric consumption}
    \item{\code{Sugar}}{Difference in average daily sugar consumption in grams}
    \item{\code{Fiber}}{Difference in average daily fiber consumption in grams}
    \item{\code{Protein}}{Difference in average daily protein consumption in grams}
    \item{\code{Fat}}{Difference in average daily fat consumption in grams}
    \item{\code{Carbs}}{Difference in average daily carbohydrate consumption in grams}
    \item{\code{Ghrelin}}{Difference in active ghrelin hormone level}
    \item{\code{Leptin}}{Difference in leptin hormone level}
    \item{\code{PYY}}{Difference in PYY hormone level}
    \item{\code{Glucagon}}{Difference in glucagon hormone level}
    \item{\code{Hungry}}{Difference in average self-reported measure of feeling hungry}
    \item{\code{Satisfied}}{Difference in average self-reported measure of feeling satisfied}
    \item{\code{Fullness}}{Difference in average self-reported measure of feeling full}
    \item{\code{EatingCapacity}}{Difference in average self-reported measure of feeling hungry}

  }
}
\details{
Twenty adults were admitted to the National Institutes of Health (NIH) Clinical Center and had all of their food supplied for 4 weeks.  Each was given a diet of ultra-processed foods for two weeks, and a diet of unprocessed foods for two weeks, with the order of the diet randomized.  The diets as presented were matched for calories, sugar, fat, fiber, and micronutrients, but then participants could choose how much of each food to eat.  The cases are the 20 participants, and the variables give the difference in responses, ultra-processed response - unprocessed response, for each person.
}
\source{
Hall KD, et al. (2019). ``Ultra-Processed Diets Cause Excess Calorie Intake and Weight Gain: An Inpatient Randomized Controlled Trial of Ad Libitum Food Intake," Cell Metabolism, 30(1): 67-77.  Thanks to the author for sharing the data.
}

\keyword{datasets}
