\name{PublicLibraries}
\alias{PublicLibraries}
\docType{data}
\title{Public Library Usage}
\description{
Frequency of public library use by political party
}
\format{
  A data frame with 2429 observations on the following 2 variables.
  \describe{
    \item{\code{LibraryUse}}{Coded as \code{Monthly}, \code{Few Times}, \code{Rarely}, or \code{Never}}
    \item{\code{Party}}{Political party (\code{Dem}=Democrat,\code{Ind}=Independent, \code{Rep}=Repubican)}
   }
}
\details{
In a survey conducted by YouGov in April 2024, a sample of adults answered a question on how frequently they use public libraries and their polical party affiliation
}
\source{
Van Dam, A, "Who uses public libraries the most?  There’s a divide by religion, and politics."" The Washington Post, October 2, 2024.  Some values approximated from data available at \url{https://today.yougov.com/}.
}
\keyword{datasets}
