\name{PumpkinBeer}
\alias{PumpkinBeer}
\docType{data}
\title{Pumpkin Beer }
\description{
Ratings of different kinds of pumpkin beer by a wife and husband 
}
\format{
  A data frame with 18 observations on the following 8 variables.
  \describe{
    \item{\code{Name}}{Name of pumpkin beer}
    \item{\code{Brewer}}{Name of brewery that produced the beer}
    \item{\code{WifeRating}}{Rating on a 0-10 scale by the wife}
    \item{\code{HusbandRating}}{Rating on a 0-10 scale by the husband}
    \item{\code{WifeComments}}{Text of comments by the wife}
    \item{\code{HusbandComments}}{Text of comments by the husband}
    \item{\code{Average}}{Average of the two ratings (wife and husband)}
    \item{\code{Year}}{Year the ratings were done (2011 to 2019)}
  }
}
\details{
A Lock wife and husband are fans of pumpkin flavored beer, so they have each rated a variety of different brands of pumpkin beer over the years. 
}
\source{
Personal records 
}
\keyword{datasets}
