\name{QuizPulse10}
\alias{QuizPulse10}
\docType{data}
\title{
Quiz vs Lecture Pulse Rates
}
\description{
Paired data with pulse rates in a lecture and during a quiz for 10 students
}
\usage{data(QuizPulse10)}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{Student}}{ID number for the student}
    \item{\code{Quiz}}{Pulse rate (beats per minute) during a quiz}
    \item{\code{Lecture}}{Pulse rate (beats per minute) during a lecture}
  }
}
\details{
Ten students in an introductory statistics class measured their pulse rate in two settings: in the middle of a regular class lecture and again while taking a quiz. 
}
\source{
In-class data collection
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(QuizPulse10)
}
\keyword{datasets}
