\name{RestaurantTips}
\alias{RestaurantTips}
\docType{data}
\title{
Restaurant Tips - 11 West
}
\description{
Bill and tip data from the restaurant 11 West
}
%\usage{data(RestaurantTips)}
\format{
  A dataset with 300 observations on the following 7 variables.
  \tabular{rl}{
    \code{Bill} \tab {Size of the bill (in dollars)} \cr
    \code{Tip} \tab {Size of the tip (in dollars)} \cr
    \code{Guests} \tab {Number of people in the group} \cr
    \code{Day} \tab {Day of the week} \cr
    \code{Server} \tab {Code for specific wait person: \code{A},  \code{B},
         \code{C}, \code{D},  \code{E},  \code{F},  \code{G}, or \code{H},} \cr
    \code{Time} \tab {\code{Lunch} or \code{Dinner}} \cr
    \code{PctTip} \tab {Tip as a percentage of the bill} \cr
  }
}
\details{
Data from a sample of bills with tips during a week in March 2025 from 11 West, a restaurant in
Canton, NY. \cr
** Updated for 4e (earlier version for First Crush Bistro is now RestaurantTips1e).  **
}
\source{
Thanks to Mike Frazer and Les Baker from 11 West for providing the tipping data.
}
\keyword{datasets}
