\name{RetailSales2011}
\alias{RetailSales2011}
\docType{data}
\title{
Retail Sales (2000-2011)
}
\description{
Monthly U.S. Retail Sales
}
%\usage{data(RetailSales2011)}
\format{
  A dataset with 136 observations on the following 3 variables.
  \tabular{rl}{
    \code{Month} \tab {Month of the year} \cr
    \code{Year} \tab {Years from 2000 to 2011 } \cr
    \code{Sales} \tab {U.S. retail sales (in billions of dollars)} \cr
  }
}
\details{
Data show the monthly retail sales (in billions) for the U.S. economy in each month from January 2000 through April 2011.\cr
** From 1e - dataset has been updated for 2e and 3e **
}
\source{
Data downloaded from \emph{http://www.census.gov/retail/}
}

\keyword{datasets}
