\name{RockandRoll2015}
\alias{RockandRoll2015}
\docType{data}
\title{
Rock & Roll Hall of Fame (2015)
}
\description{
Groups and Individuals in the Rock and Roll Hall of Fame (2015)
}
%\usage{data(RockandRoll2015)}
\format{
  A dataset with 303 observations on the following 4 variables.
  \describe{
    \item{\code{Inductee}}{Name of the group or individual}
    \item{\code{FemaleMembers}}{\code{Yes} if individual or member of the group is female, otherwise \code{No}}
    \item{\code{Category}}{Type of individual or group: \code{Early Influence}, \code{Lifetime Achievement}, \code{Non-performer}, \code{Performer}, or \code{Sideman}}
    \item{\code{People}}{Number of people in the group}
  }
}
\details{
All inductees of the Rock & Roll Hall of Fame as of 2015. \cr
** From 2e - dataset has been updated for 3e **
}
\source{
Rock & Roll Hall of Fame website, \emph{http://rockhall.com/inductees/alphabetical/}
}

\keyword{datasets}
