\name{RockandRoll2019}
\alias{RockandRoll2019}
\docType{data}
\title{Rock & Roll Hall of Fame (2019)}
\description{
Groups and Individuals in the Rock and Roll Hall of Fame as of 2019
}
\format{
  A data frame with 329 observations on the following 4 variables.
  \describe{
    \item{\code{Inductee}}{Name of the group or individual}
    \item{\code{FemaleMembers}}{\code{Yes} if individual or member of the group is female, otherwise \code{No}}
    \item{\code{Category}}{Type of individual or group: \code{Early Influence}, \code{Lifetime Achievement}, \code{Non-performer}, \code{Performer}, or \code{Sideman}}
    \item{\code{People}}{Number of people in the group}
  }
}
\details{
All inductees of the Rock & Roll Hall of Fame as of 2019. \cr
** Updated for 3e (earlier versions are now RockandRoll2015 and RockandRoll1e) **

}
\source{
Rock & Roll Hall of Fame website, \url{https://www.rockhall.com/inductees}
}
\keyword{datasets}
