\name{RockandRoll2024}
\alias{RockandRoll2024}
\docType{data}
\title{Rock & Roll Hall of Fame (2024)}
\description{
Groups and Individuals in the Rock and Roll Hall of Fame as of 2024
}
\format{
  A data frame with 393 observations on the following 4 variables.
  \describe{
    \item{\code{Inductee}}{Name of the group or individual}
    \item{\code{FemaleMembers}}{\code{Yes} if individual or member of the group is female, otherwise \code{No}}
    \item{\code{Category}}{Type of individual or group: \code{Early Influence}, \code{Lifetime Achievement}, \code{Non-performer}, \code{Performer}, or \code{Sideman}}
    \item{\code{People}}{Number of people in the group}
  }
}
\details{
All inductees of the Rock & Roll Hall of Fame as of 2024. \cr
** Updated for 4e (earlier versions are now RockandRoll2019, RockandRoll2015, and RockandRoll2012) **

}
\source{
Rock & Roll Hall of Fame website, \url{https://www.rockhall.com/inductees}
}
\keyword{datasets}
