\name{RollerCoasters}
\alias{RollerCoasters}
\docType{data}
\title{Roller Coasters}
\description{
  Characteristics of a sample of roller coasters in the United States
}
\format{
  A data frame with 157 observations on the following 13 variables.
  \describe{
    \item{\code{Coaster}}{Name of the roller coaster }
    \item{\code{Park}}{Name of the amusement park}
    \item{\code{City}}{City where the coaster is located}
    \item{\code{State}}{State where the coaster is located}
    \item{\code{Type}}{\code{Wood} or \code{Steel}}
    \item{\code{Design}}{\code{Sit Down}, \code{Stand Up}, \code{Inverted}, etc. }
    \item{\code{Year}}{Year the coaster was put into service}
    \item{\code{Speed}}{Top speed (in mph)}
    \item{\code{Height}}{Vertical distance to the tallest point (in feet)}
    \item{\code{Drop}}{Maximum vertical drop (in feet)}
    \item{\code{Length}}{Total length of the ride (in feet)}
    \item{\code{Duration}}{Duration of the ride (in seconds)}
    \item{\code{Inversions}}{Number of inversions}
  }
}
\details{
Information on a sample of roller coasters at amusement parks in the US.
}
\source{
Roller coaster data obtained from a CODAP example at  \url{https://codap.concord.org/}. Accessed May 2025. Much of the data came originally from the Roller Coaster Database \url{https://rcdb.com/}.
}
\keyword{datasets}
