\name{SalaryBySex}
\alias{SalaryBySex}
\docType{data}
\title{Salary and Sex - 2023}
\description{
Salaries for college teachers (2023)
}
%\usage{data(SalaryBySex)}
\format{
  A dataset with 200 observations on the following 4 variables.
  \tabular{rl}{
    \code{Salary} \tab {Annual salary in $1,000's}\cr
    \code{Sex} \tab {0=female or 1=male}\cr
    \code{Age} \tab {Age in years}\cr
    \code{PhD} \tab {1=have PhD or 0=no PhD}\cr
  }
}
\details{
A random sample of postsecondary teachers taken from the 2023 American Community Survey (ACS) 1-year Public Use Microdata Sample (PUMS). Updated for 4e (earler version from 2010 is SalaryGender).
}
\source{
Downloaded from \emph{https://www.census.gov/programs-surveys/acs/data/pums.html}
}

\keyword{datasets}
