\name{SalaryGender}
\alias{SalaryGender}
\docType{data}
\title{Salary and Gender}
\description{
Salaries for college teachers
}
%\usage{data(SalaryGender)}
\format{
  A dataset with 100 observations on the following 4 variables.
  \tabular{rl}{
    \code{Salary} \tab {Annual salary in $1,000's}\cr
    \code{Gender} \tab {0=female or 1=male}\cr
    \code{Age} \tab {Age in years}\cr
    \code{PhD} \tab {1=have PhD or 0=no PhD}\cr
  }
}
\details{
A random sample of college teachers taken from the 2010 American Community Survey (ACS) 1-year Public Use Microdata Sample (PUMS).
}
\source{
http://www.census.gov/acs/www/data documentation/public use microdata sample/
}

\keyword{datasets}
