\name{SampCountries1e}
\alias{SampCountries1e}
\docType{data}
\title{
Sample of Countries - 1e
}
\description{
Data on a sample of fifty countries of the world (2008)
}
%\usage{data(SampCountries)}
\format{
  A dataset with 50 observations on the following 13 variables.
  \tabular{rl}{
   \code{Country} \tab {Name of the country} \cr
   \code{LandArea} \tab {Size in sq. kilometers} \cr
   \code{Population} \tab {Population in millions} \cr
   \code{Energy} \tab {Energy usage (kilotons of oil)} \cr
   \code{Rural} \tab {Percentage of population living in rural areas} \cr
   \code{Military} \tab {Percentage of government expenditures directed toward the military} \cr
   \code{Health} \tab {Percentage of government expenditures directed towards healthcare} \cr
   \code{HIV} \tab {Percentage of the population with HIV} \cr
   \code{Internet} \tab {Percentage of the population with access to the internet} \cr
   \code{Developed} \tab {Categories for kilowatt hours per capita: \code{1}= under 2500, \code{2}=2500 to 5000, \code{3}=over 5000} \cr
   \code{BirthRate} \tab {Births per 1000 people} \cr
   \code{ElderlyPop} \tab {Percentage of the population at least 65 years old} \cr
   \code{LifeExpectancy} \tab {Average life expectancy (in years)} \cr
  }
}
\details{
A subset of data from AllCountries for a random sample of 50 countries in 2008.\cr
** From 1e - dataset has been updated for 2e and 3e **
}
\source{
Data collected from the World Bank website, \emph{http://www.worldbank.org}.
}

\keyword{datasets}
