\name{SandP500}
\alias{SandP500}
\docType{data}
\title{S&P 500 Prices}
\description{
Daily data for S&P 500 Stock Index in 2024
}
\format{
  A data frame with 252 observations on the following 6 variables.
  \describe{
    \item{\code{Date}}{Trading date (mm/dd/yyy)}
    \item{\code{Open}}{Opening value}
    \item{\code{High}}{High point for the day}
    \item{\code{Low}}{Low point for the day}
    \item{\code{Close}}{Closing value}
    \item{\code{Volume}}{Shares traded (in millions)}
  }
}
\details{
Daily prices for the S&P 500 Stock Index for trading days in 2024. \cr
** Updated for 4e (earlier versions are SandP5003e from 2018, SandP5002e from 2014, and  SandP5001e from 2010). **

}
\source{
Downloaded from \url{https://stooq.com}
}
\keyword{datasets}
