\name{SandP5003e}
\alias{SandP5003e}
\docType{data}
\title{S&P 500 Prices}
\description{
Daily data for S&P 500 Stock Index in 2018
}
\format{
  A data frame with 251 observations on the following 6 variables.
  \describe{
    \item{\code{Date}}{Trading date (mm/dd/yyy)}
    \item{\code{Open}}{Opening value}
    \item{\code{High}}{High point for the day}
    \item{\code{Low}}{Low point for the day}
    \item{\code{Close}}{Closing value}
    \item{\code{Volume}}{Shares traded (in millions)}
  }
}
\details{
Daily prices for the S&P 500 Stock Index for trading days in 2018. \cr
** Updated for 3e (earlier versions are SandP5002e from 2014 and  SandP5001e from 2010). **

}
\source{
Downloaded from \emph{https://finance.yahoo.com/quote/^GSPC/history?ltr=1}
}
\keyword{datasets}
