\name{SandwichAnts}
\alias{SandwichAnts}
\docType{data}
\title{
Sandwich Ants
}
\description{
Ant Counts on samples of different sandwiches
}
\usage{data(SandwichAnts)}
\format{
  A data frame with 24 observations on the following 5 variables.
  \describe{
    \item{\code{Butter}}{a factor with levels \code{no}}
    \item{\code{Filling}}{a factor with levels \code{Ham \& Pickles} \code{Peanut Butter} \code{Vegemite}}
    \item{\code{Bread}}{a factor with levels \code{Multigrain} \code{Rye} \code{White} \code{Wholemeal}}
    \item{\code{Ants}}{a numeric vector}
    \item{\code{Order}}{a numeric vector}
  }
}
\details{
Only sandwiches with no butter in this dataset
}
\source{
Margaret Mackisack, ``Favourite Experiments: An Addendum to What is the Use of Experiments Conducted by Statistics Students?", 
Journal of Statistics Education (1994) http://www.amstat.org/publications/jse/v2n1/mackisack.supp.html
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(SandwichAnts)
}
\keyword{datasets}
