\name{SandwichAnts2}
\alias{SandwichAnts2}
\docType{data}
\title{
Sandwich Ants - Part 2
}
\description{
Ant Counts on samples of different sandwiches
}
\usage{data(SandwichAnts2)}
\format{
  A data frame with 48 observations on the following 5 variables.
  \describe{
    \item{\code{Butter}}{a factor with levels \code{no} \code{yes}}
    \item{\code{Fillings}}{a factor with levels \code{Ham & Pickles} \code{Peanut Butter} \code{Vegemite}}
    \item{\code{Breads}}{a factor with levels \code{Multigrain} \code{Rye} \code{White} \code{Wholemeal}}
    \item{\code{Ants}}{a numeric vector}
    \item{\code{Order}}{a numeric vector}
  }
}
\details{
This version has sandwiches with and without butter
}
\source{
Margaret Mackisack, ``Favourite Experiments: An Addendum to What is the Use of Experiments Conducted by Statistics Students?", 
Journal of Statistics Education (1994) http://www.amstat.org/publications/jse/v2n1/mackisack.supp.html
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(SandwichAnts2)
}
\keyword{datasets}
