\name{SkateboardPrices}
\alias{SkateboardPrices}
\docType{data}
\title{Skateboard Prices in (2025)}
\description{
Prices of skateboards for sale online in 2025
}
%\usage{data(SkateboardPrices)}
\format{
  A dataset with 20 observations on the following variable.
  \tabular{rl}{
    \code{Price} \tab {Selling price in dollars (including delivery)}\cr
  }
}
\details{
Prices (including delivery) for skateboards offered for sale on eBay.\cr
**Updated for 4e (earlier version is nowSkateboardPrices2012)  **
}
\source{
Random sample taken from skateboards available for sale on eBay in March 2025.
}

\keyword{datasets}
