\name{SkateboardPrices2012}
\alias{SkateboardPrices2012}
\docType{data}
\title{Skateboard Prices (2012)}
\description{
Prices of skateboards for sale online in 2012
}
%\usage{data(SkateboardPrices)}
\format{
  A dataset with 20 observations on the following variable.
  \tabular{rl}{
    \code{Price} \tab {Selling price in dollars}\cr
  }
}
\details{
Prices for skateboards offered for sale on eBay.
}
\source{
Random sample taken from all skateboards available for sale on eBay on February 12, 2012.
}

\keyword{datasets}
