\name{SleepStudy}
\alias{SleepStudy}
\docType{data}
\title{Sleep Study}
\description{
Data from a study of sleep patterns for college students. 
}
%\usage{data(SleepStudy)}
\format{
  A dataset with 253 observations on the following 27 variables.
  \tabular{rl}{
    \code{Gender} \tab {1=male, 0=female}\cr
    \code{ClassYear} \tab {Year in school, 1=first year, ..., 4=senior}\cr
    \code{LarkOwl} \tab {Early riser or night owl?  \code{Lark}, \code{Neither}, or \code{Owl}}\cr
    \code{NumEarlyClass} \tab {Number of classes per week before 9 am}\cr
    \code{EarlyClass} \tab {Indicator for any early classes}\cr
    \code{GPA} \tab {Grade point average (0-4 scale)}\cr
    \code{ClassesMissed} \tab {Number of classes missed in a semester}\cr
    \code{CognitionZscore} \tab {Z-score on a test of cognitive skills}\cr
    \code{PoorSleepQuality} \tab {Measure of sleep quality (higher values are poorer sleep)}\cr
    \code{DepressionScore} \tab {Measure of degree of depression}\cr
    \code{AnxietyScore} \tab {Measure of amount of anxiety}\cr
    \code{StressScore} \tab {Measure of amount of stress}\cr
    \code{DepressionStatus} \tab {Coded depression score: \code{normal}, \code{moderate}, or \code{severe}}\cr
    \code{AnxietyStatus} \tab {Coded anxiety score: \code{normal}, \code{moderate}, or \code{severe}}\cr
    \code{Stress} \tab {Coded stress score: \code{normal} or  \code{high}}\cr
    \code{DASScore} \tab {Combined score for depression, anxiety and stress}\cr
    \code{Happiness} \tab {Measure of degree of happiness}\cr
    \code{AlcoholUse} \tab {Self-reported: \code{Abstain},  \code{Light}, \code{Moderate}, or \code{Heavy}}\cr
    \code{Drinks} \tab {Number of alcoholic drinks per week}\cr
    \code{WeekdayBed} \tab {Average weekday bedtime (24.0=midnight)}\cr
    \code{WeekdayRise} \tab {Average weekday rise time (8.0=8 am)}\cr
    \code{WeekdaySleep} \tab {Average hours of sleep on weekdays}\cr
    \code{WeekendBed} \tab {Average weekend bedtime (24.0=midnight)}\cr
    \code{WeekendRise} \tab {Average weekend rise time (8.0=8 am)}\cr
    \code{WeekendSleep} \tab {Average weekend bedtime (24.0=midnight)}\cr
    \code{AverageSleep} \tab {Average hours of sleep for all days}\cr
    \code{AllNighter} \tab {Had an all-nighter this semester?  1=yes, 0=no}\cr
  }
}
\details{
The data were obtained from a sample of students who did skills tests to measure cognitive function, 
completed a survey that asked many questions about attitudes and habits, 
and kept a sleep diary to record time and quality of sleep over a two week period.
}
\source{
Onyper, S., Thacher, P., Gilbert, J., Gradess, S., "Class Start Times, Sleep, and Academic Performance in College: 
A Path Analysis," April 2012; 29(3): 318-335.  Thanks to the authors for supplying the data.
}

\keyword{datasets}
