\name{Smiles}
\alias{Smiles}
\docType{data}
\title{
Smiles
}
\description{
Experiment to study effect of smiling on leniency in judicial matters
}
\usage{data(Smiles)}
\format{
  A data frame with 68 observations on the following 2 variables.
  \describe{
    \item{\code{Leniency}}{Score assigned by a judgement panel (higher is more lenient)}
    \item{\code{Group}}{Treatment group: \code{neutral} or \code{smile}}
  }
}
\details{
Hecht and LeFrance conducted a study examining the effect of a smile on the leniency of disciplinary action for wrongdoers.
Participants in the experiment took on the role of members of a college disciplinary panel judging students accused of cheating. 
For each suspect, along with a description of the offense, a picture was provided with
either a smile or neutral facial expression. 
A leniency score was calculated based on the disciplinary decisions
made by the participants.
}
\source{
LaFrance, M., & Hecht, M. A., "Why smiles generate leniency", Personality and Social Psychology Bulletin, 21, 1995,
207-214.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(Smiles)
}
\keyword{datasets}
