\name{SpeedDating}
\alias{SpeedDating}
\docType{data}
\title{Speed Dating}
\description{
Data from a sample of four minute speed dates. 
}
%\usage{data(SpeedDating)}
\format{
  A dataset with 276 observations on the following 22 variables.
  \tabular{rl}{
    \code{DecisionM} \tab {Would the male like another date?  1=yes 0=no}\cr
    \code{DecisionF} \tab {Would the female like another date? 1=yes 0=no}\cr
    \code{LikeM} \tab {How much the male likes his partner (1-10 scale)}\cr
    \code{LikeF} \tab {How much the female likes her partner (1-10 scale)}\cr
    \code{PartnerYesM} \tab {Male's estimate of chance the female wants another date (1-10 scale)}\cr
    \code{PartnerYesF} \tab {Female's estimate of chance the male wants another date (1-10 scale)}\cr
    \code{AgeM} \tab {Male's age (in years)}\cr
    \code{AgeF} \tab {Females age (in years)}\cr
    \code{RaceM} \tab {Male's race: \code{} \code{Asian} \code{Black} \code{Caucasian} \code{Latino} \code{Other}}\cr
    \code{RaceF} \tab {Female's race: \code{} \code{Asian} \code{Black} \code{Caucasian} \code{Latino} \code{Other}}\cr
    \code{AttractiveM} \tab {Male's rating of female's attractiveness (1-10 scale)}\cr
    \code{AttractiveF} \tab {Female's rating of male's attractiveness (1-10 scale)}\cr
    \code{SincereM} \tab {Male's rating of female's sincerity (1-10 scale)}\cr
    \code{SincereF} \tab {Female's rating of male's sincerity (1-10 scale)}\cr
    \code{IntelligentM} \tab {Male's rating of female's intelligence (1-10 scale)}\cr
    \code{IntelligentF} \tab {Female's rating of male's intelligence (1-10 scale)}\cr
    \code{FunM} \tab {Male's rating of female as fun (1-10 scale)}\cr
    \code{FunF} \tab {Female's rating of male as fun (1-10 scale)}\cr
    \code{AmbitiousM} \tab {Male's rating of female's ambition (1-10 scale)}\cr
    \code{AmbitiousF} \tab {Female's rating of male's ambition (1-10 scale)}\cr
    \code{SharedInterestsM} \tab {Male's rating of female's shared interests (1-10 scale)}\cr
    \code{SharedInterestsF} \tab {Female's rating of male's shared interests (1-10 scale)}\cr
  }
}
\details{
Participants were students at Columbia's graduate and professional schools, recruited by mass email, posted
fliers, and fliers handed out by research assistants. Each participant attended one speed dating session, in
which they met with each participant of the opposite sex for four minutes. Order and session assignments
were randomly determined. After each four minute "speed date," participants filled out a form rating their
date on a scale of 1-10 on various attributes.  Only data from the first date in each session is recorded here.
}
\source{
Gelman, A. and Hill, J., Data analysis using regression and multilevel/hierarchical models, Cambridge University Press: New York, 2007
}

\keyword{datasets}
