\name{SplitBill}
\alias{SplitBill}
\docType{data}
\title{
Split Bill vs Individual Meal Costs
}
\description{
Meal costs when ordering individually vs splitting a bill
}
%\usage{data(SplitBill)}
\format{
  A dataset with 48 observations on the following 4 variables.
  \tabular{rl}{
    \code{Payment} \tab {Payment method: \code{Individual} or \code{Split} }\cr
    \code{Sex} \tab {\code{F} = female or \code{M} = male}\cr
    \code{Items} \tab {Number of items ordered}\cr
    \code{Cost} \tab {Cost of items ordered in Israeli new shekel's (ILS)}\cr
  }
}
\details{
Subjects were 48 Israeli students who were randomly assigned to eat in groups of six (three males and three females) at a restaurant.  Half the groups were told that they would pay for meals individually and half were told that the group would split the bill equally. The number of items ordered and cost (in Israeli new shekels) was recorded for each individual.
}
\source{
Gneezy, U.,Haruvy, E., and Yafe, H. "The Inefficiency of Splitting the Bill,"" The Economic Journal, 2004; 114, 265-280.
}

\keyword{datasets}
