\name{StatGrades}
\alias{StatGrades}
\docType{data}
\title{Statistics Exam Grades}
\description{
Grades on statistics exams
}
%\usage{data(StatGrades)}
\format{
  A dataset with 50 observations on the following 3 variables.
  \tabular{rl}{
    \code{Exam1} \tab {Score (out of 100 points) on the first exam}\cr
    \code{Exam2} \tab {Score (out of 100 points) on the first exam}\cr
    \code{Final} \tab {Score (out of 100 points) on the final exam}\cr
  }
}
\details{
exam
Scores for a sample of students who completed a course using Statistics: Unlocking the Power of Data as a text. 
 The dataset contains scores on
Exam1 (Chapters 1 to 4), Exam2 (Chapters 5 to 8), and the Final exam (entire book).
}
\source{
Random selection of students in an introductory statistics course. 
}

\keyword{datasets}
