\name{StockChanges}
\alias{StockChanges}
\docType{data}
\title{
Stock Changes (2024)
}
\description{
Stock price change for a sample of stocks from the S&P 500 (December 2-6, 2024)
}
%\usage{data(StockChanges)}
\format{
  A dataset with 50 observations on the following 2 variables.
  \describe{
    \item{\code{Symbol}}{Ticker symbol for the stock}
    \item{\code{SPChange}}{Change in stock price (in dollars)}
  }
}
\details{
A random sample of 50 companies from Standard & Poor's index of 500 companies was selected. The change in the price of the stock (in dollars)
over the 5-day period from December 2 - 6, 2024 was recorded for each company in the sample.
}
\source{
Data obtained from Kaggle at \url{https://www.kaggle.com/datasets/andrewmvd/sp-500-stocks}
}

\keyword{datasets}
