\name{StudentSurvey}
\alias{StudentSurvey}
\docType{data}
\title{Student Survey Data}
\description{
Data from a survey of students in introductory statistics courses
}
\format{
  A data frame with 362 observations on the following 17 variables.
  \describe{
    \item{\code{Year}}{Year in school}
    \item{\code{Sex}}{code {F}=female or \code{M}=male}
    \item{\code{Smoke}}{Smoker? \code{No} or \code{Yes}}
    \item{\code{Award}}{Preferred award: \code{Academy}, \code{Nobel}, or \code{Olympic}}
    \item{\code{HigherSAT}}{Which SAT is higher?  \code{Math} or \code{Verbal}}
    \item{\code{Exercise}}{Hours of exercise per week}
    \item{\code{TV}}{Hours of TV viewing per week}
    \item{\code{Height}}{Height (in inches)}
    \item{\code{Weight}}{Weight (in pounds)}
    \item{\code{Siblings}}{Number of siblings}
    \item{\code{BirthOrder}}{Birth order, 1=oldest}
    \item{\code{VerbalSAT}}{Verbal SAT score}
    \item{\code{MathSAT}}{Math SAT scorer}
    \item{\code{SAT}}{Combined Verbal + Math SAT}
    \item{\code{GPA}}{College grade point average}
    \item{\code{Pulse}}{Pulse rate (beats per minute)}
    \item{\code{Piercings}}{Number of body piercings}
  }
}
\details{
Data from an in-class survey given to introductory statistics students over several years. Note the Sex variable was labeled as Gender in earlier versions of this dataset. We acknowledge that this binary dichotomization is not a complete or inclusive representation of reality.
}
\source{
In-class student survey
}
\keyword{datasets}
