\name{SynchronizedMovement}
\alias{SynchronizedMovement}
\docType{data}
\title{
Synchronized Movement
}
\description{
Effects of synchronized movement activities
}
%\usage{data(SynchronizedMovement)}
\format{
  A dataset with 264 observations on the following 11 variables.
  \tabular{rl}{
    \code{Sex} \tab {\code{f} = female or \code{m} = male}\cr
    \code{Group} \tab {Type of activity.  Coded as \code{HS+HE}, \code{HS+LE}, \code{LS+HE}, or \code{LS+LE}  }\cr
    \code{ } \tab {for High/Low Synchronization + High/Low Exertion}\cr
    \code{Synch} \tab {Synchronized activity? \code{yes} or \code{no} }\cr
    \code{Exertion} \tab {Exertion level: \code{high} or \code{low}}\cr
    \code{PainToleranceBefore} \tab {Measure of pain tolerance (mm Hg) before activity}\cr
    \code{PainTolerance} \tab {Measure of pain tolerance (mm Hg) after activity}\cr
    \code{PainTolDiff} \tab {Difference (after - before) in pain tolerance}\cr
    \code{MaxPressure} \tab {Reached the maximum pressure (300 mm Hg) when testing pain tolerance (after)}\cr
    \code{CloseBefore} \tab {Rating of closeness to the group before activity (1=least close to 7=most close)}\cr
    \code{CloseAfter} \tab {Rating of closeness to the group after activity (1=least close to 7=most close)}\cr
    \code{CloseDiff} \tab {Change on closeness rating (after - before)}\cr
  }
}
\details{
From a study of 264 high school students in Brazil to examine the effect of doing synchronized movements (such as marching in step or doing synchronized dance steps) and the effect of exertion on variables, such as pain tolerance and attitudes towards others. Students were randomly assigned to activities that involved synchronized or non-synchronized movements involving high or low levels of exertion.  Pain tolerance was measured with a blood pressure cuff, going to a maximum possible reading of 300 mmHg.
}
\source{
Tarr B, Launay J, Cohen E, and Dunbar R, "Synchrony and exertion during dance independently raise pain threshold and encourage social bonding," Biology Letters, 11(10), October 2015.
}

\keyword{datasets}
