\name{TenCountries}
\alias{TenCountries}
\docType{data}
\title{Ten Countries}
\description{
A subset of the \code{AllCountries} data for a random sample of ten countries
}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Country}}{Country name}
    \item{\code{Code}}{Three-letter country code}
    \item{\code{Area}}{Size in 1000 sq. kilometers}
    \item{\code{PctRural}}{Percentage of population living in rural areas}
  }
}
\details{
Area and percent rural for a sample of ten countries from AllCountries dataset.\cr
** Updated for 3e (earlier versions are now TenCountries2e and TenCountries1e) **
}
\source{
Data collected from the World Bank website, \url{https://www.worldbank.org/en/home}
}
\keyword{datasets}
