\name{TenCountries2e}
\alias{TenCountries2e}
\docType{data}
\title{Ten Countries - 2e}
\description{
A subset of the \code{AllCountries} data for a random sample of ten countries
}
%\usage{data(TenCountries2e)}
\format{
  A dataset with 10 observations on the following 4 variables.
  \tabular{rl}{
    \code{Country} \tab {Country name}\cr
    \code{Code} \tab {Three-letter country code}\cr
    \code{Area} \tab {Size in 1000 sq. kilometers}\cr
    \code{PctRural} \tab {Percentage of population living in rural areas}\cr
  }
}
\details{
Area and percent rural for a sample of ten countries from AllCountries dataset.\cr
** From 2e - dataset has been updated for 3e **
}
\source{
Data collected from the World Bank website, \emph{http://www.worldbank.org}.
}

\keyword{datasets}
