\name{TextbookCosts}
\alias{TextbookCosts}
\docType{data}
\title{
Textbook Costs (2025)
}
\description{
Prices for textbooks for different courses in Spring 2025
}
%\usage{data(TextbookCosts)}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
    \item{\code{Field}}{General discipline of the course:  \code{Arts}, \code{Humanities}, \code{NaturalScience}, or \code{SocialScience}}
    \item{\code{Books}}{Number of books required}
    \item{\code{Cost}}{Total cost (in dollars) for required books}
  }
}
\details{
Data are from samples of ten courses in each of four disciplines at a liberal arts college.
For each course the bookstore's website lists the required texts(s) and costs (for new books).
Data were collected for the Spring 2025 semester.
}
\source{
Bookstore online site
}

\keyword{datasets}
