\name{TextbookCosts2011}
\alias{TextbookCosts2011}
\docType{data}
\title{
Textbook Costs (2011)
}
\description{
Prices for textbooks for different courses in Fall 2011
}
%\usage{data(TextbookCosts)}
\format{
  A data frame with 40 observations on the following 3 variables.
  \tabular{rl}{
    \code{Field} \tab {General discipline of the course:  \code{Arts}, \code{Humanities}, \code{NaturalScience}, or \code{SocialScience}} \cr
    \code{Books} \tab {Number of books required}\cr
    \code{Cost} \tab {Total cost (in dollars) for required books}\cr
  }
}
\details{
Data are from samples of ten courses in each of four disciplines at a liberal arts college.  For each course the bookstore's website lists the required texts(s) and costs for new books.
Data were collected for the Fall 2011 semester.
}
\source{
Bookstore online site
}

\keyword{datasets}
