\name{USStates2e}
\alias{USStates2e}
\docType{data}
\title{
US State Data - 2e
}
\description{
Various data for all 50 US States in 2014.
}
%\usage{data(USStates2e)}
\format{
  A dataset with 50 observations on the following 22 variables.
  \tabular{rl}{
    \code{State} \tab {State name} \cr
    \code{HouseholdIncome} \tab {Median household income (in $1,000's)} \cr
    \code{Region} \tab {MW=Midwest, NE=Northeast, S=South, W=West} \cr
    \code{Population} \tab {Number of residents (in millions for 2014)} \cr
    \code{EighthGradeMath} \tab {Average score NAEP mathematics for 8th-grade students (2013)} \cr
    \code{HighSchool} \tab {Percent of residents (ages 25-34) who are high school graduates} \cr
    \code{College} \tab {Percent of residents (ages 25-34) who are college graduates} \cr
    \code{IQ} \tab {Estimated mean IQ score of residents} \cr
    \code{GSP} \tab {Gross state product (in $1,000's per capita in 2013)} \cr
    \code{Vegetables} \tab {Percent of residents eating vegetables at least once per day} \cr
    \code{Fruit} \tab {Percent of residents eating fruit at least once per day} \cr
    \code{Smokers} \tab {Percent of residents who smoke} \cr
    \code{PhysicalActivity} \tab {Percent who do 150+ minutes of aerobic physical activity per week} \cr
    \code{Obese} \tab {Percent  obese residents (BMI 30+)} \cr
    \code{NonWhite} \tab {Percent nonwhite residents (in 2013)} \cr
    \code{HeavyDrinkers} \tab {Percent heavy drinkers (men: 3+ drinks/day, women 2+ drinks/day)} \cr
    \code{Electoral} \tab {Number of state votes in the presidential electoral college} \cr
    \code{ObamaVote} \tab {Proportion of votes for Obama in 2012 presidential election} \cr
    \code{ObamaRomney} \tab {State winner in 2012 presidential election (O=Obama, R=Romney)} \cr
    \code{TwoParents} \tab {Percent of children living in two-parent households} \cr
    \code{StudentSpending} \tab {School spending (in $1,000 per pupil in 2013)} \cr
    \code{Insured} \tab {Percent  of adults (ages 18-64) who have any kind of health coverage} \cr
  }
}
\details{
Information from each of the 50 states of the United States (from 2013 or 2014). \cr
** From 2e - dataset has been updated for 3e **
}
\source{
U.S. Census Bureau, 2009-2013 5-Year American Community Survey\cr
\emph{http://factfinder.census.gov/faces/tableservices/jsf/pages/} \cr
\emph{productview.xhtml?pid=ACS_13_5YR_DP03&src=pt}\cr
\emph{http://factfinder.census.gov/faces/tableservices/jsf/pages/}\cr
\emph{productview.xhtml?pid=ACS_13_5YR_S1501&src=pt} \cr
\emph{http://factfinder.census.gov/faces/tableservices/jsf/pages/}\cr
\emph{productview.xhtml?pid=ACS_13_5YR_B02001&prodType=table }\cr
\emph{http://factfinder.census.gov/faces/nav/jsf/pages/index.xhtml} (Table C23008)


}

\keyword{datasets}
