\name{WaterTaste}
\alias{WaterTaste}
\docType{data}
\title{
WaterTaste
}
\description{
Blind taste test to compare brands of bottled water
}
%\usage{data(WaterTaste)}
\format{
  A dataset with 100 observations on the following 10 variables.
  \tabular{rl}{
    \code{Gender} \tab {Gender of respondent: \code{F}=Female \code{M}=Male}\cr
    \code{Age} \tab {Age (in years)}\cr
    \code{Class} \tab {Year in school \code{F}=First year \code{J}=Junior \code{O}=Other \code{P} \code{SO}=Sophomore \code{SR}=Senior}\cr
    \code{UsuallyDrink} \tab {Usual source of drinking water: \code{Bottled}, \code{Filtered}, or \code{Tap}}\cr
    \code{FavBotWatBrand} \tab {Favorite brand of bottled water }\cr
    \code{Preference} \tab {Order of perference: \code{A}=Sams Choice, \code{B}=Aquafina, \code{C}=Fiji, and \code{D}=Tap water }\cr
    \code{First} \tab {Top choice among \code{Aquafina}, \code{Fiji}, \code{SamsChoice}, or \code{Tap}}\cr
    \code{Second} \tab {Second choice}\cr
    \code{Third} \tab {Third choice }\cr
    \code{Fourth} \tab {Fourth choice}\cr
  }
}
\details{
Result from a blind taste test comparing different four different types of water 
(Sam's Choice, Aqufina, Fiji, and tap water).  Participants rank ordered waters when presented in a random order.  
}
\source{
"Water Taste Test Data" by M. Leigh Lunsford and Alix D. Dowling Finch in the Journal of Statistics Education (Vol 18, No, 1) 2010
}
\references{
http://www.amstat.org/publications/jse/v18n1/lunsford.pdf
}

\keyword{datasets}
