\name{Wetsuits}
\alias{Wetsuits}
\docType{data}
\title{
Wetsuits
}
\description{
Swim velocity (for 1500 meters) with and withut wearing a wetsuit
}
%\usage{data(Wetsuits)}
\format{
  A dataset with 12 observations on the following 4 variables.
  \tabular{rl}{
    \code{Wetsuit} \tab {Maximum swim velocity (m/sec) when wearing a wetsuit}\cr
    \code{NoWetsuit} \tab {Maximum swim velocity (m/sec) when wearing a regular bathing suit}\cr
    \code{Gender} \tab {Gender of swimmer: \code{F} or \code{M}}\cr
    \code{Type} \tab {Type of athlete: \code{swimmer} or \code{triathlete}}\cr
  }
}
\details{
A study tested whether wearing wetsuits influences swimming velocity. 
Twelve competitive swimmers and triathletes swam 1500m at maximum speed twice each; once wearing a wetsuit and once wearing a regular bathing suit. 
The order of the trials was randomized. 
Each time, the maximum velocity in meters/sec of the swimmer was recorded.
}
\source{
de Lucas, R.D., Balildan, P., Neiva, C.M., Greco, C.C., Denadai, B.S. (2000). "The effects of wetsuits on physiological and
biomechanical indices during swimming," Journal of Science and Medicine in Sport, 3 (1): 1-8.
}

\keyword{datasets}
