\name{XylitolEffect}
\alias{XylitolEffect}
\docType{data}
\title{Xylitol vs. Sugar - Blood Clotting}
\description{
Blood clotting measures with Xylitol and Sugar
}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Subject}}{Subject ID number}
    \item{\code{Xylitol}}{Blood clotting measure after taking xylitol}
    \item{\code{Sugar}}{Blood clotting measure after taking sugar}
    \item{\code{Diff}}{Difference=Xylitol-Sugar}
  }
}
\details{
Researchers were interested in how xylitol might affect bloood clotting. Ten subjects
had a measure of blood clotting taken 30 minutes after
ingesting a drink sweetened with xylitol (a sugar substitute) and again after
a sugar-sweetened drink.
}
\source{
Witkowski M, et al., “Xylitol is prothrombotic and associated with
cardiovascular risk,” European Heart Journal, 45 (27), July 14, 2024. \cr
Values are estimated from information given in the paper

}
\keyword{datasets}
