\name{YoungBlood}
\alias{YoungBlood}
\docType{data}
\title{
Young Blood
}
\description{
Effects of transfusions of young blood on exercise endurance in mice
}
%\usage{data(YoungBlood)}
\format{
  A dataset with 30 observations on the following 2 variables.
  \tabular{rl}{
    \code{Plasma} \tab {Whether the blood came from a \code{Young} or \code{Old} mouse}\cr
    \code{Runtime} \tab {Maximum treadmill run time (in minutes) in a 90-minute window}\cr
  }
}
\details{
The data come from a study to see if transfusions of blood plasma from young mice (equivalent to about a 25-year-old person) can counteract or reverse brain aging in old mice (equivalent to about a 70-year-old person.)  Old mice were randomly assigned to receive plasma from either a young mice or another old mouse, and exercise endurance was measured.
}
\source{
Data come from two references, and are estimated from summary statistics and graphs.\cr
Sanders L, "Young blood proven good for old brain,""  Science News, 185(11), May 31, 2014. \cr
Manisha S, et al., "Restoring Systemic GDF11 Levels Reverses Age-Related Dysfunction in Mouse Skeletal Muscle," Science, 9 May 2014.
}

\keyword{datasets}
