\name{AllCountries}
\alias{AllCountries}
\docType{data}
\title{
AllCountries
}
\description{
Data on the countries of the world
}
\usage{data(AllCountries)}
\format{
  A data frame with 213 observations on the following 18 variables.
  \describe{
    \item{\code{Country}}{Name of the country}
    \item{\code{Code}}{Three letter country code}
    \item{\code{LandArea}}{Size in sq. kilometers}
    \item{\code{Population}}{Population in millions}
    \item{\code{Energy}}{Energy usage (kilotons of oil)}
    \item{\code{Rural}}{Percentage of population living in rural areas}
    \item{\code{Military}}{Percentage of government expenditures directed toward the military}
    \item{\code{Health}}{Percentage of government expenditures directed towards healthcare}
    \item{\code{HIV}}{Percentage of the population with HIV}
    \item{\code{Internet}}{Percentage of the population with access to the internet}
    \item{\code{Developed}}{Categories for kilowatt hours per capita, 1= under 2500, 2=2500 to 5000, 3=over 5000}
    \item{\code{BirthRate}}{Births per 1000 people}
    \item{\code{ElderlyPop}}{Percentage of the population at least 65 year old}
    \item{\code{LifeExpectancy}}{Average life expectancy (years)}
    \item{\code{CO2}}{CO2 emissions (metric tons per capita)}
    \item{\code{GDP}}{Gross Domestic Product (per capita)}
    \item{\code{Cell}}{Cell phone subscriptions (per 100 people)}
    \item{\code{Electricity}}{Electric power consumption (kWh per capita)}
  }
}
\details{
Most data from 2008 to avoid many missing values in more recent years.
}
\source{
Data collected from the World Bank website, worldbank.org. 
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%%data(AllCountries)
}
\keyword{datasets}
