\name{AllCountries}
\alias{AllCountries}
\docType{data}
\title{
AllCountries
}
\description{
Data on the countries of the world
}
%\usage{data(AllCountries)}
\format{
  A dataset with 213 observations on the following 18 variables.
  \tabular{rl}{
    \code{Country} \tab {Name of the country}\cr
    \code{Code} \tab {Three letter country code}\cr
    \code{LandArea} \tab {Size in sq. kilometers}\cr
    \code{Population} \tab {Population in millions}\cr
    \code{Energy} \tab {Energy usage (kilotons of oil)}\cr
    \code{Rural} \tab {Percentage of population living in rural areas}\cr
    \code{Military} \tab {Percentage of government expenditures directed toward the military}\cr
    \code{Health} \tab {Percentage of government expenditures directed towards healthcare}\cr
    \code{HIV} \tab {Percentage of the population with HIV}\cr
    \code{Internet} \tab {Percentage of the population with access to the internet}\cr
    \code{Developed} \tab {Categories for kilowatt hours per capita, 1= under 2500, 2=2500 to 5000, 3=over 5000}\cr
    \code{BirthRate} \tab {Births per 1000 people}\cr
    \code{ElderlyPop} \tab {Percentage of the population at least 65 year old}\cr
    \code{LifeExpectancy} \tab {Average life expectancy (years)}\cr
    \code{CO2} \tab {CO2 emissions (metric tons per capita)}\cr
    \code{GDP} \tab {Gross Domestic Product (per capita)}\cr
    \code{Cell} \tab {Cell phone subscriptions (per 100 people)}\cr
    \code{Electricity} \tab {Electric power consumption (kWh per capita)}\cr
  }
}
\details{
Most data from 2008 to avoid many missing values in more recent years.
}
\source{
Data collected from the World Bank website, worldbank.org. 
}

\keyword{datasets}
