\name{APMultipleChoice}
\alias{APMultipleChoice}
\docType{data}
\title{
AP Multiple Choice
}
\description{
Correct responses on Advanced  Placement multiple choice exams
}
\usage{data(APMultipleChoice)}
\format{
  A data frame with 400 observations on the following variable.
  \describe{
    \item{\code{Answer}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E}}
  }
}
\details{
Correct responses from multiple choice sections for a sample of Advanced Placement exams
}
\source{
Sample exams from several disciplines at http://apcentral.collegeboard.com
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(APMultipleChoice)
}
\keyword{datasets}
