\name{April14Temps}
\alias{April14Temps}
\docType{data}
\title{April 14th Temperatures}
\description{
Temperatures in Des Moines, IA and San Francisco, CA on April 14th
}
\format{
  A data frame with 25 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{1995 to 2019}
    \item{\code{DesMoines}}{Temperature in Des Moines (degrees F)}
    \item{\code{SanFrancisco}}{Temperature in San Francisco (degrees F)}
  }
}
\details{
Average temperature for the day of April 14th in each of 25 years from 1995-2019 \cr
** Data set updated for 3e (earlier versions are now April14Temps1e and April14Temps2e) **
}
\source{
The University of Dayton Average Daily Temperature Archive at
\url{https://academic.udayton.edu/kissock/http/Weather/citylistUS.htm}
}
\keyword{datasets}
