\name{April14Temps}
\alias{April14Temps}
\docType{data}
\title{
April 14th Temperatures
}
\description{
Temperatures in Des Moines, IA and San Francisco, CA on April 14th
}
%\usage{data(April14Temps)}
\format{
  A dataset with 21 observations on the following 3 variables.
  \tabular{rl}{
    \code{Year} \tab {1995 to 2015} \cr
    \code{DesMoines} \tab {Temperature in Des Moines (degrees F)} \cr
    \code{SanFrancisco} \tab {Temperature in San Francisco (degrees F)}
  }
}
\details{
Average temperature for the day of April 14th in each of 21 years from 1995-2015 \cr
** Data set updated for 2e (original is now April14Temps1e) **
}
\source{
The University of Dayton Average Daily Temperature Archive at \cr
\emph{http://academic.udayton.edu/kissock/http/Weather/citylistUS.htm}
}

\keyword{datasets}
