\name{April14Temps}
\alias{April14Temps}
\docType{data}
\title{
April 14th Temperatures
}
\description{
Temperatures in Des Moines, IA and San Francisco, CA on April 14th
}
\usage{data(April14Temps)}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{1995-2010}
    \item{\code{DesMoines}}{Temperature in Des Moines (degrees F)}
    \item{\code{SanFrancisco}}{Temperature in San Francisco (degrees F)}
  }
}
\details{
Average temperature for the day of April 14th in each of 16 years from 1995-2010
}
\source{
The University of Dayton Average Daily Temperature Archive at http://academic.udayton.edu/kissock/http/Weather/citylistUS.htm
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(April14Temps)
}
\keyword{datasets}
