\name{BodyFat}
\alias{BodyFat}
\docType{data}
\title{Body Measurements}
\description{
Various body measurements for a sample of 100 men to build models to predict percent body fat.
}
%\usage{data(BodyFat)}
\format{
  A dataset with 100 observations on the following 10 variables.
  \tabular{rl}{
    \code{Bodyfat} \tab {Percent body fat}\cr
    \code{Age} \tab {Age (in years)}\cr
    \code{Weight} \tab {Weight (in pounds)}\cr
    \code{Height} \tab {Height (in inches}\cr
    \code{Neck} \tab {Neck circumference (in cm)}\cr
    \code{Chest} \tab {Chest circumference (in cm)}\cr
    \code{Abdomen} \tab {Abdomen circumference (in cm)}\cr
    \code{Ankle} \tab {Ankle circumference (in cm)}\cr
    \code{Biceps} \tab {Biceps circumference (in cm)}\cr
    \code{Wrist} \tab {Wrist circumference (in cm)}\cr
  }
}
\details{
The percentage of a person's weight that is made up of body fat is often used as an indicator of health and
fitness. However, accurate methods of measuring percent body fat are difficult to implement. One method
involves immersing the body in water to estimate its density and then applying a formula to estimate percent
body fat. An alternative is to develop a model for percent body fat that is based on body characteristics
such as height and weight that are easy to measure. The dataset BodyFat contains such measurements for
a sample of 100 men.
}
\source{
A sample taken from data provided by Johnson R., "Fitting Percentage of Body Fat to Simple Body Measurements,"
Journal of Statistics Education, 1996, \cr
http://www.amstat.org/publications/jse/v4n1/datasets.johnson.html.
}

\keyword{datasets}
