\name{BodyTemp50}
\alias{BodyTemp50}
\docType{data}
\title{
Body Temperatures
}
\description{
Sample of 50 body temperatures
}
\usage{data(BodyTemp50)}
\format{
  A data frame with 50 observations on the following 3 variables.
  \describe{
    \item{\code{BodyTemp}}{Body temperatures in degrees F}
    \item{\code{Pulse}}{Pulse rates (beats per minute)}
    \item{\code{Gender}}{F=Female, M=Male}
  }
}
\details{
Body temperatures and pulse rates for a sample of 50 healthy adults.  
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
Shoemaker, What's Normal: Temperature, Gender and Heartrate, Journal of Statistics Education, Vol. 4, No. 2
(1996)
}
\examples{
%% data(BodyTemp50)
}
\keyword{datasets}
