\name{BodyTemp50}
\alias{BodyTemp50}
\docType{data}
\title{
Body Temperatures
}
\description{
Sample of 50 body temperatures
}
%\usage{data(BodyTemp50)}
\format{
  A dataset with 50 observations on the following 3 variables.
  \tabular{rl}{
    \code{BodyTemp} \tab {Body temperatures in degrees F}\cr
    \code{Pulse} \tab {Pulse rates (beats per minute)}\cr
    \code{Gender} \tab {F=Female, M=Male}\cr
  }
}
\details{
Body temperatures and pulse rates for a sample of 50 healthy adults.  
}
\source{
Shoemaker, What's Normal: Temperature, Gender and Heartrate, Journal of Statistics Education, Vol. 4, No. 2
(1996)
}

\keyword{datasets}
