\name{BootAtlantaCorr}
\alias{BootAtlantaCorr}
\docType{data}
\title{
Bootstrap Correlations for Atlanta Commutes
}
\description{
Boostrap correlations between Time and Distance for 500 commuters in Atlanta
}
%\usage{data(BootAtlantaCorr)}
\format{
  A dataset with 1000 observations on the following variable.
  \tabular{rl}{
    \code{CorrTimeDist} \tab {Correlation between Time and Distance for a bootstrap sample of Atlanta commuters}\cr
  }
}
\details{
Correlations for bootstrap samples of Time vs. Distance for the data on Atlanta commuters in CommuteAtlanta.
}
\source{
Computer simulation
}

\keyword{datasets}
