\name{BootAtlantaCorr}
\alias{BootAtlantaCorr}
\docType{data}
\title{
Bootstrap Correlations for Atlanta Commutes
}
\description{
Boostrap correlations between Time and Distance for 500 commuters in Atlanta
}
\usage{data(BootAtlantaCorr)}
\format{
  A data frame with 1000 observations on the following variable.
  \describe{
    \item{\code{CorrTimeDist}}{Correlation between Time and Distance for a bootstrap sample of Atlanta commuters}
  }
}
\details{
Correlations for bootstrap samples of Time vs. Distance for the data on Atlanta commuters in CommuteAtlanta.
}
\source{
Computer simulation
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(BootAtlantaCorr)
}
\keyword{datasets}
