\name{CarbonDioxide}
\alias{CarbonDioxide}
\docType{data}
\title{Carbon Dioxide Levels}
\description{
Atmospheric carbon dioxide levels by year
}
\format{
  A data frame with 13 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Every five years from 1960 to 2020}
    \item{\code{CO2}}{Carbon dioxide level in parts per million}
  }
}
\details{
Carbon dioxide levels in the atmosphere over a 60 year span from 1960-2020. \cr
** Updated for 4e (earlier versions are now CarbonDioxide3e and CarbonDioxide2e) **

}
\source{
Dr. Pieter Tans, NOAA/ESRL. Values recorded at the Mauna Loa Observatory in Hawaii. \url{https://gml.noaa.gov/ccgg/trends/}
}
\keyword{datasets}
