\name{CarbonDioxide}
\alias{CarbonDioxide}
\docType{data}
\title{
CO2
}
\description{
Atmospheric carbon dioxide levels by year 
}
%\usage{data(Carbondioxide)}
\format{
  A dataset with 11 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year} \tab {Every five years from 1960-2010}\cr
    \code{C02} \tab {Carbon dioxide level in parts per million}\cr
  }
}
\details{
Carbon dioxide levels in the atmosphere over a 50 year span from 1960-2010.
}
\source{
Dr. Pieter Tans, NOAA/ESRL (www.esrl.noaa.gov/gmd/ccgg/trends/). Values recorded at the Mauna Loa Observatory
in Hawaii.
}

\keyword{datasets}
