\name{CarDepreciation}
\alias{CarDepreciation}
\docType{data}
\title{
Car Depreciation
}
\description{
Depreciation for 20 car models.
}
%\usage{data(CarDepreciation)}
\format{
  A dataset with 20 observations on the following 4 variables.
  \tabular{rl}{
    \code{Car} \tab {Name of the car model}\cr
    \code{New} \tab {Price of a new car}\cr
    \code{Used} \tab {Value after new car leaves the lot after purchase}\cr
    \code{Depreciation} \tab {Drop in value when a new car is driven away}\cr
  }
}
\details{
Twenty car models were selected at random from \emph{kellybluebook.com}.  Original price (in dollars) and value after the car has been driven 10 miles were recorded for each model.  The depreciation is the difference (New-Used).
}
\source{
New and used automobile costs determined using 2015 models selected from \emph{kellybluebook.com}.
}

\keyword{datasets}
