\name{CityTemps}
\alias{CityTemps}
\docType{data}
\title{City Temperatures}
\description{
Mean monthly temperature in Moscow, Melbourne, and San Francisco for 2023 and 2024
}
\format{
  A data frame with 24 observations on the following 5 variables.
  \describe{
    \item{\code{Year}}{2023 or 2024}
    \item{\code{Month}}{1=January through 12=December}
    \item{\code{Moscow}}{Monthly temperatures in Moscow (Russia)}
    \item{\code{Melbourne}}{Monthly temperatures in Melbourne (Australia)}
    \item{\code{SanFrancisco}}{Monthly temperatures in San Francisco (United States)}
  }
}
\details{
Mean monthly temperatures in degrees C for the years 2023 and 2024 in each of three cities. \cr
** Updated for 4e (earlier versions for 2017 and 2018 in CityTemps3e,  for 2014 and 2015 in CityTemps2e). **
}
\source{
Source:   \url{https://www.weatherandclimate.info/history/} add station codes 94866 (Melbourne), 72494 (San Francisco), 27612 (Moscow) to url.

}
\keyword{datasets}
