\name{CommuteAtlanta}
\alias{CommuteAtlanta}
\docType{data}
\title{
Commute Atlanta
}
\description{
Commute times and distance for a sample of 500 people in Atlanta
}
\usage{data(CommuteAtlanta)}
\format{
  A data frame with 500 observations on the following 5 variables.
  \describe{
    \item{\code{City}}{ \code{Atlanta}}
    \item{\code{Age}}{Age of the respondent (in years)}
    \item{\code{Distance}}{Commute distance (in miles)}
    \item{\code{Time}}{Commute time (in minutes)}
    \item{\code{Sex}}{  \code{F} or \code{M}}
  }
}
\details{
Data fro the US Census Bureua's American Housing Survey (AHS) which contains information about housing and living conditions for samples from certain metropolitan areas. These data were extracted respondents in the Atlanta metropolitan area. They include only cases where the respondent worked somewhere other than home. Values show the time (in minutes) and distance (in miles) that respondents typically traveled on their commute to work each day as well as age and sex.
}
\source{
Sample chosen using DataFerret at http://www.thedataweb.org/index.html.
}
\references{
%%
}
\examples{
%% data(CommuteAtlanta)
}
\keyword{datasets}
