\name{CricketChirps}
\alias{CricketChirps}
\docType{data}
\title{
Cricket Chirps
}
\description{
Cricket chirp rate and temperature
}
\usage{data(CricketChirps)}
\format{
  A data frame with 7 observations on the following 2 variables.
  \describe{
    \item{\code{Temperature}}{Air temperature in degrees F}
    \item{\code{Chirps}}{Cricket chirp rate (chirps per minute)}
  }
}
\details{
The data were collected by E.A. Bessey and C.A. Bessey who measured chirp rates for crickets
and temperatures during the summer of 1898.
}
\source{
From E.A Bessey and C.A Bessey, Further Notes on Thermometer Crickets, American Naturalist, (1898) 32, 263-264.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(CricketChirps)
}
\keyword{datasets}
