\name{DecemberFlights}
\alias{DecemberFlights}
\docType{data}
\title{December Flights}
\description{
Difference between actual and scheduled arrival for a sample of United and Delta flights in December 2014.
}
%\usage{data(DecemberFlights)}
\format{
  A dataset with 2000 observations on the following 2 variables.
  \tabular{rl}{
    \code{Airline} \tab {\code{Delta} or \code{United}}\cr
    \code{Difference} \tab {Difference (Actual - Scheduled arrival times)}\cr
  }
}
\details{
For a sample of 1000 December flights (in 2014) from each airline, we find the difference between actual and scheduled arrival times.  A negative value indicates the flight arrived early.
}
\source{
Downloaded from the Bureau of Transportation Statistics (\emph{https://www.bts.gov/}).  More specific URL is \emph{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236&DB_Short_Name=On-Time}.
}


\keyword{datasets}
