\name{Digits}
\alias{Digits}
\docType{data}
\title{
Digit counts
}
\description{
Digits from social security numbers and student selected random numbers
}
%\usage{data(Digits)}
\format{
  A dataset with 150 observations on the following 7 variables.
  \tabular{rl}{
    \code{Random} \tab {Four digit random numbers picked by a sample of students}\cr
    \code{RND1} \tab {First digit}\cr
    \code{RND2} \tab {Second digit}\cr
    \code{RND3} \tab {Third digit}\cr
    \code{RND4} \tab {Fourth digit}\cr
    \code{SSN8} \tab {Eighth digit of social security number}\cr
    \code{SSN9} \tab {Last digit of social security number}\cr
  }
}
\details{ A sample of students were asked to pick a random four digit number.  
The numbers are given in the dataset, along with separate columns for each of the four digits.  The data also show the last two digits of each student's social security number (SSN). 
}
\source{
In class-student surveys from several classes.
}

\keyword{datasets}
