\name{ElectionMargin}
\alias{ElectionMargin}
\docType{data}
\title{
Election Margin
}
\description{
Approval rating and election margin for recent presidential elections
}
\usage{data(ElectionMargin)}
\format{
  A data frame with 11 observations on the following 5 variables.
  \describe{
    \item{\code{Year}}{Certain election years from 1940-2004}
    \item{\code{Candidate}}{Incumbent US president}
    \item{\code{Approval}}{Presidential approval rating at time of election}
    \item{\code{Margin}}{margin of victory/defeat (as a percentage)}
    \item{\code{Result}}{Outcome of the election for the imbumbent: \code{Lost} or \code{Won}}
  }
}
\details{
Data include US Presidential elections since 1940 in which an incumbent was running for president. 
The approval rating for the sitting president is comapred to the margin of victory/defeat in the election. 
}
\source{
Silver, Nate, "Approval Ratings and Re-Election Odds", fivethirtyeight.com, posted January 28, 2011.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%%data(ElectionMargin)
}
\keyword{datasets}
