\docType{data}
\name{EmployedACS}
\alias{EmployedACS}
\title{Employed in American Community Survey}
\format{A dataset with 431 observations on the following 9 variables.
\itemize{
		\item{\code{Sex}} {\code{Female} or \code{Male} }
		\item{\code{Age}} {Age (years)}
		\item{\code{Married}} {\code{Not Married} or \code{Married}}
		\item{\code{Income}} {Wages and salary for the past 12 months (in \$1,000's)}
		\item{\code{HoursWk}} {Hours of work per week}
		\item{\code{Race}} {\code{asian}, \code{black}, \code{white}, or \code{other}}
		\item{\code{USCitizen}} {\code{Citizen} or \code{Noncitizen}}
		\item{\code{HealthInsurance}} {\code{Insured} or \code{Uninsured}}
		\item{code{Language}} {Native language: \code{English} or \code{Other}}
}}
\source{
  The full public dataset can be downloaded at
  \url{http://www.census.gov/acs/www/data_documentation/pums_data/}
  and the full list of variables is at
  \url{http://www.census.gov/acs/www/Downloads/data/documentation/pums/DataDict/PUMSDataDict10.pdf}.
}
\description{
  Employed individiuals from the American Community Survey
  (ACS) dataset
}
\details{
  This is a subset of the ACS dataset including only 431
  individuals who were employed.
}
\keyword{datasets}

