\name{ExerciseHours}
\alias{ExerciseHours}
\docType{data}
\title{
Exercise Hours
}
\description{
Amount of exercise per week for students (and other variables)
}
%\usage{data(ExerciseHours)}
\format{
  A dataset with 50 observations on the following 7 variables.
  \tabular{rl}{
    \code{Year} \tab {Year in school (1=First year,..., 4=Senior)}\cr
    \code{Gender} \tab { \code{F} or \code{M}}\cr
    \code{Hand} \tab {Left (\code{l}) or Right (\code{r}) handed?}\cr
    \code{Exercise} \tab {Hours of exercise per week}\cr
    \code{TV} \tab {Hours of TV viewing per week}\cr
    \code{Pulse} \tab {Resting pulse rate (beats per minute)}\cr
    \code{Pierces} \tab {Number of body piercings}\cr
  }
}
\details{
Data from an in-class survey of statistics students asking about amount of exercise,
TV viewing, handedness, gender, pulse rate, and number of body piercings. 
}
\source{
In-class student survey.
}

\keyword{datasets}
