\name{ExerciseHours}
\alias{ExerciseHours}
\docType{data}
\title{
Exercise Hours
}
\description{
Amount of exercise per week for students (and other variables)
}
\usage{data(ExerciseHours)}
\format{
  A data frame with 50 observations on the following 7 variables.
  \describe{
    \item{\code{Year}}{Year in school (1=First year,..., 4=Senior)}
    \item{\code{Gender}}{ \code{F} or \code{M}}
    \item{\code{Hand}}{Left (\code{l}) or Right (\code{r}) handed?}
    \item{\code{Exercise}}{Hours of exercise per week}
    \item{\code{TV}}{Hours of TV viewing per week}
    \item{\code{Pulse}}{Resting pulse rate (beats per minute)}
    \item{\code{Pierces}}{Number of body piercings}
  }
}
\details{
Data from an in-class survey of statistics students asking about amount of exercise,
TV viewing, handedness, gender, pulse rate, and number of body piercings. 
}
\source{
In-class student survey.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(ExerciseHours)
}
\keyword{datasets}
