\name{FishGills12}
\alias{FishGills12}
\docType{data}
\title{
FishGills12
}
\description{
An experiment to look at fish repiration rates in water withdiffernet levels of calcium.
}
\usage{data(FishGills12)}
\format{
  A data frame with 360 observations on the following 2 variables.
  \describe{
    \item{\code{Calcium}}{Amount of calcium in the water (mg/L)}
    \item{\code{GillRate}}{Respiration rate (beats per minute)}
  }
}
\details{
Fish were randomly assigned to twelve tanks with different levels (measured in mg/L) levels of calcium. 
Respiration rate was measured as number of gill beats per minute. 
}
\source{
Thanks to Pro. Brad Baldwin for supplying the data
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(FishGills12)
}
\keyword{datasets}
