\name{Flight179}
\alias{Flight179}
\docType{data}
\title{
Flight times
}
\description{
Flight times for Flight 179 (Boston-SF) and Flight 180 (SF-Boston)
}
\usage{data(Flight179)}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{Date}}{Date of the flight (5th, 15th and 25th of each monthin 2010}
    \item{\code{Flight179}}{Flying time (Boston-SF) in minutes}
    \item{\code{Flight180}}{Fllying time (SF-Boston) in minutes}
  }
}
\details{
United Airlines Flight 179 is a daily flight from Boston to San Francisco.  
Flight 180 goes in the other direction (SF to Boston). The data show the airborn flying times 
for each flight on the three dates each month (5th, 15th and 25th) in 2010. 
}
\source{
Data collected from the Bureau of Transportation Statistics website at
http://www.bts.gov/xml/ontimesummarystatistics/src/dstat/OntimeSummaryAirtime.xml
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(Flight179)
}
\keyword{datasets}
