\name{HockeyPenalties}
\alias{HockeyPenalties}
\docType{data}
\title{
Hockey Penalties
}
\description{
Penalty minutes (per game) for NHL teams in 2010-11
}
%\usage{data(HockeyPenalties)}
\format{
  A dataset with 30 observations on the following 2 variables.
  \tabular{rl}{
    \code{Team} \tab {Name of the team}\cr
    \code{PIMperG} \tab {Average penaly minutes per game}\cr
  }
}
\details{
Data give the average numeber of penalty minutes for each of the 30 National Hockey League (NHL) teams during the 2010-11 regualar season. 
}
\source{
Data obtained online at www.nhl.com
}

\keyword{datasets}
