\name{HollywoodMovies}
\alias{HollywoodMovies}
\docType{data}
\title{Hollywood Movies}
\description{
Data on movies released in Hollywood between 2007 and 2013
}
%\usage{data(HollywoodMovies)}
\format{
  A dataset with 970 observations on the following 16 variables.
  \tabular{rl}{
    \code{Movie} \tab {Title of movie} \cr
    \code{LeadStudio} \tab {Studio that released the movie} \cr
    \code{RottenTomatoes} \tab {Rotten Tomatoes rating (reviewers)} \cr
    \code{AudienceScore} \tab {Audience rating (via Rotten Tomatoes)} \cr
    \code{Story} \tab {General theme - one of 21 themes} \cr
    \code{Genre} \tab {One of 14 possible genres} \cr
    \code{TheatersOpenWeek} \tab {Number of screens for opening weekend} \cr
    \code{OpeningWeekend} \tab {Opening weekend gross (in $ millions)} \cr
    \code{BOAverageOpenWeek} \tab {Average opening week box office income (per theater)} \cr
    \code{DomesticGross} \tab {Gross income for domestic viewers (in $ millions)} \cr
    \code{ForeignGross} \tab {Gross income for foreign viewers (in $ millions)} \cr
    \code{WorldGross} \tab {Gross income for all viewers (in $ millions)} \cr
    \code{Budget} \tab {Production budget (in $ millions)} \cr
    \code{Profitability} \tab {WorldGross as a percentage of Budget} \cr
    \code{OpenProfit} \tab {Percentage of budget recovered on opening weekend} \cr
    \code{Year}  \tab {Year the movie was released} \cr
  }
}
\details{
Information from 970 movies released from Hollywood between 2007 and 2013. \cr
** This daatset is updated from an earlier version (HollywoodMovies2011) **
}
\source{
McCandless, D., "Most Profitable Hollywood Movies" from "Information is Beautiful" at \cr
\emph{http://www.informationisbeautiful,net.data/} and  \cr
\emph{http://bit.ly/hollywoodbudgets}.
}

\keyword{datasets}
