\name{HollywoodMovies2011}
\alias{HollywoodMovies2011}
\docType{data}
\title{Hollywood Movies in 2011}
\description{
Information for 136 movies released from Hollywood in 2011. 
}
%\usage{data(HollywoodMovies2011)}
\format{
  A dataset with 136 observations on the following 14 variables.
  \tabular{rl}{
    \code{Movie} \tab {Title of movie}\cr
    \code{LeadStudio} \tab {Studio that released the movie}\cr
    \code{RottenTomatoes} \tab {Rotten Tomatoes rating (reviewers)}\cr
    \code{AudienceScore} \tab {Audience rating (via Rotten Tomatoes)}\cr
    \code{Story} \tab {General theme - one of 21 themes}\cr
    \code{Genre} \tab {Type of movie: \code{Action}, \code{Adventure}, \code{Animation}, \code{Comedy}, \code{Drama,} \code{Fantasy},}\cr 
    \tab {\code{Horror}, \code{Romance}, or \code{Thriller}}\cr
    \code{TheatersOpenWeek} \tab {Number of screens for opening weekend}\cr
    \code{BOAverageOpenWeek} \tab {Average box office income per theater - opening weekend}\cr
    \code{DomesticGross} \tab {Gross income for domestic viewers (in millions)}\cr
    \code{ForeignGross} \tab {Gross income for foreign viewers (in millions)}\cr
    \code{WorldGross} \tab {Gross income for all viewers (in millions)}\cr
    \code{Budget} \tab {Productoin budge (in millions)}\cr
    \code{Profitability} \tab {WorldGross/Budget}\cr
    \code{OpeningWeekend} \tab {Opening weekend gross (in millions)}\cr
  }
}
\details{
There were 136 movies that came out of Hollywood in 2011. 
 The data were compiled late in 2011 so they reflect results as of December 2011.
}
\source{
McCandless, D., "Most Profitable Hollywood Movies,", from "Information is Beautiful", davidmccandless.com, accessed January 2012. 
}

\keyword{datasets}
