\name{HomesForSale}
\alias{HomesForSale}
\docType{data}
\title{Homes For Sale (2025)}
\description{
Data on homes sold in four states in 2025
}
\format{
  A data frame with 120 observations on the following 6 variables.
  \describe{
    \item{\code{State}}{Location of the home (\code{NY}, \code{IA}, \code{CA}, or \code{FL})}
    \item{\code{Price}}{Asking price (in $1,000's)}
    \item{\code{Size}}{Area of all rooms (in sq. ft.)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
    \item{\code{Zip}}{Zip code for the home}
  }
}
\details{
Data for samples of 30 homes recently sold in each state, selected from random zip codes at \emph{zillow.com}. \cr
** Updated for 4e (earlier versions are HomesForSale3e (2019) and HomesForSale2e (2010)). **
}
\source{
Data collected from \emph{https://www.zillow.com/} in March 2025.
}
\keyword{datasets}
