\name{HomesForSale}
\alias{HomesForSale}
\docType{data}
\title{
Home for Sale
}
\description{
Data on homes for sale in four states
}
\usage{data(HomesForSale)}
\format{
  A data frame with 120 observations on the following 5 variables.
  \describe{
    \item{\code{State}}{Location of the home: \code{CA} \code{NJ} \code{NY} \code{PA}}
    \item{\code{Price}}{Asking price (in $1,000's)}
    \item{\code{Size}}{Area of all rooms (in 1,000's sq. ft.)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
  }
}
\details{
Data for samples of homes for sale in each state, selected from zillow.com. 
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
Data collected from www.zillow.com in 2010. 
}
\examples{
%%data(HomesForSale)
}
\keyword{datasets}
