\name{HomesForSaleCA}
\alias{HomesForSaleCA}
\docType{data}
\title{Homes For Sale in California (2019)}
\description{
Data for a sample of homes offered for sale in California
}
\format{
  A data frame with 30 observations on the following 5 variables.
  \describe{
    \item{\code{State}}{Location of the home (\code{CA})}
    \item{\code{Price}}{Asking price (in $1,000's)}
    \item{\code{Size}}{Area of all rooms (in 1,000's sq. ft.)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
  }
}
\details{
Data fora sample of homes for sale in California, selected from \emph{zillow.com}. This is a subset of the HomesForSale dataset. \cr
** Updated for 3e (earlier version from 2010 is in HomesForSaleCA2e). **
}
\source{
Data collected from \url{https://www.zillow.com/} in 2019.
}
\keyword{datasets}
