\name{HomesForSaleCA}
\alias{HomesForSaleCA}
\docType{data}
\title{
Home for Sale in California
}
\description{
Data for a sample of homes offered for sale in California
}
%\usage{data(HomesForSaleCA)}
\format{
  A dataset with 30 observations on the following 5 variables.
  \tabular{rl}{
    \code{State} \tab {Location of the home: \code{CA} } \cr
    \code{Price} \tab {Asking price (in $1,000's)} \cr
    \code{Size} \tab {Area of all rooms (in 1,000's sq. ft.)}\cr
    \code{Beds} \tab {Number of bedrooms}\cr
    \code{Baths} \tab {Number of bathrooms}\cr
  }
}
\details{
Data for samples of homes for sale in California, selected from \emph{zillow.com}.
}
\source{
Data collected from \emph{www.zillow.com} in 2010.
}

\keyword{datasets}
