\name{HomesForSaleCanton}
\alias{HomesForSaleCanton}
\docType{data}
\title{
Homes for sale in Canton, NY
}
\description{
Prices of homes for sale in Canton, NY
}
\usage{data(HomesForSaleCanton)}
\format{
  A data frame with 10 observations on the following variable.
  \describe{
    \item{\code{Price}}{Asking price for the home (in $1,000's)}
  }
}
\details{
Data for samples of homes for sale in Canton, NY, selected from zillow.com. 
}
\source{
Data collected from www.zillow.com in 2010. 
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%data(HomesForSaleCanton)
}
\keyword{datasets}
