\name{HomesForSaleNY}
\alias{HomesForSaleNY}
\docType{data}
\title{
Home for Sale in New York
}
\description{
Data for a sample of homes offered for sale in New York State
}
%\usage{data(HomesForSaleNY)}
\format{
  A dataset with 30 observations on the following 5 variables.
  \tabular{rl}{
    \code{State} \tab {Location of the home: \code{NY}}\cr
    \code{Price} \tab {Asking price (in $1,000's)}\cr
    \code{Size} \tab {Area of all rooms (in 1,000's sq. ft.)}\cr
    \code{Beds} \tab {Number of bedrooms}\cr
    \code{Baths} \tab {Number of bathrooms}\cr
  }
}
\details{
This is a subset of the HomesForSale data with just information from homes in New York State  (NY). Data were collected offerings listed on an online site.  
}
\source{
Data collected from www.zillow.com in 2010.
}

\keyword{datasets}
