\name{HotDogs}
\alias{HotDogs}
\docType{data}
\title{Hot Dog Eating Contest}
\description{
Winning number of hot dogs consumed in an eating contest
}
%\usage{data(HotDogs)}
\format{
  A dataset with 10 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year} \tab {Year of the contest: 2002-2011}\cr
    \code{HotDogs} \tab {Winning number of hot dogs consumed}\cr
  }
}
\details{
Every Fourth of July, Nathan's Famous in New York City holds a hot dog eating contest, in which
contestants try to eat as many hot dogs (with buns) as possible in ten minutes. 
The winning number of hot dogs are given for each year from 2002-2011. 

}
\source{
Nathan's Famous webste at http://nathansfamous.com/PageFetch/getpage.php?pgid=38
}

\keyword{datasets}
