\name{InkjetPrinters}
\alias{InkjetPrinters}
\docType{data}
\title{Inkjet Printers}
\description{
Data from online reviews of inkjet printers
}
%\usage{data(InkjetPrinters)}
\format{
  A dataset with 20 observations on the following 6 variables.
  \tabular{rl}{
    \code{Model} \tab {Model name of printer}\cr
    \code{PPM} \tab {Printing rate (pages per minute) for a benchmark set of print jobs}\cr
    \code{PhotoTime} \tab {Time (in seconds) to print 4x6 color photos}\cr
    \code{Price} \tab {Typical retail price (in dollars)}\cr
    \code{CostBW} \tab {Cost per page (in cents) for printing in black & white}\cr
    \code{CostColor} \tab {Cost per page (in cents) for printing in color}\cr
  }
}
\details{
Information from reviews of inkjet printers at PCMag.com in August 2011.
}
\source{
Inkjet printer reviews found at http://www.pcmag.com/reviews/printers, August 2011.
}

\keyword{datasets}
