\name{LifeExpectancyVehicles}
\alias{LifeExpectancyVehicles}
\docType{data}
\title{Life Expectancy and Vehicle Registrations}
\description{
Yearly U.S. life expectancy and paassenger car registrations
}
%\usage{data(LifeExpectancyVehicles)}
\format{
  A dataset with 40 observations on the following 3 variables.
  \tabular{rl}{
    \code{Year} \tab {1970-2009}\cr
    \code{LifeExpectancy} \tab {Average life expectancy at birth in the United States (in years)}\cr
    \code{Vehicles} \tab {Passenger vehicles registered in the United States (in millions)}\cr
  }
}
\details{
Yearly data on the average life expectancy for babies born and 
millions of cars registered in the United States over 40 years from 1970 to 2009
}
\source{
Vehicle registrations from US Census Bureau, \cr
 http://www.census.gov/compendia/statab/cats/transportation.html.
Life expectancies from the Centers for Disease Control and Prevention, National Center for Health Statistics, Health Data Interactive,
www.cdc.gov/nchs/hdi.htm, accessed April 7, 2012.
}

\keyword{datasets}
