\name{Lock5Data-package}
\alias{Lock5Data-package}
\alias{Lock5Data}
\docType{package}
\title{
Lock5 Datasets
}
\description{
Datasets for Statistics: Unlocking the Power of Data \cr
by Lock, Lock, lock, Lock and Lock
}
\details{
\tabular{ll}{
Package: \tab Lock5Data\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-18\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Robin Lock

Maintainer: Who to complain to <rlock@stlawu.edu>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
