\name{MalevolentUniformsNFL}
\alias{MalevolentUniformsNFL}
\docType{data}
\title{
Malevolent Uniforms NFL
}
\description{
Perceived malevolence of uniforms and penalies for National Football League (NFL) teams
}
%\usage{data(MalevolentUniformsNFL)}
\format{
  A dataset with 28 observations on the following 3 variables.
  \tabular{rl}{
    \code{NFLTeam} \tab {Team name}\cr
    \code{NFL_Malevolence} \tab {Score reflecting the "malevolence" of a team's uniform}\cr
    \code{ZPenYds} \tab {Z-score for penalty yards}\cr
  }
}
\details{
Participants with no knowledge of the teams rated the jerseys 
on characteristics such as timid/aggressive, nice/mean and good/bad. The averages of these responses
produced a "malevolence" index with higher scores signifying impressions of more malevolent uniforms. To
measure aggressiveness, the authors used the amount of penalty yards converted to z-scores and averaged for each team over the seasons from 1970-1986.
}
\source{
Frank and Gilovich, "The Dark Side of Self- and Social Perception: Black Uniforms and Aggression in Professional Sports",
Journal of Personality and Social Psychology, Vol. 54, No. 1, 1988, p. 74-85.
}

\keyword{datasets}
