\name{MammalLongevity}
\alias{MammalLongevity}
\docType{data}
\title{
Mammal Longevity
}
\description{
Longevity and gestation period for mammals
}
%\usage{data(MammalLongevity)}
\format{
  A dataset with 40 observations on the following 3 variables.
  \tabular{rl}{
    \code{Animal} \tab {Species of mammal}\cr
    \code{Gestation} \tab {Time from fertilization until birth (in days)}\cr
    \code{Longevity} \tab {Average lifespan (in years)}\cr
  }
}
\details{
Dataset with average lifespan (in years) and typical gestation period (in days) for 40 different species of mammals. 
}
\source{
2010 World Almanac, pg. 292.
}

\keyword{datasets}
