\name{ManhattanApartments}
\alias{ManhattanApartments}
\docType{data}
\title{Manhattan Apartment Prices}
\description{
Monthly rent for one-bedroom apartments in Manhattan, NY
}
%\usage{data(ManhattanApartments)}
\format{
  A dataset with 20 observations on the following variable.
  \tabular{rl}{
    \code{Rent} \tab {Montly rent in dollars}\cr
  }
}
\details{
Monthly rents for a sample of 20 one-bedroom apartments in Manhattan, NY that were advertised on Craig's List in July, 2011. 
}
\source{
Apartments advertised on Craig's List at newyork.craigslist.org, July 5, 2011.
}

\keyword{datasets}
