\name{MarriageAges}
\alias{MarriageAges}
\docType{data}
\title{
Marriage Ages
}
\description{
Ages for husbands and wives from marriage licenses
}
%\usage{data(MarriageAges)}
\format{
  A dataset with 100 observations on the following 2 variables.
  \tabular{rl}{
    \code{Husband} \tab {Age of husband at marriage}\cr
    \code{Wife} \tab {Age of wife at marriage}\cr
  }
}
\details{
Data from a sample of 100 marriage licences in St. Lawrence County, NY gives the ages of husbands and wives for newly married couples.
}
\source{
Thanks to Linda Casserly, St. Lawrence County Clerk's Office
}

\keyword{datasets}
