\name{MiamiHeat}
\alias{MiamiHeat}
\docType{data}
\title{
Miami Heat basketball
}
\description{
Game log data for the Miami Heat basketball team in 2010-11
}
%\usage{data(MiamiHeat)}
\format{
  A dataset with 82 observations on the following 33 variables.
  \tabular{rl}{
    \code{Game} \tab {ID number for each game}\cr
    \code{Date} \tab {Data the game was played}\cr
    \code{Location} \tab {\code{Away}  or \code{Home}}\cr
    \code{Opp} \tab {Opponent tream}\cr
    \code{Win} \tab {Game result: \code{L} or \code{W}}\cr
    \code{FG} \tab {Field goals made}\cr
    \code{FGA} \tab {Field goals attempted}\cr
    \code{FG3} \tab {Three-point field goals made}\cr
    \code{FG3A} \tab {Three-point field goals attempted}\cr
    \code{FT} \tab {Free throws made}\cr
    \code{FTA} \tab {Free throws attempted}\cr
    \code{Rebounds} \tab {Total rebounds}\cr
    \code{OffReb} \tab {Offensive rebounds}\cr
    \code{Assists} \tab {Number of assists}\cr
    \code{Steals} \tab {Number of steals}\cr
    \code{Blocks} \tab {Number of shots blocked}\cr
    \code{Trunovers} \tab {Number of turnovers}\cr
    \code{Fouls} \tab {Number of fouls}\cr
    \code{Points} \tab {Number of points scored}\cr
    \code{OppFG} \tab {Opponet's field goals made}\cr
    \code{OppFGA} \tab {Opponent's Field goals attempted}\cr
    \code{OppFG3} \tab {Opponent's Three-point field goals made}\cr
    \code{OppFG3A} \tab {Opponent's Three-point field goals attempted}\cr
    \code{OppFT} \tab {Opponent's Free throws made}\cr
    \code{OppFTA} \tab {Opponent's Free throws attempted}\cr
    \code{OppOffReb} \tab {Opponent's Offensive rebounds}\cr
    \code{OppRebounds} \tab {Opponent's Total rebounds}\cr
    \code{OppAssists} \tab {Opponent's assists}\cr
    \code{OppSteals} \tab {Opponent's steals}\cr
    \code{OppBlocks} \tab {Opponent's shots blocked}\cr
    \code{OppTurnovers} \tab {Opponent's turnovers}\cr
    \code{OppFouls} \tab {Opponent's fouls}\cr
    \code{OppPoints} \tab {Opponent's points scored}\cr
  }
}
\details{
Information from online boxscores for all 82 regular season games payed by the Miami Heat basketball team during the 2010-11 regular season. 
}
\source{
Data for the 2010-11 Miami games downloaded from  \cr
http://www.basketball-reference.com/teams/MIA/2011/gamelog/
}

\keyword{datasets}
