\name{MindsetMatters}
\alias{MindsetMatters}
\docType{data}
\title{
Mindset Matters
}
\description{
Data from a study of perceived exercise with maids
}
%\usage{data(MindsetMatters)}
\format{
  A dataset with 75 observations on the following 14 variables.
  \tabular{rl}{
    \code{Cond} \tab {Treatment condition:  \code{0}=uninformed or \code{1}=informed}\cr
    \code{Age} \tab {Age (in years)}\cr
    \code{Wt} \tab {Original weight (in pounds)}\cr
    \code{Wt2} \tab {Weight after 4 weeks (in pounds}\cr
    \code{BMI} \tab {Original body mass index}\cr
    \code{BMI2} \tab {Body mass index after 4 weeks}\cr
    \code{Fat} \tab {Original body fat percentage}\cr
    \code{Fat2} \tab {Body fat percentage after 4 weeks}\cr
    \code{WHR} \tab {Original waist to hip ratio}\cr
    \code{WHR2} \tab {Waist to hip ratio}\cr
    \code{Syst} \tab {Original systolic blood pressure}\cr
    \code{Syst2} \tab {Systolic blood pressure after 4 weeks}\cr
    \code{Diast} \tab {Original diastolic blood pressure}\cr
    \code{Diast2} \tab {Diastolic blood pressure after 4 weeks}\cr
  }
}
\details{
In 2007 a Harvard psychologist recruited 75 female maids working in different hotels to participate in a study. She informed 41 maids (randomly chosen) that the work they do satisfies the Surgeon General's recommendations for an active lifestyle (which is true), giving them examples for how why their work is good exercise. The other 34 maids were told nothing (uninformed). Various chacteristics (weight, body mass index, ...) were recorded for each subject at the start of the experiment and again four weeks later. Maids with missing values for weight change have been removed.
}
\source{
Crum, A.J. and Langer, E.J. (2007). Mind-Set Matters: Exercise and the Placebo Effect, Psychological Science, 18:165-171. Thanks to the authors for supplying the data. 
}

\keyword{datasets}
