\name{NBAStandings}
\alias{NBAStandings}
\docType{data}
\title{
NBA 2010-11 Regular Season Standings
}
\description{
Won-Loss record and statistics for NBA Teams
}
%\usage{data(HomesForSaleCA)}
\format{
  A dataset with 30 observations on the following 6 variables.
  \tabular{rl}{
    \code{Team} \tab {Team name}\cr
    \code{Wins} \tab {Number of wins in an 82 game regular season}\cr
    \code{Losses} \tab {Number of losses}\cr
    \code{WinPct} \tab {Proportion of games won}\cr
    \code{PtsFor} \tab {Average points scored per game}\cr
    \code{PtsAgainst} \tab {Average points allowed per game}\cr
  }
}
\details{
Won-Loss record and regular season statistics for 30 teams in the National Basketball Association
for the 2010-2011 season
}
\source{
Data downloaded from http://www.basketball-reference.com/leagues/NBA_2011_games.html
}

\keyword{datasets}
