\name{NutritionStudy}
\alias{NutritionStudy}
\docType{data}
\title{
Nutrition Study
}
\description{
Variables related to nutrition and health for 315 individuals
}
%\usage{data(NutritionStudy)}
\format{
  A dataset with 315 observations on the following 17 variables.
  \tabular{rl}{
    \code{ID} \tab {ID number for each subject in this sample}\cr
    \code{Age} \tab {Subject's age (in years)}\cr
    \code{Smoke} \tab {a factor with levels \code{No} \code{Yes}}\cr
    \code{Quetelet} \tab {Weight/(Height^2)}\cr
    \code{Vitamin} \tab {Vitamin use: \code{1}=Regulary, \code{2}=Occasionally, or \code{3}=No}\cr
    \code{Calories} \tab {Number of calories consumed per day}\cr
    \code{Fat} \tab {Grams of fat consumed per day}\cr
    \code{Fiber} \tab {Grams of fiber consumed per day}\cr
    \code{Alcohol} \tab {Number of alcoholic drinks consumed per week}\cr
    \code{Cholesterol} \tab {Cholesterol consumed (mg per day)}\cr
    \code{BetaDiet} \tab {Dietary beta-carotene consumed (mcg per day)}\cr
    \code{RetinolDiet} \tab {Dietary retinol consumed (mcg per day)}\cr
    \code{BetaPlasma} \tab {Plasma beta-carotene (ng/ml)}\cr
    \code{RetinolPlasma} \tab {Plasma retinol (ng/ml)}\cr
    \code{Gender} \tab {Coded as \code{Female} or \code{Male}}\cr
    \code{VitaminUse} \tab {Coded as \code{No} \code{Occasional} \code{Regular}}\cr
    \code{PriorSmoke} \tab {Smoking status: \code{1}=Never, \code{2}=Former, or \code{3}=Current}\cr
  }
}
\details{
Data from a cross-sectional study to investigate the relationship between personal characteristics and dietary factors, and plasma concentrations of retinol, beta-carotene and other carotenoids. Study subjects were patients who had an elective surgical procedure during a three-year period to biopsy or remove a lesion of the lung, colon, breast, skin, ovary or uterus that was found to be non-cancerous. 
}
\source{
 Data downloaded from http://lib.stat.cmu.edu/datasets/Plasma_Retinol
}
\references{
Nierenberg, Stukel, Baron, Dain, and Greenberg, "Determinants of plasma levels of beta-carotene and retinol", 
American Journal of Epidemiology (1989). 
}

\keyword{datasets}
