\name{OlympicMarathon}
\alias{OlympicMarathon}
\docType{data}
\title{
2008 Olympic Men's Marathon
}
\description{
Times for all finishers in the men's marathon at the 2008 Olympics
}
\usage{data(OlympicMarathon)}
\format{
  A data frame with 76 observations on the following 5 variables.
  \describe{
    \item{\code{Rank}}{Order of finish}
    \item{\code{Athlete}}{Name of marathoner}
    \item{\code{Nationality}}{Country of marathoner}
    \item{\code{Time}}{Time as H:MM:SS}
    \item{\code{Minutes}}{Time in minutes}
  }
}
\details{
Results for all finishers in the 2008 Men's Olympic marathon in Beijing, China. 
}
\source{
http://2008olympics.runnersworld.com/2008/08/mens-marathon-results.html
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(OlympicMarathon)
}
\keyword{datasets}
