\name{OttawaSenators}
\alias{OttawaSenators}
\docType{data}
\title{
Ottaw Senators hockey team
}
\description{
Data for 24 players on the 2009-10 Ottawa Senators
}
%\usage{data(OttawaSenators)}
\format{
  A dataset with 24 observations on the following 2 variables.
  \tabular{rl}{
    \code{Points} \tab {Number of points (goals + assists) scored}\cr
    \code{PenMins} \tab {Number of penalty minutes}\cr
  }
}
\details{
Points scored and penaly minutes for 24 players (excluding goalies) playing ice hockey for the Ottawa Senators during the 2009-10 NHL regular season. 
}
\source{
Data obtained from http://senators.nhl.com/club/stats.htm.
}

\keyword{datasets}
