\name{QuizPulse10}
\alias{QuizPulse10}
\docType{data}
\title{
Quiz vs Lecture Pulse Rates
}
\description{
Paired data with pulse rates in a lecture and during a quiz for 10 students
}
%\usage{data(QuizPulse10)}
\format{
  A dataset with 10 observations on the following 3 variables.
  \tabular{rl}{
    \code{Student} \tab {ID number for the student}\cr
    \code{Quiz} \tab {Pulse rate (beats per minute) during a quiz}\cr
    \code{Lecture} \tab {Pulse rate (beats per minute) during a lecture}\cr
  }
}
\details{
Ten students in an introductory statistics class measured their pulse rate in two settings: in the middle of a regular class lecture and again while taking a quiz. 
}
\source{
In-class data collection
}

\keyword{datasets}
