\name{RandomP50N200}
\alias{RandomP50N200}
\docType{data}
\title{
Simulated proportions
}
\description{
Counts and proportions for 5000 simulated samples with n=200 and p=0.50
}
%\usage{data(RandomP50N200)}
\format{
  A dataset with 5000 observations on the following 2 variables.
  \tabular{rl}{
    \code{Count} \tab {Number of simulated "yes" responses in 200 trials}\cr
    \code{Phat} \tab {Sample proportion (Count/200)}\cr
  }
}
\details{
Results from 5000 simulations of samples of size n=200 from a population with proportoin of "yes" responses at p=0.50. 
}
\source{
Computer simulation
}

\keyword{datasets}
