\name{RandomP50N200}
\alias{RandomP50N200}
\docType{data}
\title{
Simulated proportions
}
\description{
Counts and proportions for 5000 simulated samples with n=200 and p=0.50
}
\usage{data(RandomP50N200)}
\format{
  A data frame with 5000 observations on the following 2 variables.
  \describe{
    \item{\code{Count}}{Number of simulated "yes" responses in 200 trials}
    \item{\code{Phat}}{Sample proportion (Count/200)}
  }
}
\details{
Results from 5000 simulations of samples of size n=200 from a population with proportoin of "yes" responses at p=0.50. 
}
\source{
Computer simulation
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%data(RandomP50N200)
}
\keyword{datasets}
