\name{RestaurantTips}
\alias{RestaurantTips}
\docType{data}
\title{
Restaurant Tips
}
\description{
Tip data from the First Crush Bistro
}
%\usage{data(RestaurantTips)}
\format{
  A dataset with 157 observations on the following 7 variables.
  \tabular{rl}{
    \code{Bill} \tab {Size of the bill (in dollars)}\cr
    \code{Tip} \tab {Size of the tip (in dollars)}\cr
    \code{Credit} \tab {Paid with a credit card?  \code{n} or \code{y}}\cr
    \code{Guests} \tab {Number of peole in the group}\cr
    \code{Day} \tab {Day of the week: \code{m}=Monday, \code{t}=Tuesday, \code{w}=Wednesday, \code{th}=Thursday, or \code{f}=Friday }\cr
    \code{Server} \tab {Code for waiter/waitress: \code{A},  \code{B}, or \code{C}}\cr
    \code{PctTip} \tab {Tip as a percentage of the bill}\cr
  }
}
\details{
The owner of a bistro called First Crush in Potsdam, NY was interested in studying the tipping patterns of
his customers. 
He collected restaurant bills over a two week period that he believes provide a good sample of his customers. The data recorded from 157 bills include the amount of the bill, 
size of the tip, percentage tip, number of customers in the group, whether or not a credit card was used, day of the week, and a coded identity of the server.
}
\source{
Thanks to Tom DeRosa for providing the tipping data.
}

\keyword{datasets}
