\name{RetailSales}
\alias{RetailSales}
\docType{data}
\title{Retail Sales (2009-2024)}
\description{
Monthly U.S. Retail Sales from 2009 to 2024
}
\format{
  A data frame with 192 observations on the following 3 variables.
  \describe{
    \item{\code{Month}}{Month (\code{Jan} through \code{Dec})}
    \item{\code{Year}}{Years from 2009 to 2024}
    \item{\code{Sales}}{Monthly U.S. retail sales (in billions of dollars)}
  }
}
\details{
Data show the monthly retail sales (in billions) for the U.S. economy in each month from January 2009 through December 2024. \cr
** Updated for 4e (earlier versions are RetailSales2e and RetailSales2011).  **
}
\source{
Data downloaded from \url{https://www.census.gov/retail/} (February 2025).
}
\keyword{datasets}
