\name{RetailSales}
\alias{RetailSales}
\docType{data}
\title{Retail Sales (2009-2019)}
\description{
Monthly U.S. Retail Sales from 2009 to 2019
}
\format{
  A data frame with 129 observations on the following 3 variables.
  \describe{
    \item{\code{Month}}{Month (\code{Jan} through \code{Dec})}
    \item{\code{Year}}{Years from 2009 to 2019}
    \item{\code{Sales}}{Monthly U.S. retail sales (in billions of dollars)}
  }
}
\details{
Data show the monthly retail sales (in billions) for the U.S. economy in each month from January 2009 through September 2019. \cr
** Updated for 3e (earlier versions are RetailSales2e and RetailSales1e).  **
}
\source{
Data downloaded from \url{https://www.census.gov/retail/}.
}
\keyword{datasets}
