\name{RetailSales}
\alias{RetailSales}
\docType{data}
\title{
Rtail Sales
}
\description{
Monthly U.S. Retail Sales (in billions)
}
%\usage{data(RetailSales)}
\format{
  A dataset with 144 observations on the following 3 variables.
  \tabular{rl}{
    \code{Month} \tab {Month of the year}\cr
    \code{Year} \tab {Years from 2002 to 2011 }\cr
    \code{Sales} \tab {U.S. retail sales (in billions of dollars)}\cr
  }
}
\details{
Data show the monthly retail sales (in billions) for the U.S. economy in each month from 2002 through 2011. 
}
\source{
http://www.census.gov/retail/
}

\keyword{datasets}
