\name{RetailSales}
\alias{RetailSales}
\docType{data}
\title{
Rtail Sales
}
\description{
Monthly U.S. Retail Sales (in billions)
}
\usage{data(RetailSales)}
\format{
  A data frame with 144 observations on the following 3 variables.
  \describe{
    \item{\code{Month}}{Month of the year}
    \item{\code{Year}}{Years from 2002 to 2011 }
    \item{\code{Sales}}{U.S. retail sales (in billions of dollars)}
  }
}
\details{
Data show the monthly retail sales (in billions) for the U.S. economy in each month from 2002 through 2011. 
}
\source{
http://www.census.gov/retail/
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(RetailSales)
}
\keyword{datasets}
