\name{RockandRoll}
\alias{RockandRoll}
\docType{data}
\title{
Rock & Roll Hall of fame
}
\description{
Groups and Individuals in the Rock and Roll hall of Fame
}
%\usage{data(RockandRoll)}
\format{
  A dataset with 273 observations on the following 4 variables.
  \tabular{rl}{
    \code{Inductee} \tab {Name of he group or individual}\cr
    \code{FemaleMembers} \tab {\code{Yes} if individual or membr of the group is female, otherwise \code{No}}\cr
    \code{Category} \tab {Type of indiviudal or group: \code{Performer}, \code{Non-performer}, \code{Early Influence},}\cr
    \tab {\code{Lifetime Achievement}, \code{Sideman}}\cr
    \code{People} \tab {Number of people in the group}\cr
  }
}
\details{
All inductees of the Rock & Roll Hall of Fame as of 2012. 
}
\source{
Rock & Roll Hall of Fame website, http://rockhall.com/inductees/alphabetical/
}

\keyword{datasets}
