\name{RockandRoll2012}
\alias{RockandRoll2012}
\docType{data}
\title{
Rock & Roll Hall of Fame (2012)
}
\description{
Groups and Individuals in the Rock and Roll Hall of Fame (2012)
}
%\usage{data(RockandRoll2012)}
\format{
  A dataset with 273 observations on the following 4 variables.
  \tabular{rl}{
    \code{Inductee} \tab {Name of the group or individual} \cr
    \code{FemaleMembers} \tab {\code{Yes} if individual or member of the group is female, otherwise \code{No}} \cr
    \code{Category} \tab {Type of individual or group: \code{Performer}, \code{Non-performer}, \code{Early Influence},}\cr
    { } \tab {\code{Lifetime Achievement}, \code{Sideman}} \cr
    \code{People} \tab {Number of people in the group} \cr
  }
}
\details{
All inductees of the Rock & Roll Hall of Fame as of 2012. \cr
** From 1e - dataset has been updated for 2e and 3e **
}
\source{
Rock & Roll Hall of Fame website, \emph{http://rockhall.com/inductees/alphabetical/}
}

\keyword{datasets}
