\name{SampCountries}
\alias{SampCountries}
\docType{data}
\title{Sample of Countries}
\description{
Data on a sample of fifty countries of the world (2018)
}
\format{
  A data frame with 50 observations on the following 25 variables.
  \describe{
    \item{\code{Country}}{Country name}
    \item{\code{LandArea}}{Size in 1000 sq. km.}
    \item{\code{Population}}{Population in millions}
    \item{\code{Density}}{Number of people per square kilometer}
    \item{\code{GDP}}{Gross Domestic Product (in $US) per capita}
    \item{\code{Rural}}{Percentage of population living in rural areas}
    \item{\code{CO2}}{CO2 emissions (metric tons per capita)}
    \item{\code{PumpPrice}}{Price for a liter of gasoline ($US)}
    \item{\code{Military}}{Percentage of government expenditures directed toward the military}
    \item{\code{Health}}{Percentage of government expenditures directed towards healthcare}
    \item{\code{ArmedForces}}{Number of active duty military personnel (in 1,000's)}
    \item{\code{Internet}}{Percentage of the population with access to the internet}
    \item{\code{Cell}}{Cell phone subscriptions (per 100 people)}
    \item{\code{HIV}}{Percentage of the population with HIV}
    \item{\code{Hunger}}{Percent of the population considered undernourished}
    \item{\code{Diabetes}}{Percent of the population diagnosed with diabetes}
    \item{\code{BirthRate}}{Births per 1000 people}
    \item{\code{DeathRate}}{Deaths per 1000 people}
    \item{\code{ElderlyPop}}{Percentage of the population at least 65 years old}
    \item{\code{LifeExpectancy}}{Average life expectancy (years)}
    \item{\code{FemaleLabor}}{Percent of females 15 - 64 in the labor force}
    \item{\code{Unemployment}}{Percent of labor force unemployed}
    \item{\code{EnergyUse}}{Kilotons of oil equivalent}
    \item{\code{Electricity}}{Electric power consumption (kWh per capita)}
    \item{\code{Developed}}{Categories for kilowatt hours per capita, 1= under 2500, 2=2500 to 5000, 3=over 5000}
  }
}
\details{
Data from AllCountries for a random sample of 50 countries. Data for 2016-2018 to avoid many missing values in more recent years. \cr
** Updated for 3e (earlier versions are now SampCountries2e and SampCountries1e). **

}
\source{
Data collected from the World Bank website, http://www.worldbank.org.
}
\keyword{datasets}
