\name{SampCountries}
\alias{SampCountries}
\docType{data}
\title{
AllCountries
}
\description{
Data on a sample of countries of the world
}
%\usage{data(SampCountries)}
\format{
  A dataset with 50 observations on the following 13 variables.
  \tabular{rl}{
    \code{Country} \tab {Name of the country}\cr
    \code{LandArea} \tab {Size in sq. kilometers}\cr
    \code{Population} \tab {Population in millions}\cr
    \code{Energy} \tab {Energy usage (kilotons of oil)}\cr
    \code{Rural} \tab {Percentage of population living in rural areas}\cr
    \code{Military} \tab {Percentage of government expenditures directed toward the military}\cr
    \code{Health} \tab {Percentage of government expenditures directed towards healthcare}\cr
    \code{HIV} \tab {Percentage of the population with HIV}\cr
    \code{Internet} \tab {Percentage of the population with access to the internet}\cr
    \code{Developed} \tab {Categories for kilowat hours per capita: \code{1}= under 2500, \code{2}=2500 to 5000, \code{3}=over 5000}\cr
    \code{BirthRate} \tab {Births per 1000 people}\cr
    \code{ElderlyPop} \tab {Percentage of the population at least 65 year old}\cr
    \code{LifeExpectancy} \tab {Average life expectancy (in years)}\cr
  }
}
\details{
A subset of data from AllCountries for a random sample of 50 countries.
Data for 2008 to avoid many missing values in more recent years.
}
\source{
Data collected from the World Bank website, worldbank.org. 
}

\keyword{datasets}
