\name{SampCountries}
\alias{SampCountries}
\docType{data}
\title{
AllCountries
}
\description{
Data on a sample of countries of the world
}
\usage{data(SampCountries)}
\format{
  A data frame with 50 observations on the following 13 variables.
  \describe{
    \item{\code{Country}}{Name of the country}
    \item{\code{LandArea}}{Size in sq. kilometers}
    \item{\code{Population}}{Population in millions}
    \item{\code{Energy}}{Energy usage (kilotons of oil)}
    \item{\code{Rural}}{Percentage of population living in rural areas}
    \item{\code{Military}}{Percentage of government expenditures directed toward the military}
    \item{\code{Health}}{Percentage of government expenditures directed towards healthcare}
    \item{\code{HIV}}{Percentage of the population with HIV}
    \item{\code{Internet}}{Percentage of the population with access to the internet}
    \item{\code{Developed}}{Categories for kilowat hours per capita: \code{1}= under 2500, \code{2}=2500 to 5000, \code{3}=over 5000}
    \item{\code{BirthRate}}{Births per 1000 people}
    \item{\code{ElderlyPop}}{Percentage of the population at least 65 year old}
    \item{\code{LifeExpectancy}}{Average life expectancy (in years)}
  }
}
\details{
A subset of data from AllCountries for a random sample of 50 countries.
Data for 2008 to avoid many missing values in more recent years.
}
\source{
Data collected from the World Bank website, worldbank.org. 
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(SampCountries)
}
\keyword{datasets}
