\name{SandP500}
\alias{SandP500}
\docType{data}
\title{
S \& P 500 Prices
}
\description{
Daily data for S&P 500 Stock Index
}
%\usage{data(SandP500)}
\format{
  A dataset with 252 observations on the following 6 variables.
  \tabular{rl}{
    \code{Date} \tab {Date}\cr
    \code{Open} \tab {Opening value}\cr
    \code{High} \tab {High point for the day}\cr
    \code{Low} \tab {Low point for the day}\cr
    \code{Close} \tab {Closing value}\cr
    \code{Volume} \tab {Shares traded (in millions)}\cr
  }
}
\details{
Daily prices for the S&P 500 Stock Index for trading days in 2010. 
}
\source{
Downladed from http://finance.yahoo.com/q/hp?s=^GSPC+Historical+Prices
}

\keyword{datasets}
