\name{SandP500}
\alias{SandP500}
\docType{data}
\title{
S \& P 500 Prices
}
\description{
Daily data for S\& P 500 Stock Index
}
\usage{data(SandP500)}
\format{
  A data frame with 252 observations on the following 6 variables.
  \describe{
    \item{\code{Date}}{}
    \item{\code{Open}}{Opening value}
    \item{\code{High}}{High point for the day}
    \item{\code{Low}}{Low point for the day}
    \item{\code{Close}}{Closing value}
    \item{\code{Volume}}{Shares traded (in millions)}
  }
}
\details{
Daily prices for the S\& P 500 Stock Index for trading days in 2010. 
}
\source{
Downladed from http://finance.yahoo.com/q/hp?s=^GSPC+Historical+Prices
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(SandP500)
}
\keyword{datasets}
