\name{SandwichAnts}
\alias{SandwichAnts}
\docType{data}
\title{
Sandwich Ants
}
\description{
Ant Counts on samples of different sandwiches
}
%\usage{data(SandwichAnts)}
\format{
  A dataset with 24 observations on the following 5 variables.
  \tabular{rl}{
    \code{Butter} \tab {Butter on the sandwich? \code{no}}\cr
    \code{Filling} \tab {Type of filling: \code{Ham \& Pickles}, \code{Peanut Butter}, or \code{Vegemite}}\cr
    \code{Bread} \tab {Type of bread:  \code{Multigrain}, \code{Rye}, \code{White}, or \code{Wholemeal}}\cr
    \code{Ants} \tab {Number of ants on the sandwich}\cr
    \code{Order} \tab {Trial number}\cr
  }
}
\details{
As young students, Dominic Kelly and his friends enjoyed watching ants gather on pieces of sandwiches.  Later, as a university student, Dominic decided to study this with a more formal experiment. He chose three types of sandwich fillings (vegemite, peanut butter, and ham \& pickles),  four types of bread (multigrain, rye, white, and wholemeal), and put butter on some of the sandwiches. \cr
To conduct the experiment he randomly chose a sandwich, broke off a piece, and left it on the ground near an ant hill.  After several minutes he placed a jar over the sandwich bit and counteed the number of ants. He repeated the process, allowing time for ants to return to the hill after each trial, until he had two samples for each combination of the factors. \cr
This dataset has only sandwiches with no butter. The data in SandwichAnts2 adds information for samples with butter. 
}
\source{
Margaret Mackisack, ``Favourite Experiments: An Addendum to What is the Use of Experiments Conducted by Statistics Students?", 
Journal of Statistics Education (1994) \cr
 http://www.amstat.org/publications/jse/v2n1/mackisack.supp.html
}

\keyword{datasets}
