\name{SkateboardPrices}
\alias{SkateboardPrices}
\docType{data}
\title{Skateboard Prices}
\description{
Prices of skateboards for sale online
}
%\usage{data(SkateboardPrices)}
\format{
  A dataset with 20 observations on the following variable.
  \tabular{rl}{
    \code{Price} \tab {Selling price in dollars}\cr
  }
}
\details{
Prices for skateboards offered for sale on eBay. 
}
\source{
Random sample taken from all skateboards available for sale on eBay on February 12, 2012.
}

\keyword{datasets}
