\name{SleepCaffeine}
\alias{SleepCaffeine}
\docType{data}
\title{
Sleep Caffeine
}
\description{
Experimentn to compare word recall after sleep or caffeine
}
%\usage{data(SleepCaffeine)}
\format{
  A dataset with 24 observations on the following 2 variables.
  \tabular{rl}{
    \code{Group} \tab {Treatment: \code{Caffeine} \code{Sleep}}\cr
    \code{Words} \tab {Number of words recalled}\cr
  }
}
\details{
A random sample of 24 adults were divided equally into two groups and given a list of 24 words to memorize. During a break, one group takes a 90 minute nap while another group is given a caffeine pill.  The response variable is the number of words participants are able to recall following the break. 
}
\source{
Mednick, Cai, Kanady, and Drummond, "Comparing the benefits of caffeine, naps and placebo on verbal, motor and perceptual memory", Behavioural Brain Research, 193 (2008), 79-86.
}

\keyword{datasets}
