\name{StatisticsPhD}
\alias{StatisticsPhD}
\docType{data}
\title{
Statistics PhD Programs
}
\description{
Enrollments in Statistics PhD Programs
}
%\usage{data(StatisticsPhD)}
\format{
  A dataset with 82 observations on the following 3 variables.
  \tabular{rl}{
    \code{University} \tab {Name of the school}\cr
    \code{Department} \tab {Type of department: \code{Biostatistics} or \code{Statistics}}\cr
    \code{FTGradEnrollment} \tab {Full time graduate student enrollment}\cr
  }
}
\details{
Graduate student enrollments in Statistics and Biostatistics departments in 2009. The list does not include combined departments of mathematics and statistics and does not
include departments that did not reply to the AMS survey.
}
\source{
The full list of the 82 Group IV departments was obtained at http://www.ams.org/profession/data/annual-survey/group iv.
Data on enrollment were obtained primarily from Assistantships and Graduate Fellowships in the Mathematical Sciences, 2009, American Mathematical Society. 
}

\keyword{datasets}
