\name{StockChanges}
\alias{StockChanges}
\docType{data}
\title{
Stock Changes
}
\description{
Stock price change for a smaple of stocks from the S \& P 500 (August 2-6, 2010)
}
%\usage{data(StockChanges)}
\format{
  A dataset with 50 observations on the following variable.
  \tabular{rl}{
    \code{SPChange} \tab {Change in sock price (in dollars)}\cr
  }
}
\details{
A random sample of 50 companies from Standard \& Poor's index of 500 companies was selected. The change in the price of the stock (in dollars)
over the 5-day period from August 2 - 6, 2010 was recorded for each company in the sample.
}
\source{
http://money.cnn.com/data/markets/sandp/
}

\keyword{datasets}
