\name{StudentSurvey}
\alias{StudentSurvey}
\docType{data}
\title{
Student Survey Data
}
\description{
Data from a survey of students in introductory statistics courses
}
%\usage{data(StudentSurvey)}
\format{
  A dataset with 362 observations on the following 17 variables.
  \tabular{rl}{
    \code{Year} \tab {Year in school}\cr
    \code{Gender} \tab {Student's gender:  \code{F} or \code{M}}\cr
    \code{Smoke} \tab {Smoker? \code{No} or \code{Yes}}\cr
    \code{Award} \tab {Prefered award: \code{Academy} \code{Nobel} \code{Olympic}}\cr
    \code{HigherSAT} \tab {Which SAT is higher?  \code{Math} or \code{Verbal}}\cr
    \code{Exercise} \tab {Hours of exercsie per week}\cr
    \code{TV} \tab {Hours of TV viewing per week}\cr
    \code{Height} \tab {Height (in inches)}\cr
    \code{Weight} \tab {Weight (in pounds)}\cr
    \code{Siblings} \tab {Number of siblings}\cr
    \code{BirthOrder} \tab {Birth order, 1=oldest}\cr
    \code{VerbalSAT} \tab {Verbal SAT score}\cr
    \code{MathSAT} \tab {Math SAT score}\cr
    \code{SAT} \tab {Combined Verbal + Math SAT}\cr
    \code{GPA} \tab {College grade point average}\cr
    \code{Pulse} \tab {Pulse rate (beats per minute)}\cr
    \code{Piercings} \tab {Number of body piercings}\cr
  }
}
\details{
Data from an in-class survey given to introductory statistics students over several years. 
}
\source{
In-class student survey
}

\keyword{datasets}
