\name{TenCountries}
\alias{TenCountries}
\docType{data}
\title{Ten Countries}
\description{
A subset of the ALLCountries data for a random sample of ten countries
}
%\usage{data(TenCountries)}
\format{
  A dataset with 10 observations on the following 4 variables.
  \tabular{rl}{
    \code{Country} \tab {\code{Armenia} \code{Bahamas} \code{Lebanon} \code{Macedonia} \code{North Korea} \code{Romania} \code{Serbia} \code{Slovenia} \code{Tunisia} \code{Uzbekistan}}\cr
    \code{Code} \tab {\code{ARM} \code{BHS} \code{LBN} \code{MKD} \code{PRK} \code{ROU} \code{SRB} \code{SVN} \code{TUN} \code{UZB}}\cr
    \code{Area} \tab {Area (in 1,000's sq km)}\cr
    \code{PctRural} \tab {Percentage of population living in rural areas}\cr
  }
}
\details{
A random sample of ten countries selected from the AllCountries dataset. 
}
\source{
Data collected from the World Bank website, worldbank.org. 
}

\keyword{datasets}
