\name{TextbookCosts}
\alias{TextbookCosts}
\docType{data}
\title{
Textbook Costs
}
\description{
Prices for textbooks for different courses
}
%\usage{data(TextbookCosts)}
\format{
  A dataset with 40 observations on the following 3 variables.
  \tabular{rl}{
    \code{Field} \tab {General discipline of the course:  \code{Arts}, \code{Humanities}, \code{NaturalScience}, or \code{SocialScience}}\cr
    \code{Books} \tab {Number of books requiired}\cr
    \code{Cost} \tab {Total cost (in dollars) for required books}\cr
  }
}
\details{
Data are from samples of ten courses in each of four disciplines at a liberal arts college.  For each course the bookstore's website lists the required texts(s) and costs. Data were collected for the Fall 2011 semester. 
}
\source{
Bookstore online site
}

\keyword{datasets}
