\name{ToenailArsenic}
\alias{ToenailArsenic}
\docType{data}
\title{
Toenail Arsenic
}
\description{
Arsenic in toenails of 19 people using private wells in New Hampshire
}
%\usage{data(ToenailArsenic)}
\format{
  A dataset with 19 observations on the following variable.
  \tabular{rl}{
    \code{Arsenic} \tab {Level of arsenic found in toenails}\cr
  }
}
\details{
Level of arsenic was measured in toenails of 19 subjects from New Hampshire, all with private wells as their main water source. 
}
\source{
Adapted from Karagas, et.al.,"Toenail Samples as an Indicator of Drinking Water Arsenic Exposure", Cancer Epidemiology, Biomarkers and Prevention 1996;5:849-852.
}

\keyword{datasets}
