\name{USStates}
\alias{USStates}
\docType{data}
\title{
US State Data
}
\description{
Various data for all 50 U.S. states
}
%\usage{data(USStates)}
\format{
  A dataset with 50 observations on the following 17 variables.
  \tabular{rl}{
    \code{State} \tab {Name of state}\cr
    \code{HouseholdIncome} \tab {Mean household income (in dollars)}\cr
    \code{IQ} \tab {Mean IQ score of residents}\cr
    \code{McCainVote} \tab {Percentage of votes for John McCain in 2008 Presidential election}\cr
    \code{Region} \tab {Area of the country: \code{MW}=Midwest, \code{NE}=Northeast, \code{S}=South, or \code{W}=West}\cr
    \code{ObamaMcCain} \tab {Which 2008 Presidential candidate won state?  \code{M}=McCain or \code{O}=Obama}\cr
    \code{Population} \tab {Number of residents (in millions)}\cr
    \code{EighthGradeMath} \tab {a numeric vector}\cr
    \code{HighSchool} \tab {Percentage of high school graduates}\cr
    \code{GSP} \tab {Gross State Product (dollars per capita)}\cr
    \code{FiveVegetables} \tab {Percentage of residents who eat at least five servings of fruits/vegetables per day}\cr
    \code{Smokers} \tab {Percentage of residents who smoke}\cr
    \code{PhysicalActivity} \tab {Percentage of residents who have competed in a physical activity in past month}\cr
    \code{Obese} \tab {Percentage of residents classified as obese}\cr
    \code{College} \tab {Percentage of residents with college degrees}\cr
    \code{NonWhite} \tab {Percentage of residents who are not white}\cr
    \code{HeavyDrinkers} \tab {Percentage of residents who drink heavily}\cr
  }
}
\details{
Demographic data for U.S. Sstates
}
\source{
Various online sources, mostly at www.census.gov
}

\keyword{datasets}
